% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{acc}
\alias{acc}
\title{Catch curve estimates of total mortality at age (Z)}
\usage{
acc(object, ...)
}
\arguments{
\item{object}{An input object of the right class}
}
\description{
Catch curve estimates of total mortality at age (Z)
}
\examples{
data(ple4)
}
