% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{accessors}
\alias{accessors}
\alias{catch}
\alias{catch.n}
\alias{catch.n<-}
\alias{catch.q}
\alias{catch.q<-}
\alias{catch.wt}
\title{Accesor and replacement methods for FLCore classes}
\usage{
range(x, i) <- value
}
\arguments{
\item{value}{Object to be inserted into the relevant slot}

\item{object}{The object from which a slot is to be extracted or replaced}
}
\value{
The required slot, for an accessor method, or invisible modifies the
  object, for the replacement one.
}
\description{
All S4 classes defined in FLCore have methods for accessing and replacing any
of their slots. These methods are named as the slot, and will return the
content of the slot, for the accessor method, or modify it with the provided
value.
}
\details{
Accessors and replacement methods, with some exception, are created at build
time by calls to the \code{createFLAccessors} function. An accesor method is
created for each slot, with simply calls \code{slot()} on the relevant slot
name. For slots of class \code{\link{FLQuant}}, or \code{FLArray}-based, two
methods are created: one if \code{value} is of class \code{FLQuant}, and
another for \code{value} being a numeric vector. The later would insert the
vector into the slot structure, using R's recycling rules.

Users are encouraged to use the accessor methods, rather than the '@' operator
or the \code{slot()} method, to isolate code from the internal structure of
the class. If a slot was to be altered or deleted in the future, a method
would be provided to return the same value, computed from other slots.

Some of these methods might already not access directly an slot, and instead
carry out a calculation to return the requested value, depending on the class
being called with. Please refer to the particular method implementation to
see if this is the case.

Accessor methods for slots of class \code{\link{predictModel}} behave
differently depending on the \code{compute} argument. Please refer to the
relevant help page for further clarification.
}
\examples{

data(ple4)

# To access the catch slot in an FLStock, use
catch(ple4)

# while to modify it, do
catch(ple4) <- catch(ple4) * 2

# A number can be used as input, to be recycled
m(ple4) <- 0.3
# same as a longer vector, by age
m(ple4) <- 0.4^(seq(1, 2, length=10))

# To see the methods defined by createFLAccessors, run, for example
getMethod('catch', 'FLS')

}
\seealso{
\code{\link{FLQuant}}, \code{\link{FLStock}}, \code{\link{FLIndex}},
\code{\link{FLBiol}}, \code{\link{predictModel}}
}
\author{
The FLR Team
}
\keyword{methods}
