% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\docType{methods}
\name{adjust,FLStock-method}
\alias{adjust,FLStock-method}
\title{Recalculate to adjust abundances to F and M}
\usage{
\S4method{adjust}{FLStock}(object)
}
\arguments{
\item{object}{an \code{FLStock} object}
}
\value{
\code{FLStock} object
}
\description{
An FLStock object is projected forward using the initial abundances and
the total mortality-at-age per timestep. New values for the stock.n and
catch.n slots are calculated, assuming that harvest and m are correct.
This calculation provides a test of the internal consistency of the object.
}
\examples{
data(ple4)
test <- adjust(ple4)
# Difference in catch due to estimation error
plot(FLStocks(PLE=ple4, TEST=test))
}
\seealso{
\code{\link{harvest}}
}
