% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{ageopt}
\alias{ageopt}
\title{Age at which a cohort reaches its maximum biomass, calculated by year}
\usage{
\S4method{ageopt}{FLStock}(object)
}
\arguments{
\item{object}{An object of class 'FLStock'}
}
\value{
The age at which maximum biomass is reached, an 'FLQuant'.
}
\description{
The optimal (or critical) age is the transition point when a cohort achieves
its maximum biomass in the absemce of fishing, i.e. losses due to natural
mortality are now greater than gains due to increase in individual biomass.
}
\examples{
data(ple4)
ageopt(ple4)
}
\seealso{
\link{FLStock}
}
\author{
The FLR Team
}
\keyword{methods}
