% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModel.R
\docType{methods}
\name{AIC}
\alias{AIC}
\alias{AIC,FLModel,missing-method}
\alias{AIC,FLModel,numeric-method}
\title{Method AIC}
\usage{
\S4method{AIC}{FLModel,numeric}(object, k = 2)
}
\arguments{
\item{object}{an FLModel object}

\item{k}{the penalty per parameter to be used; the default 'k = 2' is the classical AIC.}
}
\description{
Akaike's information criterion (AIC) method
}
\details{
A method to calculate Akaike's 'An Information Criterion' (AIC) of an
\link{FLModel} object from the value of the obtained log-likelihood stored
in its \code{logLik} slot.
}
\section{Generic function}{
 AIC(object, k)
}
\examples{

data(nsher)
AIC(nsher)

}
\author{
The FLR Team
}
\seealso{
\link[stats]{AIC}, \link[stats]{logLik}, \link{FLModel}
}
\keyword{methods}

