% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{flc}
\alias{flcs}
\alias{fli}
\alias{flib}
\alias{flp}
\alias{flq}
\alias{flqd}
\alias{flqp}
\alias{flqs}
\alias{fls}
\alias{flsl}
\alias{flsr}
\alias{flss}
\title{Short aliases for most FLCore classes construction methods}
\usage{
flq(...) for FLQuant(...)
flqp(...) for FLQuantPoint(...)
flqd(...) for FLQuantDistr(...)
flc(...) for FLCohort(...)
flp(...) for FLPar(...)
fls(...) for FLStock(...)
flsl(...) for FLStockLen(...)
fli(...) for FLIndex(...)
flib(...) for FLIndexBiomass(...)
flsr(...) for FLSR(...)
flqs(...) for FLQuants(...)
flcs(...) for FLCohorts(...)
flss(...) for FLStocks(...)
flis(...) for FLIndices(...)
flps(...) for FLPars(...)
flb(...) for FLBiol(...)
flbs(...) for FLBiols(...)
flms(...) for FLModelSim(...)
flmss(...) for FLModelSims(...)
pm(...) for predictModel(...)
}
\value{
An object of the requested class
}
\description{
When working interactively, the naming convention of the FLCore classes can
become tiresome. This set of short, lowercase aliases make calling the class
construction methods a bit simpler.
}
\details{
We recomend you use the full name of classes and methods when developing code
that you intend to keep or distribute. A script written using these aliases
can be later 'corrected' by substituting the aliases used in your editor.
}
\examples{

flq <- flc(1:10, units="kg")
}
\author{
The FLR Team
}
\seealso{
\link{FLQuant}, \link{FLQuantPoint}, \link{FLQuantDistr}, \link{FLCohort}, \link{FLPar},
\link{FLStock}, \link{FLStockLen}, \link{FLIndex}, \link{FLIndexBiomass}, \link{FLSR},
\link{FLQuants}, \link{FLCohorts}, \link{FLStocks}, \link{FLIndices}, \link{FLPars}, 
\link{FLBiol}, \link{FLBiols}, \link{FLModelSim}, \link{FLModelSims}, \link{predictModel}
}
\keyword{methods}

