% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R, R/FLStock.R
\docType{methods}
\name{append-FLCore}
\alias{append-FLCore}
\alias{append,FLStock,FLStock-method}
\title{Append objects along the year dimension}
\usage{
\S4method{append}{FLQuant,FLQuant}(x, values,
  after = dims(values)$minyear - 1)

\S4method{append}{FLStock,FLStock}(x, values,
  after = dims(values)$minyear - 1)
}
\arguments{
\item{x}{the object to which the values are to be appended to.}

\item{values}{to be included in the modified object.}

\item{after}{a year dimname after with the values are to be appended.}
}
\value{
An object of the same class as x with values appended.
}
\description{
Method to append objects along the \emph{year} dimensions, by extending, combining
and substituting sections of them.
}
\details{
FLR objects are commonly manipulated along rhe year dimension, and the append
method offers a single interface for substituting parts of an object with
another, or combine them into one, extending them when necessary.
The object to be included or added to the first will be placed as defined by
the \emph{year} dimnames, unless the \emph{after} input argument specifies otherwise.

The method for an *FLStock*
}
\examples{
fq1 <- FLQuant(1, dimnames=list(age=1:3, year=2000:2010))
fq2 <- FLQuant(2, dimnames=list(age=1:3, year=2011:2012))
fq3 <- FLQuant(2, dimnames=list(age=1:3, year=2014:2016))

# Appends by dimnames$year
append(fq1, fq2)
# Appends by dimnames$year with gap (2011:2013)
append(fq1, fq3)
# Appends inside x
append(fq1, fq2, after=2009)
# Appends after end of x
append(fq1, fq2, after=2013)
data(ple4)
fs1 <- window(ple4, end=2001)
fs2 <- window(ple4, start=2002)
fs3 <- window(ple4, start=2005)

# Appends by dimnames$year
stock.n(append(fs1, fs2))

# Appends by dimnames$year with gap (2011:2013)
stock.n(append(fs1, fs3))

# Appends inside x
stock.n(append(fs1, fs3, after=2000))
# Appends after end of x
stock.n(append(fs1, fs3, after=2005))
}
\seealso{
\link[base:append]{base::append}
}
\author{
The FLR Team
}
\keyword{methods}
