% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R
\docType{methods}
\name{apply}
\alias{apply}
\alias{apply,ANY,missing,missing-method}
\alias{apply,FLArray,numeric,function-method}
\alias{apply,FLPar,ANY,ANY-method}
\alias{apply,FLPar,numeric,function-method}
\alias{apply,FLQuant,numeric,function-method}
\title{Method apply}
\usage{
\S4method{apply}{FLArray,numeric,`function`}(X, MARGIN, FUN, ...)

\S4method{apply}{FLPar,ANY,ANY}(X, MARGIN, FUN, ...)
}
\description{
Functions can be applied to margins of an \code{FLQuant} or \code{FLPar} array
using this method. In contrast with the standard R method, dimensions are not collapsed
in the output object.
}
\details{
\code{FUN} in the case of an \code{FLQuant} must be a function whose results is least one
dimension less when applied over an array (e.g. sum, mean, ...).

For further details see \link[base]{apply}.
}
\section{Generic function}{
 apply(X,MARGIN,FUN)
}
\examples{
flq <- FLQuant(rlnorm(1000), dim=c(10,20,1,1,1,5))
apply(flq, 1, sum)
apply(flq, 2:6, sum)
}
\author{
The FLR Team
}
\seealso{
\link[base]{apply}
}
\keyword{methods}

