% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{ar1deviances}
\alias{ar1deviances}
\title{Extend AR(1) log-normal deviances beyond a given year}
\usage{
ar1deviances(x, year)
}
\arguments{
\item{x}{An \code{FLQuant} of log-normal deviances with historical data
up to at least \code{year}.}

\item{year}{Integer or character; the last year of the historical period.
Values in \code{x} after this year will be replaced by simulations.}
}
\value{
An \code{FLQuant} of the same dimensions as \code{x}, with years
after \code{year} replaced by AR(1) log-normal simulations estimated
from the historical data.
}
\description{
Replaces the values of an \code{FLQuant} beyond a specified year with
simulated AR(1) log-normal deviates, using the autocorrelation and
standard deviation estimated from the historical period up to and
including that year.
}
\seealso{
\link{rlnormar1}
}
\author{
The FLR Team
}
\keyword{utilities}
