% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R
\docType{methods}
\name{Arith,numeric,FLArray-method}
\alias{Arith,numeric,FLArray-method}
\alias{Arith,FLArray,numeric-method}
\alias{Arith,FLArray,FLArray-method}
\alias{Arith,FLPar,FLPar-method}
\alias{Arith,FLArray,FLPar-method}
\alias{Arith,FLPar,FLArray-method}
\title{Arithmetic operators for FLCore classes}
\usage{
\S4method{Arith}{numeric,FLArray}(e1, e2)

\S4method{Arith}{FLArray,numeric}(e1, e2)

\S4method{Arith}{FLArray,FLArray}(e1, e2)

\S4method{Arith}{FLPar,FLPar}(e1, e2)

\S4method{Arith}{FLArray,FLPar}(e1, e2)

\S4method{Arith}{FLPar,FLArray}(e1, e2)
}
\description{
Overloaded arithmetic operators for FLCore classes
}
\details{
These methods apply the standard arithmetic operators included in the
\code{\link[methods]{Arith}} group ("+", "-", "*", "^", "%%", "%/%", and
"/"), so that they return an object of the appropriate class.

When the operation involves objects of two classes (e.g. \code{\link{FLPar}} and \code{\link{FLQuant}}),
the class is the returned object is that of the more complexs object, in this
case \code{\link{FLQuant}}.
}
\examples{

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
flp <- FLPar(a=99)

# FLQuant and numeric
flq * 25
# Two FLQuant objects
flq + flq

# FLQuant and FLPar
flq / flp
}
\seealso{
\link[methods:Arith]{methods::Arith} \link[base:Arithmetic]{base::Arithmetic}
}
\author{
The FLR Team
}
\keyword{methods}
