\name{as.FLFleet}
\alias{as.FLFleet}
\alias{as.FLFleet-method}
\alias{as.FLFleet,FLStock-method}
\title{as.FLFleet method}
\description{
    Conversion to \code{FLFleet} of various other classes.
}
\usage{
as.FLFleet(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An FLR object to be converted. Currently only working for \code{FLStock} objects}
}
\details{
}
\value{
    This method should return an \code{FLFleet} object
}
\author{FLR Team}
\seealso{\code{\link{FLFleet-class}}}
\examples{
}
\keyword{methods}
\keyword{manip}
