% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSR.R
\name{as.FLSRs}
\alias{as.FLSRs}
\title{Convert an FLStock into a list of one or FLSR objects.}
\usage{
as.FLSRs(x, models = NULL, ...)
}
\arguments{
\item{x}{An estimated FLStock object to coerce.}

\item{models}{Name(s) of model(s) to fit.}

\item{...}{Any extra arguments to be passed to \emph{as.FLSR}.}
}
\value{
An objecdt of class \code{FLSRs}
}
\description{
A single \code{FLStock} can be coerced into a list with one or more objects of
class \code{FLSR}, each of them typically set to a diefferemt stock-recruit model.
}
\examples{
data(ple4)
as.FLSRs(ple4, model=c("bevholt", "segreg"))
}
\seealso{
\linkS4class{FLSRs} \linkS4class{FLSRs} \code{\link[=as.FLSR]{as.FLSR()}}
}
\author{
FLR Team, 2023.
}
\keyword{classes}
