% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R, R/oem.R
\name{bias}
\alias{bias}
\alias{bias,FLParJK-method}
\title{Bias of estimates through jackknife}
\usage{
\S4method{bias}{FLQuantJK}(x)

\S4method{bias}{FLParJK}(x)

bias(object, bias = 0.02)
}
\arguments{
\item{x}{An object holding estimates obtained through jackknife}

\item{object}{An \code{FLQuant} whose year dimension and dimnames are used.}

\item{bias}{Numeric; per-year fractional bias; defaults to 0.02 (2\\% per
year).}
}
\value{
A value for the mean bias

An \code{FLQuant} with the same dimnames as \code{object}
containing the cumulative bias multipliers.
}
\description{
Description: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.

Computes a cumulative multiplicative bias factor that increases (or
decreases) systematically over the years in an \code{FLQuant}. Starting
from 1, each subsequent year is multiplied by \code{1 + bias}, producing
a trend that can represent, e.g., a gradually increasing observation bias.
}
\details{
Details: Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa,
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien,
feugiat blandit arcu in, dapibus dictum felis.

\deqn{\widehat{Bias}_{(\theta)} = (n - 1)((\frac{1}{n}\sum\limits_{i=1}^n\hat{\theta}_{(i)})-\hat{\theta})}{}
}
\examples{

flq <- FLQuant(1:8)
flj <- jackknife(flq)
bias(flj)

}
\seealso{
\link{FLComp}

\link{biased}
}
\author{
The FLR Team
}
\keyword{classes}
\keyword{utilities}
