% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModel.R
\docType{methods}
\name{BIC}
\alias{BIC}
\alias{BIC,FLModel-method}
\title{Method BIC
Bayesian information criterion (BIC) method}
\usage{
\S4method{BIC}{FLModel}(object)
}
\arguments{
\item{object}{a fitted FLModel object for which there exists a 'logLik' method to
extract the corresponding log-likelihood.}
}
\description{
A method to calculate the Bayesian information criterion (BIC), also known
as Schwarz's Bayesian criterion of an \link{FLModel} object from the value
of the obtained log-likelihood stored in its \code{logLik} slot.
}
\section{Generic function}{
 BIC(object)
}
\examples{

data(nsher)
BIC(nsher)

}
\author{
The FLR Team
}
\seealso{
\link[stats]{AIC}, \link[stats]{BIC}, \link{FLModel},
\link[stats]{logLik}
}
\keyword{methods}

