% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{biomass_metrics}
\alias{biomass_metrics}
\alias{ssb,FLStock-method}
\alias{ssb_end}
\alias{ssb_start}
\alias{vb,FLStock,ANY-method}
\alias{vb,FLStock,missing-method}
\alias{exb,FLStock-method}
\alias{biomass_end}
\alias{tsb,FLStock-method}
\alias{ssb_next}
\alias{biomass_spawn}
\title{Biomass metrics for FLStock}
\usage{
\S4method{ssb}{FLStock}(object, byage = FALSE, byunit = FALSE, ...)

ssb_end(object, byage = FALSE, byunit = FALSE, ...)

ssb_start(object, byage = FALSE, byunit = FALSE, ...)

\S4method{vb}{FLStock,ANY}(x, sel, byage = FALSE, ...)

\S4method{vb}{FLStock,missing}(x, byage = FALSE, ...)

\S4method{exb}{FLStock}(x, sel = catch.sel(x), wt = catch.wt(x), byage = FALSE, ...)

biomass_end(x)

\S4method{tsb}{FLStock}(object, time = m.spwn(object), byage = FALSE, byunit = FALSE, ...)

ssb_next(x, fbar = 0, wts.nyears = 3, fbar.nyears = 3)

biomass_end(x)

biomass_spawn(x)
}
\arguments{
\item{object, x}{An object of class \linkS4class{FLStock}.}

\item{byage}{Logical; if \code{TRUE} return values by age rather than
summing over ages. Default \code{FALSE} (except \code{biomass_end}
where it is \code{TRUE}).}

\item{byunit}{Logical; if \code{TRUE} return values by unit rather than
summing over units. Default \code{FALSE}.}

\item{...}{Named slot–value pairs to temporarily override slots of
\code{object} before calculation (e.g. \code{m = 0.2}).}

\item{sel}{Selectivity-at-age as an \linkS4class{FLQuant}. Defaults to
\code{catch.sel(x)} in \code{vb} and \code{exb}.}

\item{wt}{Weight-at-age as an \linkS4class{FLQuant}. Defaults to
\code{catch.wt(x)} in \code{exb}.}

\item{time}{Fraction of year at which biomass is evaluated in \code{tsb};
defaults to \code{m.spwn(object)}.}

\item{fbar}{Assumed mean fishing mortality for the projected extra year in
\code{ssb_next}. Default \code{0}.}

\item{wts.nyears}{Number of years over which to average weights-at-age and
maturity for the projected year in \code{ssb_next}. Default \code{3}.}

\item{fbar.nyears}{Number of years over which to average selectivity, M and
spawning fractions for the projected year in \code{ssb_next}. Default \code{3}.}
}
\value{
An \linkS4class{FLQuant} with the requested biomass quantity.
Dimensions depend on \code{byage} and \code{byunit}.
}
\description{
A family of functions and methods to compute different biomass quantities
from an \linkS4class{FLStock} object, varying in which component of the
stock is selected and at what point in the time step mortality is evaluated.
}
\details{
All calculations use the internal \code{.biomass} workhorse, which applies
fishing mortality (F or harvest rate) and natural mortality proportionally
according to the \code{harvest.spwn} and \code{m.spwn} timing fractions.

\describe{
\item{\code{ssb}}{Spawning stock biomass at the spawning fraction of the
time step, weighted by maturity-at-age. Mortality applied up to
\code{harvest.spwn} and \code{m.spwn}.}
\item{\code{ssb_end}}{Spawning stock biomass at the \emph{end} of the
time step (\code{ph = pm = 1}): full annual F and M applied before
weighting by maturity.}
\item{\code{ssb_start}}{Spawning stock biomass at the \emph{start} of the
time step (\code{ph = pm = 0}): no mortality applied, numbers weighted
by maturity and stock weight only.}
\item{\code{ssb_next}}{Projected SSB for the year immediately following
the last year in the object, calculated from survivors and an assumed
\code{fbar}. Weights-at-age and maturity are averaged over the last
\code{wts.nyears} years; selectivity, M and spawning fractions over the
last \code{fbar.nyears} years.}
\item{\code{tsb}}{Total stock biomass (all ages, \code{sel = 1}) at the
spawning fraction of the time step.}
\item{\code{biomass_end}}{Total stock biomass at the end of the time step
(\code{ph = pm = 1}, \code{sel = 1}).}
\item{\code{biomass_spawn}}{Total stock biomass surviving to the spawning
fraction, using the stored \code{harvest.spwn} and \code{m.spwn}
slots directly.}
\item{\code{vb}}{Vulnerable biomass at the \emph{start} of the time step
(\code{ph = pm = 0}), using a user-supplied or default
(\code{catch.sel}) selectivity.}
\item{\code{exb}}{Exploitable biomass at the start of the time step,
using \code{catch.sel} and \code{catch.wt} by default, or
user-supplied \code{sel} and \code{wt}.}
}
}
\examples{
data(ple4)

# SSB at spawning fraction of the year
ssb(ple4)

# SSB at start and end of the time step
ssb_start(ple4)
ssb_end(ple4)

# Projected SSB for year after last, assuming fbar = 0.3
ssb_next(ple4, fbar = 0.3)

# Total stock biomass
tsb(ple4)
biomass_end(ple4)
biomass_spawn(ple4)

# Vulnerable and exploitable biomass
vb(ple4)
exb(ple4)
}
\seealso{
\code{\link{fbar}}, \code{\link{catch}}, \linkS4class{FLStock}
}
\author{
FLR Team
}
\keyword{manip}
