\name{bkey}
\docType{methods}
\alias{bkey}
\alias{bkey-methods}
\alias{bkey,list-method}
\title{Generate key for bubbles plot}
\description{
\code{bkey} generates the bubble scale that shows on bubbles plot for \code{FLQuants} objects. It is not intended to be used directly by users but you may try.
}
\section{Generic function}{
bkey(object) 
}
\section{Methods}{\describe{
\item{ signature(object=list) :}{The list must have at least the elements \code{data}, \code{cex}, \code{bub.col} and \code{bub.scale}.}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
