% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{bubbles}
\alias{bubbles}
\alias{bubbles-methods}
\alias{bubbles,formula,FLQuant-method}
\alias{bubbles,formula,data.frame-method}
\alias{bubbles,formula,FLCohort-method}
\alias{bubbles,formula,FLQuants-method}
\title{Method Bubbles plot}
\usage{
\S4method{bubbles}{formula,FLQuant}(x, data, bub.scale = 2.5,
  col = c("blue", "red"), ...)

\S4method{bubbles}{formula,data.frame}(x, data, bub.scale = 2.5,
  col = c("blue", "red"), ...)

\S4method{bubbles}{formula,FLCohort}(x, data, bub.scale = 2.5, ...)

\S4method{bubbles}{formula,FLQuants}(x, data, bub.scale = 2.5,
  bub.col = gray(c(0.1, 0.1)), ...)
}
\description{
This method plots three dimensional data such as matrices by age and year or
age-class, very common in fisheries. The area of each bubble is proportional
to the corresponding value in the matrix. Note that \code{bubbles} accepts
an argument \code{bub.scale} to control the relative size of the bubbles.
Positive and negative values have separate colours.
}
\section{Generic function}{
 bubbles(x, data)
}

\examples{

data(ple4)
bubbles(age~year, data=catch.n(ple4))
bubbles(age~year, data=catch.n(ple4), bub.scale=5)
bubbles(age~cohort, data=FLCohort(catch.n(ple4)), bub.scale=5)

qt01 <- log(catch.n(ple4)+1)
qt02 <- qt01+rnorm(length(qt01))
flqs <- FLQuants(qt01=qt01, qt02=qt02)
bubbles(age~year|qname, data=flqs, bub.scale=1)

qt03 <- FLQuant(rnorm(100),dimnames=list(age=as.character(1:10),
  year=as.character(1:10)))
bubbles(age~year, data=qt03, bub.scale=7, col=c("black","red"), pch=16)

}
\seealso{
\link[lattice]{lattice}, \code{\linkS4class{FLQuant}},
\code{\linkS4class{FLQuants}},\code{\linkS4class{FLCohort}}
}
\author{
The FLR Team
}
\keyword{methods}
