% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{catch.n,FLQuant-method}
\alias{catch.n,FLQuant-method}
\title{catch.n calculation method}
\usage{
\S4method{catch.n}{FLQuant}(object, harvest, m)
}
\description{
Calculate catch.n (catch-at-age/length) from abundances, F and M using the catch equation
}
\details{
The catch-at-age/length, commonly found in the catch.n slot of an
\code{FLStock} object, can be simply calculated from abundances-at-age/length,
and natural and fishing mortalities-at-age/length by applying the catch
equation
\deqn{C = N \cdot F \frac{F}{M+F} \cdot (1 - {\rm e}^(-M-F))}{C = N *F/(M+F) * (1-exp(-M-F))}
}
\examples{
data(ple4)
res <- catch.n(stock.n(ple4), harvest(ple4), m(ple4))
catch.n(ple4) / res
}
\seealso{
\linkS4class{FLStock}
}
\author{
The FLR Team
}
\keyword{methods}
