% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{catchInmature}
\alias{catchInmature}
\alias{catchMature}
\title{Proportion of mature and inmature fish in the catch}
\usage{
catchInmature(object)

catchMature(object)
}
\arguments{
\item{object}{An age-structured FLStock object}
}
\value{
An FLQuant object
}
\description{
The proportion in weight of mature and inmature fish in the catch can
be computed using catchMature and catchInmature.
}
\examples{
data(ple4)
catchInmature(ple4)
catchMature(ple4)
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{ts}
