% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\docType{methods}
\name{coerce-methods}
\alias{coerce-methods}
\title{Convert Objects Between Classes}
\arguments{
\item{object}{Object to be converted.}

\item{Class}{Name of the class to convert the object to, \code{character}.}
}
\value{
An object of the requested class.
}
\description{
Objects of various \strong{FLCore} classes can be converted into other classes,
both basic R ones, like \code{data.frame}, and others defined in the package. For
the specifics of the precise calculations carried out for each pair of
classes, see below.
}
\section{FLArray to data.frame}{

 The six dimensions of an `FLArray` are converted into seven columns, named
 `quant` (or any other name given to the first dimension in the object),
 `year`, `unit`, `season`, `area`, `iter`  and `data`. The last one contains
 the actual numbers stored in the array. `units` are stored as an attribute
 to the `data.frame`. The `year` and `data` columns are of type `numeric`,
 while all others are `factor`.
}

\section{FLPar to data.frame}{

 The two or more dimensions of an *FLPar* objects are converted into three or
 more columns. For a 2D objects, they are named *params*, *iter* and *data*.
 The last one contains the actual numbers stored in the array, in a column
 type `numeric`, while all others are `factor`.
}

\examples{
# from FLQuant to data.frame
as(FLQuant(rnorm(100), dim=c(5, 20)), "data.frame")
# from FLPar to data.frame
as(FLPar(phi=rnorm(10), rho=rlnorm(10)), "data.frame")
}
\seealso{
\link[base:as]{base::as}, \link[base:coerce]{base::coerce}
}
\author{
The FLR Team
}
\keyword{methods}
