% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLlst-methods.R
\docType{methods}
\name{collapse}
\alias{collapse}
\alias{collapse-methods}
\alias{collapse,FLlst-method}
\title{Collapse the iterations stored across an FLlst object}
\usage{
collapse(x, ...)
}
\arguments{
\item{x}{An \code{\link{FLlst}} object whose elements share compatible dimensions.}

\item{...}{Unused.}
}
\value{
An \code{\link{FLQuant}} containing all iterations from the elements in
\code{x}.
}
\description{
\code{collapse()} concatenates the iterations contained in each element of an
\code{\link{FLlst}} object into a single \code{\link{FLQuant}}.
}
\seealso{
\linkS4class{FLlst}, \link{lock}, \link{iter}, \link{propagate}
}
\author{
The FLR Team
}
\keyword{methods}
