% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{compare}
\alias{compare}
\title{A method for comparing FLR objects}
\usage{
compare(result, target, ...)
}
\arguments{
\item{result}{First element in comparison, result of method or operation.}

\item{target}{Second element, desired output.}
}
\value{
A table of comparisons, of class data.frame.
}
\description{
Comparisons of complete objects of FLR classes can be carried out and
a report table is generated to better identify differences. Comparisons
do not substitute but complement those provided by R's all.equal and
identical.
}
\author{
Iago Mosqueira (WMR)
}
\keyword{methods}
