% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLModelDeriv.R
\docType{methods}
\name{computeD}
\alias{computeD}
\alias{computeD-methods}
\alias{computeD,FLModel-method}
\title{Compute first and second derivative terms for an FLModel}
\usage{
computeD(object, ...)

\S4method{computeD}{FLModel}(object, params = as(object@params, "list"),
  method = "Richardson", eps = 1e-04, d = 1e-04,
  zero.tol = sqrt(.Machine$double.eps/7e-07), r = 4, v = 2)
}
\arguments{
\item{object}{An \code{\link{FLModel}} object.}

\item{params}{Parameter values passed to the model likelihood. Defaults to
\code{as(object@params, "list")}.}

\item{method}{Numerical differentiation method. Currently only Richardson
extrapolation is implemented.}

\item{eps}{A small perturbation used in the numerical approximation.}

\item{d}{Relative step size used to initialise perturbations.}

\item{zero.tol}{Threshold used to treat values as zero.}

\item{r}{Number of Richardson extrapolation steps.}

\item{v}{Reduction factor applied to the perturbation at each Richardson
iteration.}
}
\value{
A numeric matrix containing first derivatives and second derivative terms.
}
\description{
\code{computeD()} numerically approximates first derivatives and the lower
triangle of the Hessian for an \code{\link{FLModel}} object, returning the
intermediate matrix used by \code{\link{computeHessian}()}.
}
\seealso{
\link{computeHessian}, \link{gradient}, \linkS4class{FLModel}
}
\author{
The FLR Team
}
\keyword{methods}
