% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{computeHarvest,FLStock,missing-method}
\alias{computeHarvest,FLStock,missing-method}
\title{Computes fishing mortality from abundances, catches and natural mortality}
\usage{
\S4method{computeHarvest}{FLStock,missing}(object, units = NULL)
}
\arguments{
\item{units}{Harvest to be computed as 'f' or 'hr', 'character'.}

\item{x}{An object of class 'FLStock'.}
}
\value{
An 'FLQuant' with the calculated fishing mortalities at age.
}
\description{
Objects or class 'FLStock' already contain a 'harvest' slot to store
estimates of fishing mortality at age, for example those obtained from
a stock assessment method. Fishing mortality at age can be recalculated
using two methods:
}
\examples{
data(ple4)
# Compute 'f' from stock.n and Baranov
computeHarvest(ple4)
# Recomputes all F at age by solving catch Baranov
recomputeHarvest(ple4)
}
\seealso{
\linkS4class{FLStock} \code{\link[=harvest]{harvest()}} \linkS4class{FLQuant}
}
\author{
The FLR Team
}
\keyword{manip}
