% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLModelDeriv.R
\docType{methods}
\name{computeHessian}
\alias{computeHessian}
\alias{computeHessian-methods}
\alias{computeHessian,FLModel-method}
\title{Compute the Hessian matrix for an FLModel}
\usage{
computeHessian(object, ...)

\S4method{computeHessian}{FLModel}(object, eps = 1e-04, d = 1e-04,
  zero.tol = sqrt(.Machine$double.eps/7e-07), r = 4, v = 2)
}
\arguments{
\item{object}{An \code{\link{FLModel}} object.}

\item{eps}{A small perturbation used in the numerical approximation.}

\item{d}{Relative step size used to initialise perturbations.}

\item{zero.tol}{Threshold used to treat values as zero.}

\item{r}{Number of Richardson extrapolation steps.}

\item{v}{Reduction factor applied to the perturbation at each Richardson
iteration.}
}
\value{
An array containing the approximated Hessian matrix, with parameter names used
as dimnames.
}
\description{
\code{computeHessian()} numerically approximates the Hessian matrix of the
likelihood for an \code{\link{FLModel}} object.
}
\seealso{
\link{computeD}, \link{gradient}, \linkS4class{FLModel}
}
\author{
The FLR Team
}
\keyword{methods}
