% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\docType{methods}
\name{computeQ}
\alias{computeQ}
\alias{computeQ-methods}
\alias{computeQ,FLI,FLStock,FLQuant-method}
\title{Compute catchability from model fit and observations}
\usage{
\S4method{computeQ}{FLIndices,FLStock,FLQuants}(indices, stock, fit)

\S4method{computeQ}{FLI,FLStock,FLQuant}(indices, stock, fit)
}
\arguments{
\item{indices}{An \code{FLIndices} or \code{FLI} (the virtual base class
shared by \code{FLIndex} and \code{FLIndexBiomass}) object containing the
survey index definitions (timing, selectivity pattern, etc.).}

\item{stock}{An \code{FLStock} operating model providing stock numbers and
weight-at-age.}

\item{fit}{An \code{FLQuants} or \code{FLQuant} of fitted index values from
the stock assessment model.}

\item{...}{Additional arguments (unused).}
}
\value{
An \code{FLQuants} or \code{FLQuant} of estimated catchability
coefficients, one per index (or age for biomass indices).
}
\description{
Estimates the catchability coefficient (\eqn{Q}) for one or more survey
indices given an \code{FLStock} operating model and the fitted index values
from a stock assessment. The calculation accounts for survey timing,
correcting stock numbers for total mortality (\eqn{Z}) at the time of the
survey.
}
\section{Generic function}{
 computeQ(indices, stock, fit, ...)
}

\examples{
# Load datasets
data(ple4)
data(ple4.indices)
# Compute for each index and pseudo-noisy fit
pseudofit <- lapply(ple4.indices, function(x) index(x) * 0.8)
computeQ(ple4.indices, ple4, pseudofit)
#  Compute for single index and pseudo-noisy fit
computeQ(ple4.indices[[1]], ple4, rlnorm(1, log(index(ple4.indices[[1]])), 0.1))
}
\seealso{
\link{survey}, \linkS4class{FLIndex}, \linkS4class{FLStock}
}
\author{
The FLR Team
}
\keyword{methods}
