% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\docType{methods}
\name{cpue}
\alias{cpue}
\alias{cpue-methods}
\alias{cpue,FLStock,missing-method}
\alias{cpue,FLStock-method}
\title{Compute a CPUE index of abundance from an FLStock}
\usage{
cpue(object, index, ...)

\S4method{cpue}{FLStock,missing}(
  object,
  sel.pattern = harvest(object),
  effort = units(harvest(object)),
  biomass = TRUE
)
}
\arguments{
\item{object}{An \code{FLStock} from which to generate the CPUE observation.}

\item{index}{An optional index object (currently unused for the
\code{FLStock, missing} method).}

\item{...}{Additional arguments passed to methods.}

\item{sel.pattern}{Selectivity pattern to apply; defaults to
\code{harvest(object)}.}

\item{effort}{Units of effort used to mimic the effort series in the
fishery: \code{"f"} (default, fishing mortality), \code{"hr"} (harvest
rate), or a numeric value.}

\item{biomass}{Logical; if \code{TRUE} (default) the index is summed across
ages and weighted by catch weight-at-age to give a biomass CPUE.}
}
\value{
An \code{FLQuant} containing the CPUE index: age-disaggregated
if \code{biomass = FALSE}, or aggregated biomass CPUE if
\code{biomass = TRUE}.
}
\description{
The observation of stock abundance by CPUE series from commercial fleets is an
important step in the generation of management advice that needs to be replicated
on an Operating Model during any simulation exercise. This method generates
an observation of biomass or numbers-at-age from an \code{FLStock} being used as OM.
}
\section{Generic function}{
 cpue(object, index, ...)
}

\examples{

data(ple4)

cpue(ple4)
# An aggregated biomass CPUE
quantSums(cpue(ple4))

\dontrun{
plot(FLQuants(om=stock(ple4), cpue=quantSums(cpue(ple4)),
  hr=quantSums(cpue(ple4, effort="hr"))))
}
}
\seealso{
\link{survey}, \linkS4class{FLStock}
}
\author{
The FLR Team
}
\keyword{methods}
