% $Id: data.Rd,v 1.5.2.1 2005/12/19 11:13:48 iagoazti Exp $
\name{data}
\alias{ple4}
\alias{bt4}
\alias{ple4sex}
\alias{ple4.biol}
\alias{ple4.index}
\title{FLCore datasets}
\description{
    \item{ple4}{A dataset of North Sea (ICES Area IV) plaice catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
    \item{ple4sex}{A dataset of North Sea (ICES Area IV) plaice disaggregated by sex catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
    \item{ple4.index}{A dataset of North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndex object.}
    \item{ple.biol}{A dataset of North Sea plaice population, numbers, natural mortality, mass and fecundity-at-age, contained in an FLBiol object.}

	Datasets can be loaded by issuing \code{data(ple4)} and \code{data(bt4)} respectively.
}
\seealso{
\code{\link{FLStock}}, \code{\link{FLSR}}, \code{\link{FLIndex}, \code{\link{FLStock}}, \code{\link{FLIndex}}, \code{\link{FLBiol}}}
}
\references{
ICES.
}
\keyword{datasets}
