% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{datasets}
\alias{datasets}
\alias{nsher}
\alias{ple4}
\alias{ple4.biol}
\alias{ple4.index}
\alias{ple4.indices}
\alias{ple4sex}
\title{FLCore datasets}
\description{
Example datasets for the classes defined in FLCore.
}
\details{
\itemize{
   \item{\code{ple4}, \code{\link{FLStock}}}{A dataset for North Sea (ICES Area IV)
     plaice. Catch, landings, discards, natural mortality, weight-at-age and maturity,
     together with the VPA estimated abundances and fishing mortalities.}
\item{\code{ple4sex}, \code{\link{FLStock}}}{A dataset of North Sea (ICES
     Area IV) plaice disaggregated by sex. Catch, yield, landings, discards,
     natural mortality, weight-at-age and maturity, together with the VPA
     estimated abundances and fishing mortalities.}
\item{ple4.index, \code{\link{FLIndex}}}{A dataset of North Sea (ICES Area IV) plaice
     survey catch per unit effort, index and index variance.}
\item{ple4.indices, \code{\link{FLIndices}}}{A dataset of three North Sea (ICES Area IV)
     plaice survey catch per unit effort series. Index and index variance.}
\item{ple4.biol, \code{\link{FLBiol}}}{A dataset of the North Sea plaice population.
     Numbers, natural mortality, mass and fecundity-at-age.}
\item{nsher , \code{\link{FLSR}}}{Stock and recruit data and fitted relationship for
     autumn spawning North Sea herring.}
}

Datasets can be loaded by issuing the \code{data} command, like in
\code{data(ple4)}.
}
\examples{

data(ple4)
summary(ple4)

data(nsher)
is(nsher)

}
\references{
ICES.
}
\seealso{
\linkS4class{FLStock}, \linkS4class{FLSR}, \linkS4class{FLIndex},
\linkS4class{FLStock}, \linkS4class{FLIndex}, \linkS4class{FLBiol}
}
\keyword{datasets}

