% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{discardsRatio}
\alias{discardsRatio}
\title{Compute the ratio of discards to total catch in numbers or weight}
\usage{
discardsRatio(object, value = c("numbers", "weight", "total"))
}
\arguments{
\item{object}{An object of class 'FLStock'}

\item{value}{One of 'numbers' (default), 'weight' or 'total'.}
}
\value{
The discards ratio (between 0 and 1), 'FLQuant'
}
\description{
A calculation is made of the proportion of discards over total catch at age,
either as numbers (value = 'numbers') or weight (value = 'weight'), or for
the total discards and catch in biomass (value = 'total').
}
\examples{
data(ple4)
# Discards ratio at age in numbers
discardsRatio(ple4)
# Total proportion of discards by year
discardsRatio(ple4, value="total")
}
\seealso{
\linkS4class{FLStock}
}
\author{
The FLR Team
}
\keyword{arith}
