% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{.biomass}
\alias{.biomass}
\title{FLStodck biomass metrics}
\usage{
.biomass(n, h, ph, m, pm, wt, sel, byage = FALSE)
}
\description{
FLStodck biomass metrics
}
\examples{
data(ple4)
# SSB at spawning time
ssb(ple4)
# SSB at end of time step (year)
ssb_end(ple4)
# SSB at start of time step (year)
ssb_start(ple4)
# Vulnerable biomass at start of time step (year)
vb(ple4)
# Exploitable biomass at start of time step (year)
exb(ple4)
}
