% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R
\docType{methods}
\name{drop,FLArray-method}
\alias{drop,FLArray-method}
\alias{drop,FLQuant-method}
\title{drop method for FLCore array-based classes}
\usage{
\S4method{drop}{FLArray}(x)
}
\description{
Delete the dimensions of an array which have only one level.
}
\details{
This method calls R's \link[base:drop]{base::drop} on the \code{@.Data} slot of an \link{FLArray}.
Dimensions of length one are thus dropped, as is the class attribute and the
\code{units} slot, and an array of equal or less
dimensions, a matrix or a vector is returned.

On an FLQuant object with
}
\examples{
x <- FLQuant(1:3, dim=c(3,3))
drop(x)
is(drop(x))
dim(drop(x))

# Result of drop can be used for matrix algebra
# for example to calculate aging error

data(ple4)
aging.error <- diag(0.8, 10)
diag(aging.error[-1,]) <- c(rep(0.1, 8), 0.2)
diag(aging.error[, -1]) <- c(0.2, rep(0.1, 8))
t(aging.error) \%*\% drop(catch.n(ple4))
}
\seealso{
\link[base:drop]{base::drop}
}
\author{
The FLR Team
}
\keyword{methods}
