% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\docType{methods}
\name{exb}
\alias{exb}
\alias{exb-methods}
\alias{exb,FLStock-method}
\title{Calculate exploitable biomass}
\usage{
exb(x, ...)

\S4method{exb}{FLStock}(x, sel = catch.sel(x), wt = catch.wt(x), byage = FALSE, ...)
}
\arguments{
\item{x}{An \code{\link{FLStock}} object.}

\item{sel}{Selection pattern to apply, defaulting to \code{\link{catch.sel}(x)}.}

\item{wt}{Mean weight to apply, defaulting to \code{\link{catch.wt}(x)}.}

\item{byage}{Should exploitable biomass be returned by age rather than
aggregated over age?}

\item{...}{Additional named slot values temporarily used in the calculation.}
}
\value{
An \code{\link{FLQuant}} containing exploitable biomass.
}
\description{
\code{exb()} returns exploitable biomass for an \code{\link{FLStock}} object
by applying the selected catchability pattern and weights to stock numbers.
}
\details{
Set \code{byage = TRUE} to keep the age structure instead of aggregating over
age.
}
\examples{
data(ple4)
exb(ple4)
}
\seealso{
\link{biomass}, \link{vb}, \link{ssb}, \linkS4class{FLStock}
}
\author{
The FLR Team
}
\keyword{methods}
