% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R
\name{exp,FLQuant-method}
\alias{exp,FLQuant-method}
\alias{log,FLQuant-method}
\title{exp and log methods FLCore array-based classes}
\usage{
\S4method{exp}{FLQuant}(x)

\S4method{log}{FLQuant}(x, ...)
}
\description{
Compute the exponential and logarithmic functions
}
\details{
This method simply calls R's \link[base:exp]{base::exp} and \link[base:drop]{base::drop}, but take care of
returning the right units of measurement, that is "" or character(1).
}
\examples{
x <- FLQuant(c(4,2,7,4,2,9), units="1000")
log(x)
units(log(x))
}
\seealso{
\link[base:exp]{base::exp} \link[base:log]{base::log}
}
\author{
The FLR Team
}
\keyword{methods}
