% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{ffwd}
\alias{ffwd}
\title{Project forward an FLStock for a fbar target}
\usage{
ffwd(object, sr, fbar = control, control = fbar, deviances = "missing")
}
\arguments{
\item{object}{An \emph{FLStock}}

\item{sr}{A stock-recruit relationship, \emph{FLSR} or \emph{predictModel}.}

\item{fbar}{Yearly target for average fishing mortality, \emph{FLQuant}.}

\item{control}{Yearly target for average fishing mortality, \emph{fwdControl}.}

\item{deviances}{Deviances for the strock-recruit relationsip, \emph{FLQuant}.}
}
\value{
The projected \emph{FLStock} object.
}
\description{
Projection of an FLStock object for a fishing mortality target does not
always require the features of fwd().Fast-forward an FLStock object for a fishing mortality yearly target only.
}
\examples{
data(ple4)
sr <- predictModel(model=bevholt, params=FLPar(a=140.4e4, b=1.448e5))
# Project for fixed Fbar=0.21
run <- ffwd(ple4, sr=sr, fbar=FLQuant(0.21, dimnames=list(year=1958:2017)))
plot(run)
}
\seealso{
\link{fwd}
}
\author{
Iago MOSQUEIRA (MWR), Henning WINKEL (JRC).
}
\keyword{models}
