\name{FLCatch-accesors}
\alias{FLCatch-accesors}
\docType{methods}
% catch
\alias{catch,FLCatch-method}
\alias{catch,FLCatches-method}
\alias{catch,FLMetier-method}
\alias{catch,FLMetiers-method}
\alias{catch,FLFleet-method}
% catch<-
\alias{catch<-,FLCatch,FLQuant-method}
\alias{catch<-,FLCatches,FLQuant-method}
\alias{catch<-,FLMetier,FLQuant-method}
\alias{catch<-,FLMetiers,FLQuant-method}
\alias{catch<-,FLFleet,FLQuant-method}
%catch.n
\alias{catch.n,FLCatch-method}
\alias{catch.n,FLCatches-method}
\alias{catch.n,FLMetier-method}
\alias{catch.n,FLMetiers-method}
\alias{catch.n,FLFleet-method}
% catch.n<-
\alias{catch.n<-,FLCatch,FLQuant-method}
\alias{catch.n<-,FLCatches,FLQuant-method}
\alias{catch.n<-,FLMetier,FLQuant-method}
\alias{catch.n<-,FLMetiers,FLQuant-method}
\alias{catch.n<-,FLFleet,FLQuant-method}
%catch.wt
\alias{catch.wt,FLCatch-method}
\alias{catch.wt,FLCatches-method}
\alias{catch.wt,FLMetier-method}
\alias{catch.wt,FLMetiers-method}
\alias{catch.wt,FLFleet-method}
% catch.wt<-
\alias{catch.wt<-,FLCatch,FLQuant-method}
\alias{catch.wt<-,FLCatches,FLQuant-method}
\alias{catch.wt<-,FLMetier,FLQuant-method}
\alias{catch.wt<-,FLMetiers,FLQuant-method}
\alias{catch.wt<-,FLFleet,FLQuant-method}
% catch.q
\alias{catch.q,FLCatch-method}
\alias{catch.q,FLCatches-method}
\alias{catch.q,FLMetier-method}
\alias{catch.q,FLMetiers-method}
\alias{catch.q,FLFleet-method}
 % catch.q
\alias{catch.q<-,FLCatch,FLQuant-method}
\alias{catch.q<-,FLCatches,FLQuant-method}
\alias{catch.q<-,FLMetier,FLQuant-method}
\alias{catch.q<-,FLMetiers,FLQuant-method}
\alias{catch.q<-,FLFleet,FLQuant-method}
% catch.sel
\alias{catch.sel,FLCatch-method}
\alias{catch.sel,FLCatches-method}
\alias{catch.sel,FLMetier-method}
\alias{catch.sel,FLMetiers-method}
\alias{catch.sel,FLFleet-method}
% catch.sel<-
\alias{catch.sel<-,FLCatch,FLQuant-method}
\alias{catch.sel<-,FLCatches,FLQuant-method}
\alias{catch.sel<-,FLMetier,FLQuant-method}
\alias{catch.sel<-,FLMetiers,FLQuant-method}
\alias{catch.sel<-,FLFleet,FLQuant-method} 
% catches
\alias{catches,FLFleet-method}
\alias{catches,FLMetiers-method}
\alias{catches,FLMetier-method}
% catches<-
\alias{catches<-,FLMetier,ANY,FLCatch-method}
\alias{catches<-,FLMetier,ANY,FLCatches-method}
% discards
\alias{discards,FLCatch-method}
\alias{discards,FLCatches-method}
\alias{discards,FLMetier-method}
\alias{discards,FLMetiers-method}
\alias{discards,FLFleet-method}
% discards<-
\alias{discards<-,FLCatch,FLQuant-method}
\alias{discards<-,FLCatches,FLQuant-method}
\alias{discards<-,FLMetier,FLQuant-method}
\alias{discards<-,FLMetiers,FLQuant-method}
\alias{discards<-,FLFleet,FLQuant-method}
%discards.n
\alias{discards.n,FLCatch-method}
\alias{discards.n,FLCatches-method}
\alias{discards.n,FLMetier-method}
\alias{discards.n,FLMetiers-method}
\alias{discards.n,FLFleet-method}
% discards.n<-
\alias{discards.n<-,FLCatch,FLQuant-method}
\alias{discards.n<-,FLCatches,FLQuant-method}
\alias{discards.n<-,FLMetier,FLQuant-method}
\alias{discards.n<-,FLMetiers,FLQuant-method}
\alias{discards.n<-,FLFleet,FLQuant-method}
%discards.wt
\alias{discards.wt,FLCatch-method}
\alias{discards.wt,FLCatches-method}
\alias{discards.wt,FLMetier-method}
\alias{discards.wt,FLMetiers-method}
\alias{discards.wt,FLFleet-method}
% discards.wt<-
\alias{discards.wt<-,FLCatch,FLQuant-method}
\alias{discards.wt<-,FLCatches,FLQuant-method}
\alias{discards.wt<-,FLMetier,FLQuant-method}
\alias{discards.wt<-,FLMetiers,FLQuant-method}
\alias{discards.wt<-,FLFleet,FLQuant-method}
% discards.q
\alias{discards.q,FLCatch-method}
\alias{discards.q,FLCatches-method}
\alias{discards.q,FLMetier-method}
\alias{discards.q,FLMetiers-method}
\alias{discards.q,FLFleet-method}
 % discards.q
\alias{discards.q<-,FLCatch,FLQuant-method}
\alias{discards.q<-,FLCatches,FLQuant-method}
\alias{discards.q<-,FLMetier,FLQuant-method}
\alias{discards.q<-,FLMetiers,FLQuant-method}
\alias{discards.q<-,FLFleet,FLQuant-method}
% discards.sel
\alias{discards.sel,FLCatch-method}
\alias{discards.sel,FLCatches-method}
\alias{discards.sel,FLMetier-method}
\alias{discards.sel,FLMetiers-method}
\alias{discards.sel,FLFleet-method}
% discards.sel<-
\alias{discards.sel<-,FLCatch,FLQuant-method}
\alias{discards.sel<-,FLCatches,FLQuant-method}
\alias{discards.sel<-,FLMetier,FLQuant-method}
\alias{discards.sel<-,FLMetiers,FLQuant-method}
\alias{discards.sel<-,FLFleet,FLQuant-method} 
% landings
\alias{landings,FLCatch-method}
\alias{landings,FLCatches-method}
\alias{landings,FLMetier-method}
\alias{landings,FLMetiers-method}
\alias{landings,FLFleet-method}
% landings<-
\alias{landings<-,FLCatch,FLQuant-method}
\alias{landings<-,FLCatches,FLQuant-method}
\alias{landings<-,FLMetier,FLQuant-method}
\alias{landings<-,FLMetiers,FLQuant-method}
\alias{landings<-,FLFleet,FLQuant-method}
%landings.n
\alias{landings.n,FLCatch-method}
\alias{landings.n,FLCatches-method}
\alias{landings.n,FLMetier-method}
\alias{landings.n,FLMetiers-method}
\alias{landings.n,FLFleet-method}
% landings.n<-
\alias{landings.n<-,FLCatch,FLQuant-method}
\alias{landings.n<-,FLCatches,FLQuant-method}
\alias{landings.n<-,FLMetier,FLQuant-method}
\alias{landings.n<-,FLMetiers,FLQuant-method}
\alias{landings.n<-,FLFleet,FLQuant-method}
%landings.wt
\alias{landings.wt,FLCatch-method}
\alias{landings.wt,FLCatches-method}
\alias{landings.wt,FLMetier-method}
\alias{landings.wt,FLMetiers-method}
\alias{landings.wt,FLFleet-method}
% landings.wt<-
\alias{landings.wt<-,FLCatch,FLQuant-method}
\alias{landings.wt<-,FLCatches,FLQuant-method}
\alias{landings.wt<-,FLMetier,FLQuant-method}
\alias{landings.wt<-,FLMetiers,FLQuant-method}
\alias{landings.wt<-,FLFleet,FLQuant-method}
% landings.q
\alias{landings.q,FLCatch-method}
\alias{landings.q,FLCatches-method}
\alias{landings.q,FLMetier-method}
\alias{landings.q,FLMetiers-method}
\alias{landings.q,FLFleet-method}
 % landings.q
\alias{landings.q<-,FLCatch,FLQuant-method}
\alias{landings.q<-,FLCatches,FLQuant-method}
\alias{landings.q<-,FLMetier,FLQuant-method}
\alias{landings.q<-,FLMetiers,FLQuant-method}
\alias{landings.q<-,FLFleet,FLQuant-method}
% landings.sel
\alias{landings.sel,FLCatch-method}
\alias{landings.sel,FLCatches-method}
\alias{landings.sel,FLMetier-method}
\alias{landings.sel,FLMetiers-method}
\alias{landings.sel,FLFleet-method}
% landings.sel<-
\alias{landings.sel<-,FLCatch,FLQuant-method}
\alias{landings.sel<-,FLCatches,FLQuant-method}
\alias{landings.sel<-,FLMetier,FLQuant-method}
\alias{landings.sel<-,FLMetiers,FLQuant-method}
\alias{landings.sel<-,FLFleet,FLQuant-method} 
% price
\alias{price,FLCatch-method}
\alias{price,FLCatches-method}
\alias{price,FLMetier-method}
\alias{price,FLMetiers-method}
\alias{price,FLFleet-method}
% price<-
\alias{price<-,FLCatch,FLQuant-method}
\alias{price<-,FLCatches,FLQuant-method}
\alias{price<-,FLMetier,FLQuant-method}
\alias{price<-,FLMetiers,FLQuant-method}
\alias{price<-,FLFleet,FLQuant-method}

\title{Accesor and replacement methods for the FLCatch class}
\description{
	Slots in object of class FLCatch can be accessed and modified using these accesor
	and replacement methods. Methods share their names with the slots they apply to. Slots
	are accesible directly from an FLCatch object but also through the \link{FLFleet},
	\link{FLMetiers}, \link{FLMetier} and \link{FLCatches} object where the \link{FLCatch}
	object is itself contained, by specifying the \code{spp} and/or \code{metier}
	arguments.
}
\section{Generic function}{
catch(object) 
catch<-(object,value) 
catch.n(object) 
catch.n<-(object,value) 
catch.wt(object) 
catch.wt<-(object,value) 
catch.q(object) 
catch.q<-(object,value) 
catch.sel(object) 
catch.sel<-(object,value) 
effort(object) 
effort<-(object,value) 
price(object) 
price<-(object,value) 
}
\section{Methods}{\describe{
\item{ signature(object=FLCatch) :}{Describe method}
\item{ signature(object=FLCatches) :}{Describe method}
\item{ signature(object=FLMetier) :}{Describe method}
\item{ signature(object=FLMetiers) :}{Describe method}
\item{ signature(object=FLFleet) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLFleet}, \link{FLMetiers}, \link{FLMetier}, \link{FLCatches},
	\link{FLCatch}}
\examples{
data(bt4)
catch.n(bt4, 'TBB', 'ple')
}
\keyword{methods}
