% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/classesLst.R
\docType{class}
\name{FLCohorts}
\alias{FLCohorts}
\alias{FLCohorts,ANY-method}
\alias{FLCohorts,FLCohorts-method}
\alias{FLCohorts,list-method}
\alias{FLCohorts,missing-method}
\alias{FLCohorts-class}
\alias{FLCohorts-methods}
\title{Class FLCohorts}
\usage{
FLCohorts(object, ...)
}
\arguments{
\item{object}{unnamed object to be added to the list}

\item{...}{other named or unnamed objects}
}
\description{
\code{FLCohorts} is a class that extends \code{list} through \code{FLlst}
but implements a set of features that give a little more structure to
list objects. The elements of \code{FLCohorts} must all be of class
\code{FLCohort}. It implements a lock mechanism that, when turned on, does
not allow the user to increase or decrease the object length.
}
\section{Slots}{
 \describe{ \item{.Data}{The data. \code{list}}
\item{names}{Names of the list elements. \code{character}}
\item{desc}{Description of the object. \code{character}} \item{lock}{Lock
mechanism, if turned on the length of the list can not be modified by adding
or removing elements. \code{logical}} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}
\author{
The FLR Team
}
\seealso{
\link[base]{*}, \link[methods]{Arith}, \link[base]{as.data.frame},
\link{bubbles}, \link{catch<-}, \link{iter}, \link[stats]{model.frame},
\link[methods]{show}, \link[base]{summary}, \link[lattice]{xyplot},
\link{FLlst}, \link[base]{list}
}
\keyword{classes}

