% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesLst.R
\docType{class}
\name{FLComps}
\alias{FLComps}
\alias{FLComps-class}
\title{Class FLComps}
\arguments{
\item{object}{unnamed object to be added to the list}

\item{...}{other named or unnamed objects}
}
\description{
A virtual class that forms the basis for many FLR list classes. No objects
of this class can be constructed.
}
\section{Validity}{
 \describe{
    \item{Elements}{All elements must be of a class that inherits from FLComp}
}
}

\section{Slots}{
 \describe{
  \item{.Data}{The data. \code{list}.}
  \item{names}{Names of the list elements. \code{character}.}
  \item{desc}{Description of the object. \code{character}.}
  \item{lock}{Lock mechanism, if turned on the length of the list can not be
    modified by adding or removing elements. \code{logical}.} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}

\seealso{
\link{FLlst}, \link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
