\name{FLCore Generic Functions}
\docType{methods}
\alias{FLGenerics-methods}
\alias{FLGenerics}
\title{FLCore S4 Generic Functions}
\description{This is a list of S4 generic functions implemented in FLCore. If a generic function is already in FLCore the package developers should not created it again.}
\section{Generic Function}{
\describe{
\item{revenue}{function(object,...)}
\item{as.FLFleet}{function(x)}
\item{as.FLIndex}{function(object, ...)}
\item{trim}{function(object, ...)}
\item{add<-}{function(object,value)}
\item{quant}{function(object, ...)}
\item{quant<-}{function(object, value)}
\item{units}{function(object, ...)}
\item{units<-}{function(object, value)}
\item{names}{function(x)}
\item{dims}{function(obj, ...)}
\item{summary}{useAsDefault = summary}
\item{plot}{useAsDefault = plot}
\item{xyplot}{useAsDefault = xyplot}
\item{bwplot}{useAsDefault = bwplot}
\item{dotplot}{useAsDefault = dotplot}
\item{barchart}{useAsDefault = barchart}
\item{stripplot}{useAsDefault = stripplot}
\item{histogram}{useAsDefault = histogram}
\item{bubbles}{function(x, data, ...)}
\item{as.data.frame}{function(x, row.names, optional)}
\item{apply}{function(X, MARGIN, FUN, ...)}
\item{window}{function(x, ...)}
\item{trim}{function(obj, ...)}
\item{predict}{function(object, ...)}
\item{residuals}{function(object, ...)}
\item{fitted}{function(object, ...)}
\item{params}{function(object, ...)}
\item{params<-}{function(object, ..., value)}
\item{landings}{function(object, ...)}
\item{discards}{function(object, ...)}
\item{catch}{function(object, ...)}
\item{window}{function(x, ...)}
\item{ssb}{function(object, ...)}
\item{setunits}{function(object, ...)}
\item{transform}{function(x, ...)}
\item{as.FLStock}{function(object, ...)}
\item{harvest}{function(object, ...)}
\item{harvest<-}{function(object, value)}
\item{trim}{function(object, ...)}
}
}
\author{
	FLR Team	
}
\keyword{methods}
