% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCore-package.R
\docType{package}
\name{FLCore-package}
\alias{FLCore}
\alias{FLCore-package}
\title{Core package of FLR, fisheries modelling in R.}
\description{
FLCore contains the core classes and methods for FLR, a framework for
fisheries modelling and management strategy simulation in R. Developed by a
team of fisheries scientists in various countries. More information can be
found at http://flr-project.org/, including a development mailing list.
}
\details{
\tabular{ll}{ Package: \tab FLCore\cr Version: \tab 3.0\cr Date: \tab
2009\cr Depends: \tab methods, R(>= 2.8.1), graphics, stats, stats4, grid,
lattice\cr License: \tab GPL 2 or above\cr LazyLoad: \tab Yes\cr LazyData:
\tab Yes\cr Built: \tab R 2.8.1; i686-pc-linux-gnu; 2009-01-22 12:08:46;
unix\cr }
}
\author{
FLR Team and various contributors. Initial design by Laurence T.
Kell & Philippe Grosjean.
}
\references{
Website at \url{http://flr-project.org/}.

Kell L.T., Mosqueira I., Grosjean P., Fromentin J-M., Garcia D., Hillary R.,
Jardim E., Pastoors M., Poos J.J., Scott F. & Scott R.D. 2007.  FLR: an
open-source framework for the evaluation and development of management
strategies. ICES J. of Mar. Sci. 20: 289-290. doi: 10.1093/icesjms/fsm012
}
\keyword{package}

