% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLGrowth.R
\docType{class}
\name{FLGrowth}
\alias{FLGrowth}
\alias{FLGrowth-class}
\alias{FLGrowth-methods}
\title{Class FLGrowth}
\usage{
FLGrowth(model, ...)
}
\description{
A class for growth models.
}
\section{Slots}{
 \describe{
  \item{mass}{Mass (weight) at age. \code{FLArray}.} }
}
\examples{

data(ple4)
FLGrowth(mass=stock.wt(ple4)[,1])

}
\author{
The FLR Team
}
\seealso{
\link{FLModel},
}
\keyword{classes}

