% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesComp.R
\docType{class}
\name{FLI}
\alias{FLI}
\alias{FLI-class}
\title{Class FLI}
\arguments{
\item{object}{FLQuant object used for sizing}

\item{...}{Other objects to be assigned by name to the class slots}
}
\description{
A VIRTUAL class that holds data and parameters related to abundance indices.
}
\section{Slots}{
 \describe{
    \item{distribution}{Statistical distribution of the index values (\code{character}).}
    \item{index}{Index values (\code{FLQuant}).}
    \item{index.var}{Variance of the index (\code{FLQuant}).}
    \item{catch.n}{Catch numbers used to create the index (\code{FLQuant}).}
    \item{catch.wt}{Catch weight of the index (\code{FLQuant}).}
    \item{effort}{Effort used to create the index (\code{FLQuant}).}
    \item{sel.pattern}{Selection pattern for the index (\code{FLQuant}).}
    \item{index.q}{Catchability of the index (\code{FLQuant}).}
    \item{name}{Name of the stock (\code{character}).}
    \item{desc}{General description of the object (\code{character}).}
    \item{range}{Range of the object (\code{numeric})}
}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Validity}{
 \describe{
    \item{Dimensions}{All FLQuant slots must have iters equal to 1 or 'n'.}
    \item{Iters}{The dimname for iter[1] should be '1'.}
    \item{Dimnames}{The name of the quant dimension must be the same for all FLQuant slots.}
}
}

\seealso{
\link{computeCatch}, \link{dims},
\link{iter}, \link[graphics]{plot}, \link[FLCore]{propagate},
\link[base]{summary}, \link[base]{transform}, \link{trim},
\link[stats]{window}, \link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
