% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/classesComp.R, R/FLIndex.R
\docType{class}
\name{FLIndex}
\alias{FLIndex}
\alias{FLIndex,FLQuant-method}
\alias{FLIndex,missing-method}
\alias{FLIndex-class}
\alias{FLIndex-methods}
\alias{catch.n,FLIndex-method}
\alias{catch.n<-,FLIndex,FLQuant-method}
\alias{catch.n<-,FLIndex,numeric-method}
\alias{catch.wt,FLIndex-method}
\alias{catch.wt<-,FLIndex,FLQuant-method}
\alias{catch.wt<-,FLIndex,numeric-method}
\alias{desc,FLIndex-method}
\alias{desc<-,FLIndex,character-method}
\alias{distribution,FLIndex-method}
\alias{distribution<-,FLIndex,character-method}
\alias{effort,FLIndex-method}
\alias{effort<-,FLIndex,FLQuant-method}
\alias{index,FLIndex-method}
\alias{index.q,FLIndex-method}
\alias{index.q<-,FLIndex,FLQuant-method}
\alias{index.q<-,FLIndex,numeric-method}
\alias{index.var,FLIndex-method}
\alias{index.var<-,FLIndex,FLQuant-method}
\alias{index.var<-,FLIndex,numeric-method}
\alias{index<-,FLIndex,FLQuant-method}
\alias{index<-,FLIndex,numeric-method}
\alias{name,FLIndex-method}
\alias{name<-,FLIndex,character-method}
\alias{range,FLIndex-method}
\alias{range<-,FLIndex,numeric-method}
\alias{sel.pattern,FLIndex-method}
\alias{sel.pattern<-,FLIndex,FLQuant-method}
\alias{sel.pattern<-,FLIndex,numeric-method}
\alias{type,FLIndex-method}
\alias{type<-,FLIndex,character-method}
\title{Class FLIndex}
\usage{
FLIndex(object, ...)

\S4method{FLIndex}{FLQuant}(object, plusgroup = dims(object)$max, ...)

\S4method{FLIndex}{missing}(object, ...)
}
\description{
A class for modelling abundance indices.
}
\details{
The \code{FLIndex} object holds data and parameters related to abundance
indices.
}
\section{Slots}{
 \describe{ \item{type}{Type of index (\code{character}).}
\item{distribution}{Statistical distribution of the index values
(\code{character}).} \item{index}{Index values (\code{FLQuant}).}
    \item{index.var}{Variance of the index (\code{FLQuant}).}
    \item{catch.n}{Catch numbers used to create the index (\code{FLQuant}).}
    \item{catch.wt}{Catch weight of the index (\code{FLQuant}).}
    \item{effort}{Effort used to create the index (\code{FLQuant}).}
    \item{sel.pattern}{Selection pattern for the index (\code{FLQuant}).}
\item{index.q}{Catchability of the index (\code{FLQuant}).} \item{name}{Name
of the stock (\code{character}).} \item{desc}{General description of the
object (\code{character}).} \item{range}{Named numeric vector containing the
quant and year ranges, the plusgroup, and the period of the year, expressed
as proportions of a year, that corresponds to the index (\code{numeric}).} }
}
\examples{

# Create an FLIndex object.
fli <- FLIndex(index=FLQuant(rnorm(8), dim=c(1,8)), name="myTestFLindex")
summary(fli)
index(fli)

# Creat an FLIndex object using an existing FLQuant object.
  data(ple4)
  # Create a perfect index of abundance from abundance at age
    fli2 <- FLIndex(index=stock.n(ple4))
  # Add some noise around the signal
    index(fli2) <- index(fli2)*exp(rnorm(1, index(fli2)-index(fli2), 0.1))

}
\author{
The FLR Team
}
\seealso{
\link{computeCatch}, \link{dims}, \link{iter},
\link[graphics]{plot}, \link{propagate}, \link[base]{summary},
\link[base]{transform}, \link{trim}, \link[stats]{window}, \link{FLComp}
}
\keyword{classes}

