% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesLst.R
\docType{class}
\name{FLIndices}
\alias{FLIndices}
\alias{FLIndices,ANY-method}
\alias{FLIndices,list-method}
\alias{FLIndices,missing-method}
\alias{FLIndices-class}
\alias{FLIndices-methods}
\title{Class FLIndices}
\description{
\code{FLIndices} is a class that extends \code{list} through \code{FLlst}
but implements a set of features that give a little more structure to
list objects. The elements of \code{FLIndices} must all be of class
\code{FLIndex}. It implements a lock mechanism that, when turned on, does
not allow the user to increase or decrease the object length.
}
\section{Slots}{
 \describe{ \item{.Data}{The data. \code{list}.}
\item{names}{Names of the list elements. \code{character}.}
\item{desc}{Description of the object. \code{character}.} \item{lock}{Lock
mechanism, if turned on the length of the list can not be modified by adding
or removing elements. \code{logical}.} }
}
\examples{

data(ple4.index)
flis <- FLIndices(INDa=ple4.index, INDb=window(ple4.index, end=2000))

}
\author{
The FLR Team
}
\seealso{
\linkS4class{FLlst}, \link[base]{list}
}
\keyword{classes}

