% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesLst.R
\docType{class}
\name{FLModelSims}
\alias{FLModelSims}
\alias{FLModelSims-class}
\alias{FLModelSims-methods}
\title{Class FLModelSims}
\arguments{
\item{object}{unnamed object to be added to the list}

\item{...}{other named or unnamed objects}
}
\description{
A list of \code{\linkS4class{FLModelSim}} objects.
}
\section{Slots}{
 \describe{
  \item{.Data}{The data. \code{list}.}
  \item{names}{Names of the list elements. \code{character}.}
  \item{desc}{Description of the object. \code{character}.}
  \item{lock}{Lock mechanism, if turned on the length of the list can not be
    modified by adding or removing elements. \code{logical}.} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}
\author{
The FLR Team
}
\seealso{
\link{FLlst}, \link[base]{list}, \link[base]{vector}
}
\keyword{classes}

