% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesArr.R
\docType{class}
\name{FLParJK}
\alias{FLParJK}
\alias{FLParJK-class}
\title{Class FLParJK}
\description{
A class for storing parameters of a jackknifed model fit.
}
\section{Slots}{
 \describe{ \item{.Data}{Describe slot. \code{array}.}
\item{units}{Units of measurement. \code{character}.} \item{orig}{\code{array}}}
}

\seealso{
\link[FLCore]{FLPar}
}
\author{
The FLR Team
}
\keyword{classes}
