% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/classesArr.R,
%   R/FLQuantDistr.R
\docType{class}
\name{FLQuantDistr}
\alias{FLQuantDistr}
\alias{FLQuantDistr-methods}
\alias{FLQuantDistr-class}
\alias{FLQuantDistr,ANY,ANY-method}
\alias{FLQuantDistr,FLQuant,FLQuant-method}
\title{A class for samples of a probability distribution}
\usage{
FLQuantDistr(object, var, ...)

\S4method{FLQuantDistr}{ANY,ANY}(object, var, ...)

\S4method{FLQuantDistr}{FLQuant,FLQuant}(object, var, units = object@units, distr = "norm")
}
\arguments{
\item{object}{Input numeric object}

\item{...}{Additonal arguments}
}
\description{
This extended FLQuant class holds both a measure of central tendendy (mean,
median) and of dispersion (tipically variance), to be later used to generate,
for example, random numbers with those mean and variances.
}
\section{Slots}{

    \describe{
    \item{.Data}{Unnamed slot for storing the mean (or other measure of
      expectation) (\code{FLQuant}).}
    \item{var}{Variance, or other measure of dispersion, (\code{FLQuant}).}
    \item{distr}{Name of the probability distribution, see Details
      (\code{character}).}
}
}

\section{Validity}{


  \describe{
    \item{slot dims}{.Data and var slots must have the same dimensions.}
    \item{slot dimnames}{.Data and var slots must have the same dimnames.}
}

You can inspect the class validity function by using
   \code{getValidity(getClassDef('FLQuantDistr'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.

The contents of the unnamed slot (.Data) can be accessed through the
\code{e()} method, see Example below.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for the .Data slot.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
    \item{Arith}{.}
}
}

\section{Arithmetic}{

The methods under the \emph{Arith} group have been defined for objects of
this class, both for operations between two \code{FLQuantDistr} objects and
with objects of class \code{FLQuant} (\code{FLArray}) as follows:

\describe{
    \item{+, FLQuantDistr,FLArray}{.}
    \item{-, FLQuantDistr,FLArray}{.}
    \item{*, FLQuantDistr,FLArray}{.}
    \item{/, FLQuantDistr,FLArray}{.}
    \item{+, FLQuantDistr,FLQuantDistr}{.}
    \item{-, FLQuantDistr,FLQuantDistr}{.}
    \item{*, FLQuantDistr,FLQuantDistr}{.}
}
}

\examples{

data(ple4)
fqd <- FLQuantDistr(catch.n(ple4), var=catch.n(ple4) * 10, distr='norm')

}
\seealso{
\link{FLQuant}
}
\author{
The FLR Team
}
\keyword{classes}
