% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuantJK.R
\docType{class}
\name{FLQuantJK}
\alias{FLQuantJK}
\alias{FLQuantJK-class}
\alias{FLQuantJK-methods}
\title{A class for jackkifing of FLQuant objects}
\description{
Jackknife resampling works by calculating a given statistic on replicates of
the original data by systematically leaving out each observation. The mean of
the statistic value is then computed. It can be considered a linear approximation
of the bootstrap.
}
\details{
The \code{FLQuantJK} class contains those N replicates of the original
\code{\linkS4class{FLQuant}} object, in which each \code{iter} corresponds
contains one less observation. Jackknifing on an \code{\linkS4class{FLQuant}}
object is carried out along a single dimension. See \code{\link{FLCore}{jackknife}}
for further details.
}
\section{Slots}{

 \describe{
   \item{.Data}{A jackknifed object with replicates stored along the \code{iter}
dimension, (\code{FLQuant}).}
   \item{orig}{The original \code{FLQuant}.}
}
}

\section{Validity}{


 \describe{
   \item{Element dimensions}{The objects in the \code{.Data} and \code{orig}
   slots must share all but rehe last dimension.}
}

You can inspect the class validity function by using
   \code{getValidity(getClassDef("FLQuantJK"))}
}

\section{Constructor}{

A construction method exists for this class that takes a single object of
class \code{FLQuant}, in which the \code{iter} dimensions if of length 1. See
\code{\link[FLCore]{jackknife}} for further details.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:
}

\section{Accessors}{

\describe{
 \item{orig}{Returns the original \code{FLQuant} stored in the \code{orig} slot.}
}
}
\examples{

flq <- FLQuant(rnorm(50), dim=c(5,10))
flj <- FLQuantJK(flq)

# using 'quant' dim
flj <- FLQuantJK(flq, dim='quant')

}
\author{
The FLR Team
}
\seealso{
\code{\linkS4class{FLQuant}} \link{FLCore}{jackknife}
}
\keyword{classes}

