% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesComp.R
\docType{class}
\name{FLS}
\alias{FLS}
\alias{FLS-class}
\title{Class FLS}
\description{
A virtual class that forms the basis for the \code{\linkS4class{FLStock}}
and \code{\linkS4class{FLStockLen}} classes. No objects of this class can be
constructed.
}
\section{Validity}{
 \describe{
    \item{None}{No particular validity checks}
}
}

\section{Slots}{

\describe{
 \item{catch}{Total catch weight (\code{FLQuant}).}
 \item{catch.n}{Catch numbers (\code{FLQuant}).}
 \item{catch.wt}{Mean catch weights (\code{FLQuant}).}
 \item{desc}{Description of the stock (\code{character}).}
 \item{discards}{Total discards weight (\code{FLQuant}).}
 \item{discards.n}{Discard numbers (\code{FLQuant}).}
 \item{discards.wt}{Mean discard weights (\code{FLQuant}).}
 \item{landings}{Total landings weight (\code{FLQuant}).}
 \item{landings.n}{Landing numbers (\code{FLQuant}).}
 \item{landings.wt}{Landing weights (\code{FLQuant}).}
 \item{stock}{Total stock weight (\code{FLQuant}).}
 \item{stock.n}{Stock numbers (\code{FLQuant}).}
 \item{stock.wt}{Mean stock weights (\code{FLQuant}).}
 \item{m}{Natural mortality (\code{FLQuant}).}
 \item{m.spwn}{Proportion of natural mortality before spawning (\code{FLQuant}).}
 \item{mat}{Proportion mature (\code{FLQuant}).}
 \item{harvest}{Harvest rate or fishing mortality. The units of this slot
  should be set to 'harvest' or 'f' accordingly (\code{FLQuant}).}
 \item{harvest.spwn}{Proportion of harvest/fishing mortality before
  spawning (\code{FLQuant}).}
 \item{name}{Name of the stock (\code{character}).}
 \item{range}{Named numeric vector containing the quant and year ranges,
  the plusgroup and the quant range that the average fishing mortality should
  be calculated over (\code{numeric}).} }
}
\author{
The FLR Team
}
\seealso{
\link[base]{[}, \link[base]{[<-}, \link[base]{as.data.frame},
\link{iter}, \link{propagate}, \link{qapply}, \link[base]{summary},
\link[base]{transform}, \link{trim}, \link{units,FLComp-method},
\link{units<-,FLComp,list-method}, \link[stats]{window}
}
\keyword{classes}

