\name{FLSR.control-class}
\docType{class}
\alias{FLSR.control-class}
\alias{FLSR.control}
\alias{show,FLSR.control-method}
\alias{summary,FLSR.control-method}

\title{Class FLSR.control}
\description{A class for parameters controlling the stocvk/recruitment fitting procedure}
\section{Objects from the Class}{
Objects can be created by calls to the creator function \code{FLSR.control()}
}
\section{Creator Function}{
    FLSR.control()
}
\section{Slots}{
  \describe{
    \item{\code{supplied}:}{Object of class \code{"logical"} ~~ }
    \item{\code{best}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{use.bound}:}{Object of class \code{"logical"} ~~ }
    \item{\code{lower}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{upper}:}{Object of class \code{"numeric"} ~~ }
  }
}
\section{Methods}{
  \describe{
    \item{show}{\code{signature(object = "FLSR.control")}: ... }
    \item{summary}{\code{signature(object = "FLSR.control")}: ... }
  }
}
\references{ ~put references to the literature/web site here ~ }
\author{FLR Team}
\seealso{
  \code{\link{FLSR-class}}
}
\examples{
}
\keyword{classes}
\keyword{methods}
