% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/classesLst.R
\docType{class}
\name{FLStocks}
\alias{FLStocks}
\alias{FLStocks-methods}
\alias{FLStocks}
\alias{FLStocks-class}
\alias{FLStocks-methods}
\alias{FLStocks,ANY-method}
\alias{FLStocks,missing-method}
\alias{FLStocks,list-method}
\alias{FLStocks,FLStock-method}
\alias{FLStocks,missing-method}
\alias{FLStocks,list-method}
\title{Class FLStocks}
\usage{
FLStocks(object, ...)

\S4method{FLStocks}{FLStock}(object, ...)

\S4method{FLStocks}{missing}(object, ...)

\S4method{FLStocks}{list}(object, ...)
}
\arguments{
\item{object}{unnamed object to be added to the list}

\item{...}{other named or unnamed objects}
}
\description{
\code{FLStocks} is a class that extends \code{list} through \code{FLlst} but
implements a set of features that give a little bit more structure to list
objects. The elements of \code{FLStocks} must all be of class
\code{FLStock}. It implements a lock mechanism that, when turned on, does
not allow the user to increase or decrease the object length.
}
\section{Slots}{
 \describe{ \item{.Data}{The data. \code{list}.}
\item{names}{Names of the list elements. \code{character}.}
\item{desc}{Description of the object. \code{character}.} \item{lock}{Lock
mechanism, if turned on the length of the list can not be modified by adding
or removing elements. \code{logical}.} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}

\examples{

data(ple4)
fls <- FLStocks(sa=ple4, sb=window(ple4, end=1980))
summary(fls)

}
\seealso{
\link[graphics]{plot}, \link{FLlst}, \link[base]{list}
}
\author{
The FLR Team
}
\keyword{classes}
