% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModel.R, R/FLSR.R
\docType{methods}
\name{fmle}
\alias{fmle}
\alias{fmle,ANY,missing-method}
\alias{fmle,FLModel,ANY-method}
\alias{fmle,FLModel,FLPar-method}
\alias{fmle,FLModel-method}
\alias{fmle,FLSR,ANY-method}
\title{Method fmle}
\usage{
\S4method{fmle}{FLModel,FLPar}(object, start, method = "Nelder-Mead",
  fixed = list(), control = list(trace = 1), lower = rep(-Inf,
  dim(params(object))[2]), upper = rep(Inf, dim(params(object))[2]), ...)

\S4method{fmle}{FLSR,ANY}(object, start, ...)
}
\description{
Method for maximum likelihood estimation
}
\details{
The \code{fmle} method fits the model specified in an \code{FLModel} object
using maximum likelihood estimation (MLE) by minimizing the negative of the
log-likelihood function given in the \code{logl} slot, through calls to the
\code{\link[stats]{optim}} minimizaton routine.

For a given model and log-likelihood function, the \code{fmle} method will
use the \code{optim} function in R to calculate the parameter vector which
maximises the log-likelihood (and, hence, the likelihood function), and
therefore provides the optimum parameter values for the given problem and
data.

Be advised that for non-informative or conflicting data, the maximum
likelihood estimate can be highly dependent on initial values, and if the
optimiser is given poor initial estimates, it may converge falsely. Always
try multiple start points to be assured that the true MLE solution is found.
}
\section{Generic function}{
 fmle(object,start)
}
\examples{

# use an example FLModel object
  data(nsher)
  summary(nsher)

# inspect the logl function...
  logl(nsher)
# ...and the function providing initial values to the optimizer
  initial(nsher)

# lower and upper limits for the parameters are set, and used if method
# 'L-BFGS-B' is used in the call to optim, as is default in fmle
  lower(nsher)
  upper(nsher)

# fit it with fmle
  nsher <- fmle(nsher)

# fixed values can be chosen for any parameter...
  nsher_fixed_a <- fmle(nsher, fixed=list(a=125))
# ...and results compared, for example using AIC
  AIC(nsher)
  AIC(nsher_fixed_a)

}
\author{
The FLR Team
}
\seealso{
\link{FLComp}
}
\keyword{methods}

