% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{FUNCTION}
\alias{FUNCTION}
\alias{computeHarvest}
\title{Extract and modify the recruitment time series}
\usage{
\S4method{rec}{FLStock}(object, rec.age = as.character(object@range["min"]))

computeHarvest(x, units = NULL)
}
\arguments{
\item{object}{An object of class 'FLStock'}

\item{rec.age}{What age to extract, defaults to first one. As 'character' to select by name or as 'numeric' by position.}

\item{PARAM}{Lorem ipsum dolor sit amet}
}
\value{
RETURN Lorem ipsum dolor sit amet

RETURN Lorem ipsum dolor sit amet
}
\description{
Recruitment in number of fish is the first row of the 'stock.n' slot of
an age-structured 'FLStock'. These convenience functions allow a clearer
syntax when retrieving of altering the content of 'stock.n\link{rec.age,}', where
'rec.age' is usually the first age in the object.

Objects or class 'FLStock' conatin a 'harvest' slot to store estimates of
fihsing mortality at age
}
\examples{
data(ple4)
rec(ple4)
# Multiple recruitment by a factor of 2
rec(ple4) <- rec(ple4) * 2
data(ple4)
# Compute 'f' from stock.n and Baranov
computeHarvest(ple4)
# Recomputes all F at age by solving catch Baranov
recomputeHarvest(ple4)
}
\seealso{
\link{FLComp}

\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
