% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{Funwanted}
\alias{Funwanted}
\title{Calculate the discards and landings-associated fishing mortalities}
\usage{
Funwanted(x, ages = dimnames(x)$age)
}
\arguments{
\item{x}{An FLStock object, with harvest}

\item{ages}{Ages over which the respective Fbar calculation applies}
}
\value{
An FLQuant
}
\description{
Computes the fishing mortality at age (harvest) associated with either
landings (*Fwanted*) or discards (*Funwanted*) through the respective
proportions at age. The function names reflect the convention used in
ICES.
}
\examples{
data(ple4)
Fwanted(ple4, ages=2:6)
Funwanted(ple4, ages=1:3)
}
