% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\docType{methods}
\name{fwd,FLStock,missing,missing-method}
\alias{fwd,FLStock,missing,missing-method}
\title{Project forward to reconstruct abundances from F and M}
\usage{
\S4method{fwd}{FLStock,missing,missing}(object)
}
\arguments{
\item{object}{an \code{FLStock} object}
}
\value{
\code{FLStock} object
}
\description{
An FLStock object is projected forward using the initial abundances and
the total mortality-at-age per timestep. New values for the stock.n and
catch.n slots are calculated, assuming that harvest and m are correct.
This calculation provides a test of the internal consistency of the object.
}
\examples{
data(ple4)
test <- fwd(ple4)
plot(FLStocks(PLE=ple4, TEST=test))
}
\seealso{
\code{\link{harvest}}
}
