% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPlural.R
\docType{methods}
\name{getPlural,ANY-method}
\alias{ac}
\alias{convert6d}
\alias{convertFLPar}
\alias{convertFLPar,FLModel-method}
\alias{convertFLPar,FLPar-method}
\alias{convertFLPar-methods}
\alias{getPlural}
\alias{getPlural,ANY-method}
\title{FLCore-internal}
\usage{
\S4method{getPlural}{ANY}(object)
}
\description{
FLCore-internal
}
\author{
The FLR Team
}
\keyword{methods}

