% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/getPlural.R
\docType{methods}
\name{getPlural}
\alias{getPlural}
\alias{getPlural-methods}
\alias{getPlural,ANY-method}
\alias{getPlural,FLQuant-method}
\alias{getPlural,FLCohort-method}
\alias{getPlural,FLS-method}
\alias{getPlural,FLI-method}
\alias{getPlural,FLBiol-method}
\alias{getPlural,FLSR-method}
\alias{getPlural,FLModelSim-method}
\title{Return the plural class name associated with an object}
\usage{
getPlural(object, ...)
}
\arguments{
\item{object}{An object for which the plural class name is required.}

\item{...}{Further arguments passed to methods.}
}
\value{
A character vector of length one giving the name of the plural class, or
\code{"list"} when no FLCore plural class is defined.
}
\description{
\code{getPlural()} maps an FLCore object class to the corresponding container
class used to store multiple objects of that type.
}
\seealso{
\link{FLlst}, \link{FLQuants}, \link{FLCohorts}, \link{FLStocks},
\link{FLIndices}, \link{FLBiols}, \link{FLSRs}, \link{FLModelSims}
}
\author{
The FLR Team
}
\keyword{methods}
