% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLAccesors.R
\name{getSlotNamesClass}
\alias{getSlotNamesClass}
\title{Names of slots of a given class}
\usage{
getSlotNamesClass(object, class)
}
\arguments{
\item{object}{An S4 object to check slots from.}

\item{class}{The name of the class to match, 'character'.}
}
\description{
This function returns the names, as a character vector, of the slots in an
S4 object that are of the class specified by the 'class' argument. Comparison
is done using is(), so class inheritance is matched.
}
\examples{
data(ple4)
getSlotNamesClass(ple4, 'FLQuant')

}
\author{
The FLR Team
}
\keyword{methods}

