% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLModelDeriv.R
\docType{methods}
\name{gradient}
\alias{gradient}
\alias{gradient-methods}
\alias{gradient,function,FLPar-method}
\title{Numerically differentiate an FLModel likelihood}
\usage{
gradient(func, x, ...)

\S4method{gradient}{function,FLPar}(func, x, method = "Richardson", eps = 1e-04,
  d = 1e-04, zero.tol = sqrt(.Machine$double.eps/7e-07), r = 4, v = 2,
  show.details = FALSE, ...)
}
\arguments{
\item{func}{A function, generally the likelihood component to be evaluated.}

\item{x}{An \code{\link{FLPar}} object containing parameter values.}

\item{method}{Numerical differentiation method. Currently only Richardson
extrapolation is implemented.}

\item{eps}{A small perturbation used in the numerical approximation.}

\item{d}{Relative step size used to initialise perturbations.}

\item{zero.tol}{Threshold used to treat values as zero.}

\item{r}{Number of Richardson extrapolation steps.}

\item{v}{Reduction factor applied to the perturbation at each Richardson
iteration.}

\item{show.details}{Should intermediate Richardson approximations be printed?}

\item{...}{Additional arguments passed to \code{func}.}
}
\value{
For \code{gradient()}, a numeric vector of first derivatives.
}
\description{
\code{gradient()} approximates the gradient of a model function with respect to
an \code{\link{FLPar}} parameter vector.
}
\seealso{
\link{computeD}, \link{computeHessian}, \linkS4class{FLModel},
\linkS4class{FLPar}
}
\author{
The FLR Team
}
\keyword{methods}
