% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLQuant.R
\name{group}
\alias{group}
\alias{group,FLQuant,function-method}
\title{Group objects over some index by applying a function over a single dimension}
\usage{
group(x, FUN, ...)

\S4method{group}{FLQuant,function}(x, FUN = sum, ...)
}
\arguments{
\item{x}{An object to group.}

\item{FUN}{A function to apply along the chosen dimension, defaults to 'sum'.}

\item{...}{An expression or indexing vector, named as the chosen dimension. Extra arguments to FUN can also be provided, but cannmot match names in x.}
}
\value{
A single object with reduced dimensionality.
}
\description{
Array objects (e.g. FLQuant or FLQuants) are divided along a single
dimnension following a given index or expression, an aggregating
function is applied to each subset, and the results are joined again.
Data can be added, for example, by decade or for two age groups.
}
\examples{
data(ple4)
# Add catch-at-age along two age groups, 'juv'eniles and 'adu'lts
group(catch.n(ple4), sum, age=c('juv', 'juv', rep('adu', 8)))
# An expression can use based on dimnames
group(catch.n(ple4), sum, age=age < 3)
# Mean by lustrum, by using 'year - year \%\% 5'
group(catch.n(ple4), mean, year = year - year \%\% 5)
}
\author{
Iago Mosqueira (WMR)
}
\keyword{methods}
