% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{vonbert}
\alias{vonbert}
\alias{ivonbert}
\alias{gompertz}
\alias{richards}
\title{Growth models}
\usage{
vonbert(linf, k, t0, age)

ivonbert(linf, k, t0, len)

gompertz(linf, a, k, age)

richards(linf, k, b, m, age)
}
\description{
Growth models

ivonbert

gompertz

richards
}
\examples{
data(ple4)
vonbert

vonbert(35, 0.352, -0.26, 1:14)
ivonbert(35, 0.352, -0.26, 1:34)
gompertz(linf=179.13, k=0.4088, a=1.7268, age=1:12)
richards(linf=178.63, k=0.424, b=-7.185, m=2880.4, age=1:12)
}
