\name{harvest}
\docType{methods}
\alias{harvest,FLBiol-method}
\alias{harvest,FLBiol,missing-method}
\alias{harvest,FLStock,missing-method}
\alias{harvest<-,FLStock,numeric-method}
\title{Harvest calculations for FLBiol}
\description{
  The fishing mortality endured by a population, represented by an \code{FLBiol} object,
  can be approximated by
}

\section{Generic function}{
 harvest(object) 
}
\section{Methods}{\describe{
\item{ signature(object=FLBiol) :}{Uses the method above to calculate an approximation of harvest rate}
}}
\author{The FLR Team}
\seealso{ \link{FLComp}, \link{FLStock-class}}
\examples{
}
\keyword{methods}
