% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\name{iav}
\alias{iav}
\title{Compute the inter-annual variability of a time series}
\usage{
iav(object)
}
\value{
An object of the same class as object.
}
\description{
The inter-annual variability of a time series stored in an FLQuant object,
is computed as \eqn{|x_y - x_{y-1}) / x_{y-1}|}. The resulting object will
be one year shorter than the input. The first year will be missing as values
are assigned to the final year of each pair.
}
\examples{
data(ple4)
# Compute inter-annual variability in catch
iav(catch(ple4))
}
\author{
The FLR Team
}
\keyword{methods}
