% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R
\name{intersect}
\alias{intersect}
\title{Returns FLR objects trimmed to their shared dimensions.}
\usage{
\S4method{intersect}{FLArray,FLArray}(x, y)
}
\arguments{
\item{x}{First object to be compared and subset}

\item{y}{Second object to be compared and subset}
}
\value{
And object of the corresponding \emph{FLsdt}-based plural class.
}
\description{
Objects sharing certain dimensions, as inferred by their \emph{dimnames}, are
subset to the common ones along all dimensions. The returned object is of one
of the \emph{FLlst} classes, as corresponds to the input class. The objects in
the list can then be, for example, combined or directly compared, as shown
in the examples.
}
\examples{
big <- FLQuant(64.39, dimnames=list(age=1:4, year=2001:2012))
small <- FLQuant(3.52, dimnames=list(age=2:3, year=2001:2005))
intersect(big, small)

# Two FLQuant objects can be added along their common dimension using Reduce() 
Reduce('+', intersect(big, small))
}
\seealso{
\link[base:intercept]{base::intercept}
}
\author{
The FLR Team
}
\keyword{methods}
