\name{is.FLBiols}
\alias{is.FLBiols}
\title{Checks for objects of class FLBiols}
\description{
	Checks whether an object is of class \code{\link{FLBiols-class}}
}
\usage{
is.FLBiols(x)
}
\arguments{
  \item{x}{An object to be tested}
}
\value{
	TRUE or FALSE
}
\author{FLR Team. L.T. Kell}
\seealso{\code{\link{FLBiols-class}}}
\keyword{methods}
