% $Id: is.FLFLeet.Rd,v 1.1.2.1 2005/12/19 11:13:48 iagoazti Exp $
\name{is.FLFleet}
\alias{is.FLFleet}
\title{FLFleet}
\description{Testing for FLFleet list objects}
\usage{
  is.FLFleet(x)
}
\arguments{
  \item{x}   {a \code{FLFleet} object.}
}
\value{
	\code{is.FLFleet} returns \code{TRUE} or \code{FALSE} depending on whether the object is or not of class \code{FLFleet}.
}
\seealso{
    \code{\link{FLFleet-class}}
}
\examples{
}
\author{FLR Team}
\keyword{classes}
\keyword{manip}
