% $Id: is.FLFleets.Rd,v 1.1.2.1 2005/12/19 11:13:48 iagoazti Exp $
\name{is.FLFleets}
\alias{is.FLFleets}
\title{FLFleets}
\description{Testing for FLFleets list objects}
\usage{
  is.FLFleets(x)
}
\arguments{
  \item{x}   {a \code{FLFleets} object.}
}
\value{
	\code{is.FLFleets} returns \code{TRUE} or \code{FALSE} depending on whether the object is or not of class \code{FLFleets}.
}
\seealso{
    \code{\link{FLFleets-class}}
}
\examples{
}
\author{FLR Team}
\keyword{classes}
\keyword{manip}
