% $Id: is.FLIndices.Rd,v 1.1.2.1 2005/12/19 11:13:48 iagoazti Exp $
\name{is.FLIndices}
\alias{is.FLIndices}
\title{FLIndices}
\description{Testing for FLIndices list objects}
\usage{
  is.FLIndices(x)
}
\arguments{
  \item{x}   {a \code{FLIndices} object.}
}
\value{
	\code{is.FLIndices} returns \code{TRUE} or \code{FALSE} depending on whether the object is or not of class \code{FLIndices}.
}
\seealso{
    \code{\link{FLIndices-class}}
}
\examples{
}
\author{FLR Team}
\keyword{classes}
\keyword{manip}
