% $Id: is.FLSR.Rd,v 1.1.2.1 2005/12/19 11:13:48 iagoazti Exp $
\name{is.FLSR}
\alias{is.FLSR}
\title{FLSR}
\description{Testing for FLSR objects}
\usage{
  is.FLSR(x)
}
\arguments{
  \item{x}   {a \code{FLSR} object.}
}
\value{
	\code{is.FLSR} returns \code{TRUE} or \code{FALSE} depending on whether the object is or not of class \code{FLSR}.
}
\seealso{
    \code{\link{FLSR-class}}
}
\examples{
}
\author{FLR Team}
\keyword{classes}
\keyword{manip}
