% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{is.FLStock}
\alias{is.FLStock}
\title{Is this an FLStock object?}
\usage{
is.FLStock(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical value, TRUE if x inherits from FLStock
}
\description{
Checks whether an object inherits from the FLStock class.
}
\examples{
data(ple4)
is.FLStock(ple4)
is.FLStock(stock.n(ple4))
}
\seealso{
\linkS4class{FLStock}
}
\author{
FLR Team
}
\keyword{classes}
