% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{iters}
\alias{iters}
\alias{iters-methods}
\alias{iters,FLQuant-method}
\title{Display the iterations stored in an FLQuant}
\usage{
iters(object, ...)
}
\arguments{
\item{object}{An \code{\link{FLQuant}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
\code{iters()} is called for its side effect of printing each iteration and
returns \code{NULL} invisibly.
}
\description{
\code{iters()} prints each iteration contained in an \code{\link{FLQuant}}
together with the units stored in the object. It is intended as a compact
inspection helper when working with iterated FLR objects.
}
\examples{
flq <- FLQuant(1:8, dim=c(2, 2, 1, 1, 1, 2))
iters(flq)
}
\seealso{
\link{iter}, \link{propagate}, \linkS4class{FLQuant}
}
\author{
The FLR Team
}
\keyword{methods}
