% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuantJK.R
\docType{methods}
\name{jackknife}
\alias{jackknife}
\title{A method for creating an object of class FLQuantJK, a jackknifed FLQuant}
\usage{
\S4method{jackknife}{FLQuant}(object, dim = "year", rm.na = TRUE)
}
\arguments{
\item{object}{An object to be jackknifed, \code{\linkS4class{FLQuant}}.}

\item{dim}{Name or position of the dimension to resample through, defaults to "year".}
}
\value{
RETURN Lorem ipsum dolor sit amet
}
\description{
Jackknife resampling works by calculating a given statistic on replicates of
the original data by systematically leaving out each observation. The mean of
the statistic value is then computed. It can be considered a linear approximation
of the bootstrap.
}
\details{
This method resamples along the stated dimension name (defaults to \code{year})
and returns an object with as many \code{iter} as the length of the selected
dimension.
}
\examples{

}
\author{
The FLR Team
}
\seealso{
\code{\linkS4class{FLQuantJK}}, \code{\linkS4class{FLQuant}}
}
\keyword{classes}

