% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuantJK.R
\docType{methods}
\name{jacknife}
\alias{jacknife}
\alias{jacknife,FLQuant-method}
\alias{jacknife-methods}
\title{Method jacknife}
\usage{
\S4method{jackknife}{FLQuant}(object, dim = "year", rm.na = TRUE)
}
\description{
Jacknife resampling
}
\details{
The \code{jacknife} method sets up objects ready for jacknifing, i.e. to
systematically recompute a given statistic leaving out one observation at a
time. From this new set of "observations" for the statistic, estimates for
the bias and variance of the statstic can be calculated.

Input objects cannot have length > 1 along the \code{iter} dimension, and
the resulting object will have one more \code{iter} than the number of
elements in the original object.
}
\section{Generic function}{
 jacknife(object, ...)
}
\examples{

flq <- FLQuant(1:8)
flj <- jacknife(flq)
iters(flj)

# Calculate the bias of the mean and variance estimators
(mean(iter(yearMeans(flj),2:9))-c(iter(yearMeans(flj),1)))*7
(mean(iter(yearVars(flj),2:9))-c(iter(yearVars(flj),1)))*7

}
\author{
The FLR Team
}
\seealso{
\code{\link{FLQuant}}
}
\keyword{methods}

