% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLQuant.R, R/FLQuants.R
\name{join}
\alias{join}
\alias{join,FLQuant,FLQuant-method}
\alias{join,FLQuants,missing-method}
\title{Joins objects along a dimensions where dimnames differ}
\usage{
join(x, y, ...)

\S4method{join}{FLQuant,FLQuant}(x, y)

\S4method{join}{FLQuants,missing}(x, y)
}
\arguments{
\item{x}{An object to join}

\item{y}{An object to join}
}
\value{
A single object
}
\description{
FLQuant objects are joined along a single dimension, on which dimnames
are different. This is the reverse operation to divide.
}
\examples{
data(ple4)
# JOIN over age dimension
x <- catch.n(ple4)[1,]
y <- catch.n(ple4)[2,]
join(x, y)
# JOIN over year dimension
x <- catch.n(ple4)[,10:20]
y <- catch.n(ple4)[,21:25]
join(x, y)
div <- divide(catch.n(ple4), dim=1)
is(div)
length(div)
join(div)
all.equal(join(divide(catch.n(ple4), dim=1)), catch.n(ple4))
}
\author{
Iago Mosqueira (WMR)
}
\keyword{methods}
