% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{indicators.len}
\alias{indicators.len}
\alias{lenquantile}
\alias{lmax5}
\alias{l95}
\alias{l25}
\alias{lc50}
\alias{lmode}
\alias{lbar}
\alias{lmean}
\alias{lmaxy}
\alias{pmega}
\alias{bheqz}
\title{Calculate quantile(s) of length distribution}
\usage{
indicators.len(
  object,
  indicators = "lbar",
  model = vonbert,
  params,
  cv = 0.1,
  lmax = 1.25,
  bin = 1,
  n = 500,
  metric = catch.n,
  ...
)

lenquantile(x, quantile = 0.5)

lmax5(x)

l95(x)

l25(x)

lc50(x)

lmode(x)

lbar(x)

lmean(x)

lmaxy(x, lenwt)

pmega(x, linf, lopt = linf * 2/3)

bheqz(x, linf, k, t0, lc = lc50(x))
}
\description{
z = (k * (linf - lmean)) / (lmean - lc)
lmean = sum(naa * len) / sum(naa)
lc, length at first capture
}
\examples{
data(ple4)
indicators.len(ple4, indicators=c('lbar', 'lmaxy'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='catch.n',
  lenwt=FLPar(a=0.01030, b=2.975))
indicators.len(ple4, indicators=c('pmega'),
  params=FLPar(linf=60, k=2.29e-01, t0=-1.37), metric='catch.n')
data(ple4.index)
indicators.len(ple4.index, indicators=c('lbar', 'lmean'),
  params=FLPar(linf=132, k=0.080, t0=-0.35), metric='index')
#
ialk <- invALK(params=FLPar(linf = 60, k = 2.29e-01, t0 = -1.37e+00),
  model=vonbert, age=1:10, lmax=1.2)
samps <- lenSamples(catch.n(ple4), invALK=ialk, n=250)
lenquantile(samps, 0.50)
lmax5(samps)
l95(samps)
l25(samps)
lc50(samps)
lmode(samps)
lbar(samps)
lmean(samps)
# Linf(ple4) = 60
lmean(samps) / (0.75 * lc50(samps) + 0.25 * 60) #
lenwt <- FLPar(a=0.01030, b=2.975)
lmaxy(samps, lenwt)
pmega(samps, linf=60)
linf <- 60
k <- 2.29e-01
t0 <- -1.37e+00
bheqz(samps, linf = 60, k = 2.29e-01, t0 = -1.37e+00)
}
\references{
- Kell, L.T., Minto, C., Gerritsen, H.D. 2022. Evaluation of the skill of length-based indicators to identify stock status and trends. ICES Journal of Marine Science. doiu: 10.1093/icesjms/fsac043.
- ICES. 2015. Report of the Fifth Workshop on the Development of Quantitative Assessment Methodologies based on Life-history Traits, Exploitation Characteristics and other Relevant Parameters for Data-limited Stocks (WKLIFE V), 5–9 October 2015, Lisbon, Portugal. ICES CM 2015/ACOM:56. 157 pp.
- ICES. 2020. Tenth Workshop on the Development of Quantitative Assessment Methodologies based on LIFE-history traits, exploitation characteristics, and other relevant parameters for data-limited stocks (WKLIFE X). ICES Scientific Reports. 2:98. 72 pp. http://doi.org/10.17895/ices.pub.5985
}
