% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLlst-methods.R
\docType{methods}
\name{lock}
\alias{lock}
\alias{lock-methods}
\alias{lock,FLlst-method}
\title{Lock an FLlst object against changes in length}
\usage{
lock(object, ...)
}
\arguments{
\item{object}{An \code{\link{FLlst}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A modified \code{\link{FLlst}} object with its \code{lock} slot updated.
}
\description{
\code{lock()} prevents elements being added to or removed from an
\code{\link{FLlst}} object by setting its internal lock flag to \code{TRUE}.
}
\examples{
fll <- FLlst(a=1:3, b=4:6)
slot(lock(fll), "lock")
}
\seealso{
\link{unlock}, \link{collapse}, \linkS4class{FLlst}
}
\author{
The FLR Team
}
\keyword{methods}
