% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SRmodels.R
\docType{methods}
\name{loglAR1}
\alias{loglAR1}
\alias{loglAR1,FLQuant,FLQuant-method}
\alias{loglAR1,numeric,numeric-method}
\title{FLSR log-likelihood function}
\usage{
\S4method{loglAR1}{FLQuant,FLQuant}(obs, hat, rho = 0)

\S4method{loglAR1}{numeric,numeric}(obs, hat, rho = 0)
}
\description{
DESC
}
\details{
TEXT
}
\references{
Seber, G.A.F., Wild, C.J. 2005. Autocorrelated Errors. In Seber, G.A.F., Wild, C.J.
  Nonlinear regression, pages 271-323. doi: 10.1002/0471725315.ch6.
}

