% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{mbar}
\alias{mbar}
\title{Computes the mean natural mortality acros the fully selected ages}
\usage{
mbar(object, ...)
}
\arguments{
\item{object}{An object of class 'FLStock'.}
}
\value{
An object of class 'FLQuant'.
}
\description{
Equivalent to the mean fishing mortality metric returned by 'fbar', 'mbar'
calculates the mean natural mortality across the ages inside the range defined
by 'minfbar' and 'maxfbar'.
}
\examples{
data(ple4)
mbar(ple4)
}
\seealso{
\link{fbar}
}
\author{
The FLR Team, proposal by H. Winker.
}
