% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{meanage}
\alias{meanage}
\alias{meanageCatch}
\alias{meanwt}
\title{Calculate mean age in stock and catch}
\usage{
meanage(object)

meanageCatch(object)

meanwt(object)
}
\arguments{
\item{object}{An age-structured FLStock object}
}
\value{
An FLQuant object
}
\description{
Computes the average age in the stock numbers, weighted by abundance.
}
\section{Functions}{
\itemize{
\item \code{meanageCatch()}: Computes the average age in the catch-at-age, weighted by numbers caught.

\item \code{meanwt()}: Calculate the mean weight-at-age in stock and catch

}}
\examples{
data(ple4)
meanage(ple4)
meanageCatch(ple4)
meanwt(ple4)
}
\seealso{
\link{meanageCatch}
}
\author{
FLR Team
}
\keyword{ts}
