% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\docType{methods}
\name{mergeFL}
\alias{+,FLStock,FLStock-method}
\alias{mergeFL}
\alias{mergeFLStock}
\title{Method mergeFL}
\usage{
mergeFLStock(x, y)
}
\description{
Merging FLStock objects
}
\details{
Two FLStock objects with the same dimensions can be merged using this method
 or a plus sign. Catch slots are added, and weight slots are averaged,
weighted by the relative catch numbers. No meaningful calculation is
currently done for harvest, harvest.spwn, m and m.spwn; currently, NAs are
inserted into the harvest slot, and unweighted averages used for the others.
}
\section{Methods}{
 \describe{ \item{ signature(e1=FLStock, e2=FLStock)
:}{Adds two FLStock objects} }
}
\author{
The FLR Team
}
\seealso{
\link{FLStock}
}
\keyword{methods}

