% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{mergeFLStock}
\alias{mergeFLStock}
\title{Merge two FLStock objects}
\usage{
mergeFLStock(x, y)
}
\arguments{
\item{x}{An object of class FLStock}

\item{y}{An object of class FLStock}
}
\value{
A merged FLStock object
}
\description{
Combines two FLStock objects by adding their abundances and catches,
and computing weighted mean biological parameters.
}
\examples{
data(ple4)
# Split into two stocks
stk1 <- ple4
stk2 <- ple4
# Merge them back
mergeFLStock(stk1, stk2)
}
\seealso{
\linkS4class{FLStock}
}
\author{
FLR Team
}
\keyword{manip}
