% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLComp.R, R/FLlst-methods.R
\docType{methods}
\name{model.frame}
\alias{model.frame}
\alias{model.frame,FLComp-method}
\alias{model.frame,FLlst-method}
\title{Method model.frame}
\usage{
\S4method{model.frame}{FLComp}(formula, mcf = TRUE, ...)

\S4method{model.frame}{FLlst}(formula, drop = FALSE, ...)
}
\description{
\code{model.frame} returns a \code{\link[base]{data.frame}} with the
variables in a wide format, to be used by a \code{formula} in any model
method.
}
\section{Generic function}{
 model.frame(formula)
}
\examples{

data(ple4)
flqs <- FLQuants(stock=stock.n(ple4), catch=catch.n(ple4))
fmf <- model.frame(flqs)
head(fmf)

fqs <- FLQuants(catch=catch(ple4), stock=stock(ple4))
mdf <- model.frame(fqs)
head(mdf)

}
\author{
The FLR Team
}
\seealso{
\link[stats]{model.frame}, \linkS4class{FLQuants},
\linkS4class{FLlst}
}
\keyword{methods}

