% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{mohnMatrix}
\alias{mohnMatrix}
\title{Generate a matrix to compute Mohn's rho for a single metric}
\usage{
mohnMatrix(stocks, metric = "fbar", ...)
}
\arguments{
\item{stocks}{An FLStocks object from a restrospective analysis}

\item{metric}{Metric to be computed, as a character vector or function}

\item{...}{Additional arguments passed to the metric function}
}
\value{
A metrics of n + 2 x n, where n is the numbers of objects in stocks.
}
\description{
Creates a matrix to compute Mohn's rho for a single metric from
retrospective analysis results. A common measure of the strength of stock
out the calculation but returns a matrix with the metrics value for the n
restrospective runs, in columns, and n + 2 years, in rows.
}
\examples{
data(ple4)
# Create retrospective runs
stks <- FLStocks(lapply(0:5, function(i) window(ple4, end=2017-i)))
mohnMatrix(stks, "fbar")
}
\seealso{
\linkS4class{FLStocks}
}
\author{
FLR Team
}
\keyword{ts}
