% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{msy}
\alias{msy}
\alias{msy-methods}
\alias{bmsy}
\alias{bmsy-methods}
\alias{sbmsy}
\alias{sbmsy-methods}
\alias{fmsy}
\alias{fmsy-methods}
\title{msy: A series of methods to extract or compute MSY-based reference points}
\usage{
msy(x, ...)

bmsy(x, ...)

sbmsy(x, ...)

fmsy(x, ...)
}
\arguments{
\item{x}{An input object from which to extract or compute a reference point}
}
\value{
A value for the requested reference point, 'FLPar'
}
\description{
Reference points based on equilibirum calculations of Maximum Sustainable
Yield (MSY) are computed by various FLR packages. The methods' generics are
defined here for convenience. Please refer to the help pages of particular
methods for further details
}
\details{
The four methods provide the following parameter estimates:
\itemize{
\item \code{msy} Maximum Sustainable Yield (MSY)
\item \code{fmsy} Fishing mortality level expected to produce on average MSY
\item \code{bmsy} Total biomass that should produce MSY
\item \code{sbmsy} Spawning biomass that should produce MSY
}
}
\seealso{
\code{\link{FLPar}}
}
\author{
The FLR Team
}
\keyword{methods}
