% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R, R/FLlst-methods.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,FLArray-method}
\alias{names,FLPar-method}
\alias{names,FLlst-method}
\alias{names<-,FLPar,character-method}
\title{Method names}
\usage{
\S4method{names}{FLArray}(x)

\S4method{names}{FLPar}(x)

\S4method{names}{FLlst}(x)
}
\description{
The \code{names} method returns the names of the dimnames of an object. For
some classes, the names attribute can be modified directly using names<-.
}
\section{Generic function}{
 names(x) names<-(x, value)
}
\examples{
# FLQuant
data(ple4)
names(catch.n(ple4))

# Contrast this with
dimnames(catch.n(ple4))

# FLPar
data(nsher)
names(params(nsher))

# FLlst, e.g. FLIndices
data(ple4.indices)
names(ple4.indices)

}
\author{
The FLR Team
}
\seealso{
\link[base]{names}
}
\keyword{methods}

