% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModel.R
\docType{methods}
\name{nls}
\alias{nls}
\alias{nls,FLModel,missing,missing,missing,missing,missing,missing,missing,missing,missing,missing,missing-method}
\alias{nls,FLModel,missing-method}
\alias{nls-FLCore}
\title{Method nls}
\usage{
\S4method{nls}{FLModel,missing}(formula, data = parent.frame(), start,
  control = nls.control(), algorithm = c("default", "plinear", "port"),
  trace = FALSE, subset, weights, na.action, model = FALSE, lower = -Inf,
  upper = Inf, ...)
}
\description{
For a given forumla (describing a model) and data, this method applies the
simple non-linear least squares algorithm, which calculates the parameters
that minimise the sum of squares difference between the observed (data) and
predicted (model) values.
}
\details{
The algorithm can be sensitive to initial values, so the user should try
different start points and check they converge to the same estimates.
}
\section{Generic function}{

nls(formula,data,start,control,algorithm,trace,subset,weights,na.action,model,lower,upper)
}
\examples{

# An example FLSR (FLModel) object
  data(nsher)

  # set bevholt model
    model(nsher) <- bevholt

  # fit through nls
    nsher <- nls(nsher)
    summary(nsher)

}
\author{
The FLR Team
}
\seealso{
\link{FLComp}
}
\keyword{methods}

