% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/operators.R
\docType{methods}
\name{\%+\%}
\alias{\%*\%,FLPar,FLPar-method}
\alias{\%*\%,FLPar,FLQuant-method}
\alias{\%*\%,FLQuant,FLPar-method}
\alias{\%+\%}
\alias{\%+\%,FLPar,FLPar-method}
\alias{\%+\%,FLPar,FLQuant-method}
\alias{\%+\%,FLQuant,FLPar-method}
\alias{\%+\%,FLQuant,FLQuant-method}
\alias{\%-\%}
\alias{\%-\%,FLPar,FLPar-method}
\alias{\%-\%,FLPar,FLQuant-method}
\alias{\%-\%,FLQuant,FLPar-method}
\alias{\%-\%,FLQuant,FLQuant-method}
\alias{\%/\%,FLPar,FLPar-method}
\alias{\%/\%,FLPar,FLQuant-method}
\alias{\%/\%,FLQuant,FLPar-method}
\alias{\%/\%,FLQuant,FLQuant-method}
\alias{\%^\%}
\alias{\%^\%,FLPar,FLPar-method}
\alias{\%^\%,FLPar,FLQuant-method}
\alias{\%^\%,FLQuant,FLPar-method}
\alias{\%^\%,FLQuant,FLQuant-method}
\alias{operators}
\title{FLQuant arithmetic operators}
\usage{
x \%+\% y

x \%-\% y

x \%^\% y

\S4method{\%*\%}{FLQuant,FLQuant}(x, y)

\S4method{\%/\%}{FLQuant,FLQuant}(e1, e2)

\S4method{\%+\%}{FLQuant,FLQuant}(x, y)

\S4method{\%-\%}{FLQuant,FLQuant}(x, y)

\S4method{\%^\%}{FLQuant,FLQuant}(x, y)

\S4method{\%*\%}{FLPar,FLQuant}(x, y)

\S4method{\%/\%}{FLPar,FLQuant}(e1, e2)

\S4method{\%+\%}{FLPar,FLQuant}(x, y)

\S4method{\%-\%}{FLPar,FLQuant}(x, y)

\S4method{\%^\%}{FLPar,FLQuant}(x, y)

\S4method{\%*\%}{FLQuant,FLPar}(x, y)

\S4method{\%/\%}{FLQuant,FLPar}(e1, e2)

\S4method{\%+\%}{FLQuant,FLPar}(x, y)

\S4method{\%-\%}{FLQuant,FLPar}(x, y)

\S4method{\%^\%}{FLQuant,FLPar}(x, y)

\S4method{\%*\%}{FLPar,FLPar}(x, y)

\S4method{\%+\%}{FLPar,FLPar}(x, y)

\S4method{\%-\%}{FLPar,FLPar}(x, y)

\S4method{\%/\%}{FLPar,FLPar}(e1, e2)

\S4method{\%^\%}{FLPar,FLPar}(x, y)
}
\description{
Arithmetic operations between two \linkS4class{FLQuant} objects using the
standars operators (`+`, `-`, `*`, `/`, `^`, see \link{Arith}) need all dimensions in
both objects to match. This requirement is relaxed by using the percent
version of those four: `%+%`, `%-%`, `%*%`, `%/%` and `%^%`.
}
\details{
If any of the objects is of length one in a dimensions where the other is
longer, the dimensions will be extended and the element-by-element operation
then conducted. Dimensions and dimnames of the output will be those of the
larger object. See the examples to observe their behaviour.

Please note that this behjaviour is present on the \link{Arith} methods for
\linkS4class{FLArray}-derived classes but only on the 6th, `iter`, dimension.

The original use of the `%*%` operator, as vector product, is not available
for \linkS4class{FLQuant} objects, but can be applied to the \link{array}
inside them, as in the example below.

Methods for operations between an \linkS4class{FLQuant} and an
\linkS4class{FLPar} object will match dimensions by names of dimnames,
regardless of position.
}
\section{Generic function}{
 %+%, %-%, %*%, %/%, %^%
}
\examples{

a <- FLQuant(2, dim=c(3,3,2))
b <- FLQuant(3, dim=c(3,3,1))

# This should fail
\dontrun{ a * b }

a \%*\% b
a \%+\% b
# To use base's \%*\% vector product, apply it to a matrix from @.Data
b@.Data[,,,,,] \%*\% 1:3
# or
b[,,drop=TRUE] \%*\% 1:3

# FLPar vs. FLQuant works by dimnames' names
flp <- FLPar(2, dimnames=list(params='a', year=2000:2005, iter=1))
flq <- FLQuant(3, dimnames=list(year=2000:2005))
flp \%*\% flq
}
\author{
The FLR Team
}
\keyword{methods}

