% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,FLBiol,missing-method}
\alias{plot,FLCohort,missing-method}
\alias{plot,FLIndex,missing-method}
\alias{plot,FLIndices,missing-method}
\alias{plot,FLPar,missing-method}
\alias{plot,FLQuant,missing-method}
\alias{plot,FLQuantPoint,missing-method}
\alias{plot,FLSR,missing-method}
\alias{plot,FLStock,missing-method}
\alias{plot,FLStocks,FLPar-method}
\alias{plot,FLStocks,missing-method}
\title{Method plot}
\usage{
\S4method{plot}{FLQuant,missing}(x, xlab = "year", ylab = paste("data (",
  units(x), ")", sep = ""), type = "p", ...)

\S4method{plot}{FLStock,missing}(x, auto.key = TRUE, ...)

\S4method{plot}{FLBiol,missing}(x, y, ...)

\S4method{plot}{FLIndex,missing}(x, type = c("splom"), ...)

\S4method{plot}{FLSR,missing}(x, main = "Functional form",
  log.resid = FALSE, cex = 0.8)

\S4method{plot}{FLPar,missing}(x, y = "missing", ...)
}
\description{
Standard plot methods for every FLR class. FLR plot methods are based on
\code{\link[lattice]{lattice}}, and attempt to show a general view of the
object contents.
}
\details{
Users are encouraged to write their own plotting code and make use of the
overloaded \code{\link[lattice]{lattice}} methods, for example
\code{\link[lattice]{xyplot}} or \code{\link[lattice]{bwplot}}. See also
\code{\link{lattice-FLCore}}.
}
\section{Generic function}{
 plot(x,y)
}
\examples{

data(ple4)

# FLQuant
plot(catch.n(ple4)[, 1:20])
plot(catch.n(ple4)[, 1:20], type='b', pch=19, cex=0.5)

# FLStock
data(ple4sex)
plot(ple4)
plot(ple4sex)

# FLBiol
data(ple4.biol)
plot(ple4.biol)

# FLIndex
data(ple4.index)
plot(ple4.index)

# FLSR
data(nsher)
plot(nsher)

# FLPar
fpa <- FLPar(a=rnorm(100, 1, 20), b=rlnorm(100, 0.5, 0.2))
plot(fpa)

}
\author{
The FLR Team
}
\seealso{
\link[graphics]{plot}
}
\keyword{methods}

