% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotDiagnostics.R
\docType{methods}
\name{plotA}
\alias{plotA}
\alias{plotA-methods}
\alias{plotA,FLSR,missing-method}
\title{Plot diagnostic panels for an FLSR fit}
\usage{
plotA(x, y, ...)
}
\arguments{
\item{x}{An \code{\link{FLSR}} object.}

\item{y}{Unused.}

\item{...}{Unused.}
}
\value{
\code{plotA()} is called for its side effect of producing diagnostic plots and
returns \code{NULL} invisibly.
}
\description{
\code{plotA()} draws a set of lattice-based diagnostic plots for an
\code{\link{FLSR}} object, including fitted relationships, residual plots,
autocorrelation checks and Q-Q plots.
}
\examples{
\dontrun{
data(nsher)
plotA(nsher)
}
}
\seealso{
\link{plot}, \link{residuals}, \link{fitted}, \linkS4class{FLSR}
}
\author{
The FLR Team
}
\keyword{methods}
