% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{+,FLStock,FLStock-method}
\alias{+,FLStock,FLStock-method}
\title{Adds two FLStock objects}
\usage{
\S4method{+}{FLStock,FLStock}(e1, e2)
}
\arguments{
\item{e1}{An object of class FLStock}

\item{e2}{An object of class FLStock}
}
\value{
A summed-up FLStock object
}
\description{
Combines two FLStock objects by adding their abundances and catches,
and computing weighted mean biological parameters.
}
\examples{
data(ple4)
# Addc ple4 to itself
ple4 + ple4
}
\seealso{
\linkS4class{FLStock}
}
\author{
FLR Team
}
\keyword{manip}
