% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModel.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,FLModel-method}
\title{Method predict}
\usage{
\S4method{predict}{FLModel}(object, ...)
}
\description{
\emph{predict} returns predicted values according to the parameter values
and model formula in an \code{\linkS4class{FLModel}} object. If no extra
input is given, \emph{predict} will use the input values contained in the
relevant slots. If any extra named argument is provided, this is used
instead, and the corresponding predicted values are returned.
}
\section{Generic function}{
 predict(object, ...)
}
\examples{

# nsher FLSR dataset
  data(nsher)

# Predict with no extra arguments returns the final values predicted during
# model fitting...
predict(nsher)
# ...which can also be extracted from the 'fitted' slot
fitted(nsher)

# A different ssb vector can be provided
predict(nsher, ssb=FLQuant(seq(10, 150, by=5)))

}
\author{
The FLR Team
}
\seealso{
\link{FLComp}
}
\keyword{methods}

