% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R, R/FLPar.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,FLPar-method}
\alias{print,FLQuant-method}
\title{Method print}
\usage{
\S4method{print}{FLQuant}(x)

\S4method{print}{FLPar}(x)
}
\description{
\code{print} prints its argument and returns it invisibly (via
\link[base]{invisible}(x)).
}
\section{Generic function}{
 print(x)
}
\examples{

fq <- FLQuant(1:6, dim = c(2,3))
for(i in 1:3) print(fq[,1:i])

for(i in 1:3) fq[,1:i]

fp <- print(FLPar())
fp

}
\author{
The FLR Team
}
\seealso{
\link{show}
}
\keyword{methods}

