% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{properties}
\alias{properties}
\title{Returns a series of properties of the fisheries element represented by the class.}
\usage{
properties(object, ...)
}
\arguments{
\item{object}{An object from which properties can be extracted.}
}
\value{
The correspodning properties, an FLPar.
}
\description{
Returns a series of properties of the fisheries element represented by the class.
}
\author{
Laurie Kell (Sea++), Iago Mosqueira (WMR)
}
\keyword{methods}
