% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{pv}
\alias{pv}
\alias{pv-methods}
\alias{pv,FLQuant-method}
\title{Calculate plotless variance for a yearly FLQuant time series}
\usage{
pv(object, ...)

\S4method{pv}{FLQuant}(object, dist = FALSE)
}
\arguments{
\item{object}{An \code{\link{FLQuant}} containing a yearly time series.}

\item{dist}{Should the distribution of pairwise distances be returned instead
of the scalar plotless variance summary?}

\item{...}{Further arguments passed to methods.}
}
\value{
If \code{dist = FALSE}, a numeric scalar containing plotless variance. If
\code{dist = TRUE}, a numeric vector with pairwise scaled distances.
}
\description{
\code{pv()} calculates the plotless variance index described by Heath (2006)
for a yearly time series stored in an \code{\link{FLQuant}}.
}
\references{
Heath MR. 2006. Plotted and tabulated indices of change in the state of marine
ecosystems. \emph{Oikos} 115:573-581.
}
\seealso{
\linkS4class{FLQuant}, \link{yearVars}
}
\author{
The FLR Team
}
\keyword{methods}
