% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLComp.R
\docType{methods}
\name{qapply}
\alias{qapply}
\alias{qapply,FLComp,function-method}
\alias{qapply-methods}
\title{Method qapply}
\usage{
\S4method{qapply}{FLComp,`function`}(X, FUN, ..., exclude = missing)
}
\description{
Returns a \code{\link[base]{list}} or \code{\link{FLlst}} containing values
obtained by applying a function to margins for each FLQuant in a composite
FLR object.
}
\section{Generic function}{
 qapply(X,FUN)
}
\examples{

data(ple4)

# Returns a list containing the max value for each quant
qapply(ple4, max)

# Returns an FLStock of means across all dimensions except year
qapply(ple4, apply, 2, mean, na.rm=TRUE)

# Returns an FLStock of max values across all dimensions except year and age
qapply(ple4, apply, c(1,2), max)

}
\author{
The FLR Team
}
\seealso{
\link{FLComp} \link{apply}
}
\keyword{methods}

