% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R
\docType{methods}
\name{quant}
\alias{quant}
\alias{quant,FLArray-method}
\alias{quant-methods}
\alias{quant<-}
\alias{quant<-,FLArray,character-method}
\alias{quant<-,FLArray,missing-method}
\alias{quant<-,FLArray-method}
\alias{quant<--methods}
\title{Method quant}
\usage{
\S4method{quant}{FLArray}(object)
}
\description{
Function to get or set the name of the first dimension (quant) in an object
of class \code{FLQuant} or \code{FLCohort}.
}
\section{Generic function}{
 quant(object) quant<-(object,value)
}
\examples{

# quant is 'quant' by default
  quant(FLQuant())

flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age')
quant(flq)
quant(flq) <- 'length'
summary(flq)

}
\author{
The FLR Team
}
\seealso{
\linkS4class{FLQuant}, \linkS4class{FLCohort}
}
\keyword{methods}

