% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R, R/FLBiol.R
\docType{methods}
\name{r,FLStock,missing-method}
\alias{r}
\alias{r,FLBiol,missing-method}
\alias{r,FLQuant,FLQuant-method}
\alias{r,FLStock,missing-method}
\alias{r-methods}
\title{Methods r}
\usage{
\S4method{r}{FLStock,missing}(m, by = "year", method = "el", ...)

\S4method{r}{FLBiol,missing}(m, by = "year", method = "el", ...)

\S4method{r}{FLQuant,FLQuant}(m, fec, by = "year", method = "el", ...)
}
\arguments{
\item{object}{An object of type \code{\linkS4class{FLBiol}}.}

\item{\dots}{Extra arguments accepted by each implementation.}
}
\value{
An object of class \code{\linkS4class{FLQuant}}.
}
\description{
Intrinsic rate of increase from the Leslie-transition matrix
}
\details{
From two \code{FLQuant} objects representing the natural mortality-at-age
and fecundity, this method calculates the intrinsic rate of increase,
\code{r}, for the given population. Methods exist for objects of classes
\code{FLBiol} and \code{FLStock}, where the \code{fec} and \code{m} slots
are used.

This calculation can be carried out using two methods:
(1) Solving the Euler-Lotka equation.
(2) Calculating the logarithm of the real part of the largest/lead
eigenvalue of the Leslie transition matrix.

These two methods are not identical but give similar answers for the same
data and parameters.

To chose the method used to estimate \code{r} (Euler-Lotka or Leslie matrix)
either 'el' or 'leslie' is supplied as the 'method' argument (see below). To
calculate \code{r} along years or cohorts, either 'year' or 'cohort' is
supplied as the 'by' argument (see below).

The method can handle Monte Carlo samples (i.e. with iterations) in either
the fec or m (or both) slots required to calculate \code{r}, and the
conversion is done internally so that an \code{FLQuant} of the correct
dimensions is obtained.
}
\examples{

# Call the North Sea plaice stock and biol objects
  data(ple4)
  data(ple4.biol)

# Calculate the gradient at the origin of recruits vs. spawning stock numbers
  tmp <- ple4.biol
  n(tmp) <- stock.n(ple4)
  m(tmp) <- harvest(ple4)+m(ple4.biol)

# Use nls to calculate gradient of (recruits/spawning stock numbers)
# assuming log-normal errors
  dfx <- data.frame(rec=as.vector(n(tmp)[1,]),ssn=as.vector(ssn(tmp)))
  res <- nls(log(rec)~log(a)+log(ssn),dfx,start=list(a=mean(n(tmp)[1,]/
    ssn(tmp))))

# Use this value of recruits per spawner times maturity as the birth function
# for the Leslie transition matrix/Euler-Lotka equation
  alpha <- coef(res)[[1]]
  fec(tmp)[] <- fec(tmp)[] * alpha

# calculate r assuming only natural mortality and by year using both the
# Euler-Lotka (el) and Leslie matrix (leslie) method
  m(tmp) <- m(ple4.biol)
  r.ple.el <- r(tmp,by='year',method='el')
  r.ple.lm <- r(tmp,by='year',method='leslie')

}
\author{
FLR Team
}
\seealso{
\code{\linkS4class{FLBiol}}
}
\keyword{methods}

