% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.FLIndices.R
\name{readFLIndices}
\alias{readFLIndices}
\alias{readFLIndex}
\alias{read.FLIndices}
\alias{read.FLIndex}
\title{Read one or more survey indices from exchange files}
\usage{
readFLIndices(file, file2, type = "VPA", index.names, descs, desc = paste("Imported from ", 
  type, " file '", file, "'", sep = ""), na.strings = "NA", sep = "")

readFLIndex(file, type = "VPA", index.names, descs, desc = paste("Imported from ", 
  type, " file '", file, "'", sep = ""))

read.FLIndex(...)

read.FLIndices(...)
}
\arguments{
\item{file}{Path to the main indices input file.}

\item{file2}{Optional secondary file required by some formats, such as
\code{"ICA"}.}

\item{type}{File format to read. Supported values are \code{"VPA"},
\code{"Adapt"}, \code{"CSA"} and \code{"ICA"}.}

\item{index.names}{Optional character vector used to override the names of
individual indices in the returned object.}

\item{descs}{Optional character vector of descriptions for individual indices.}

\item{desc}{Description to assign to the returned object when the file does
not already provide one.}

\item{na.strings}{Character vector used to identify missing values.}

\item{sep}{Field separator passed to the file reader.}

\item{...}{Arguments passed to \code{readFLIndex()} or
\code{readFLIndices()} by the legacy wrapper functions.}
}
\value{
\code{readFLIndices()} returns an \code{\link{FLIndices}} object.
\code{readFLIndex()} returns a single \code{\link{FLIndex}} object, warning
when the input contains more than one index.
}
\description{
\code{readFLIndices()} imports an indices file and returns an
\code{\link{FLIndices}} container. \code{readFLIndex()} is a convenience
wrapper returning the first \code{\link{FLIndex}} in the imported object. The
legacy \code{read.FLIndices()} and \code{read.FLIndex()} wrappers are retained
for backward compatibility.
}
\seealso{
\linkS4class{FLIndex}, \linkS4class{FLIndices}, \link{readFLStock}
}
\author{
The FLR Team
}
