% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.VPAsuite.R
\name{readVPAInterCatch}
\alias{readVPAInterCatch}
\title{Reads a single file with one year of data in VPA format as output by
ICES intercatch}
\usage{
readVPAInterCatch(file)
}
\arguments{
\item{file}{Itercatch VPA to load}
}
\value{
An object of class FLQuant.
}
\description{
Reads a single file with one year of data in VPA format as output by
ICES intercatch
}
