% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLQuantPoint.R
\docType{methods}
\name{rgamma}
\alias{rgamma}
\alias{rgamma,numeric,FLQuantPoint,missing,missing-method}
\title{Method rgamma}
\usage{
\S4method{rgamma}{numeric,FLQuantPoint,missing,missing}(n = 1, shape)
}
\description{
Random generation for the Gamma distribution with parameters 'shape' and
'scale'. 'shape' can be of class \code{\link{FLQuantPoint}} in which case
'shape' and 'scale' are set as \eqn{\hat{x}^2 / \sigma^2}{mean^2 / var} and
\eqn{\sigma^2 / \hat{x}}{var / mean} respectively.
}
\section{Generic function}{
 rgamma(n,shape,rate,scale)
}
\examples{
flq <- FLQuant(rnorm(1000,mean=10,sd=2),dim=c(1,10,1,1,1,100))
flqp <- FLQuantPoint(flq)
rgamma(10,shape=flqp)

data(ple4)
rgamma(10,FLQuantPoint(rnorm(200, catch(ple4), 20000)))
}
\author{
The FLR Team
}
\seealso{
\link[stats]{rgamma}, \linkS4class{FLQuantPoint}
}
\keyword{methods}

