% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{rlnorm}
\alias{rlnorm}
\alias{rlnorm,numeric,FLQuant,FLQuant-method}
\alias{rlnorm,numeric,FLQuant,missing-method}
\alias{rlnorm,numeric,FLQuant,numeric-method}
\alias{rlnorm,numeric,missing,FLQuant-method}
\alias{rlnorm,numeric,numeric,FLQuant-method}
\title{Method rlnorm}
\usage{
\S4method{rlnorm}{numeric,FLQuant,FLQuant}(n = 1, meanlog = 0, sdlog = 1)
}
\description{
Random generation for the lognormal distribution whose logarithm has mean
equal to \emph{meanlog} and standard deviation equal to \emph{sdlog}.
\emph{meanlog} and \emph{sdlog} can be given as \code{FLQuant} objects. If
both are given as \code{FLQuant} objects their dimensions must be the same.
If either of these arguments are \code{FLQuant} objects, \code{rlnorm}
returns an \code{FLQuant}.
}
\section{Generic function}{
 rlnorm(n,meanlog,sdlog)
}
\examples{

out <- rlnorm(1000, meanlog=FLQuant(rep(5,5)), sdlog=FLQuant(0:4))
apply(log(out),2,sd)
apply(log(out),2,mean)

}
\author{
The FLR Team
}
\seealso{
\link[stats]{rlnorm}, \linkS4class{FLQuant},
\linkS4class{FLQuantPoint}
}
\keyword{methods}

