% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\docType{methods}
\name{rnoise,numeric,FLQuant-method}
\alias{rnoise,numeric,FLQuant-method}
\alias{rnoise}
\alias{rnoise-method}
\alias{rnoise,numeric,missing-method}
\alias{rlnoise}
\alias{rlnoise-method}
\alias{rlnoise,numeric,FLQuant-method}
\alias{rlnoise,numeric,missing-method}
\title{Random noise with different frequencies}
\usage{
\S4method{rnoise}{numeric,FLQuant}(
  n = n,
  len = len,
  sd = 1,
  b = 0,
  burn = 0,
  trunc = 0,
  what = c("year", "cohort", "age"),
  seed = NA
)

\S4method{rnoise}{numeric,missing}(n = n, sd = 1, b = 0, burn = 0, trunc = 0, seed = NA)

\S4method{rlnoise}{numeric,FLQuant}(
  n = n,
  len = len,
  sd = 1,
  b = 0,
  burn = 0,
  trunc = 0,
  what = c("year", "cohort", "age"),
  seed = NA
)
}
\arguments{
\item{n}{number of iterations}

\item{len}{an \code{FLQuant}}

\item{sd}{standard error for simulated series}

\item{b}{autocorrelation parameter a real number in [0,1]}

\item{burn}{gets rid of 1st values i series}

\item{trunc}{get rid of values > abs(trunc)}

\item{what}{returns time series for year, cohort or age"}

\item{...}{any}
}
\value{
A \code{FLQuant} with autocorrelation equal to B.
}
\description{
A noise generator
}
\examples{
\dontrun{
flq <- FLQuant(1:100, quant="age")
white <- rnoise(100,flq,sd=.3,b=0)
plot(white)
acf(white)

red <- rnoise(100,flq,sd=.3,b=0.7)
plot(red)
acf(red)

res <- rnoise(100,flq,sd=.3,b=0)

ggplot() +
  geom_point(aes(year,age,size=data),
    data=subset(as.data.frame(res), data>0)) +
geom_point(aes(year,age,size=-data),
            data=subset(as.data.frame(res),data<=0),colour="red")+
scale_size_area(max_size=4, guide="none")+
facet_wrap(~iter)

data(ple4)
res <- rnoise(4,m(ple4),burn=10,b=0.9,what="cohort")
ggplot()+
geom_point(aes(year,age,size= data),
          data=subset(as.data.frame(res),data>0))+
geom_point(aes(year,age,size=-data),
          data=subset(as.data.frame(res),data<=0),colour="red")+
scale_size_area(max_size=4, guide="none")+
facet_wrap(~iter)

}
}
\references{
Ranta and Kaitala 2001 Proc. R. Soc.
vt = b * vt-1 + s * sqrt(1 - b^2)
s is a normally distributed random variable with mean = 0
b is the autocorrelation parameter
}
