% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{rnorm}
\alias{rnorm}
\alias{rnorm,numeric,FLQuant,FLQuant-method}
\alias{rnorm,numeric,FLQuant,missing-method}
\alias{rnorm,numeric,FLQuant,numeric-method}
\alias{rnorm,numeric,missing,FLQuant-method}
\alias{rnorm,numeric,numeric,FLQuant-method}
\title{Method rnorm}
\usage{
\S4method{rnorm}{numeric,FLQuant,FLQuant}(n = 1, mean = 0, sd = 1)
}
\description{
Generates random numbers following a normal distribution. \emph{mean} and
\emph{sd} can be specified as objects of class \code{\linkS4class{FLQuant}},
of the same dimensions, but any of the two could be given as a numeric, in
which case the value will be reused accordingly. If either is of class
\code{FLQuant} then the resultant object will be of class \code{FLQuant}.
}
\section{Generic function}{
 rnorm(n, mean, sd)
}
\examples{

data(ple4)
rnorm(10,mean=harvest(ple4)[,"2001"], sd=harvest(ple4)[,"2001"])

}
\author{
The FLR Team
}
\seealso{
\link[stats]{rnorm}, \linkS4class{FLQuant},
\linkS4class{FLQuantPoint}
}
\keyword{methods}

