% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{rpois}
\alias{rpois}
\alias{rpois,numeric,FLQuant-method}
\title{Method rpois}
\usage{
\S4method{rpois}{numeric,FLQuant}(n = 1, lambda)
}
\description{
Generates random numbers following a Poisson distribution. \emph{lambda},
the (non-negative) mean, can be specified as an object of class
\code{\linkS4class{FLQuant}}.
}
\section{Generic function}{
 rpois(n, lambda)
}
\examples{

data(ple4)
rpois(50,lambda=harvest(ple4))

}
\author{
The FLR Team
}
\seealso{
\link{rpois}, \link{FLQuant}
}
\keyword{methods}

