% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLStock.R
\docType{methods}
\name{ruleset}
\alias{ruleset}
\alias{ruleset,FLStock-method}
\title{Set of verify rules for an FLR class}
\usage{
ruleset(object, ...)

\S4method{ruleset}{FLStock}(object, ...)
}
\arguments{
\item{object}{An object of any FLR class for which the method has been defined.}

\item{...}{Names of positions in the standard list to subset.}
}
\value{
A named list containing the rules defined for for the class object belongs to.
}
\description{
Returns a set of standard rules to be used by the verify method for an
object of a given class.
}
\details{
A standard minimal set of rules to check FLStock objects against using the
verify method. The included rules are (with names in italics) check that:

- there are no NAs in any slot, *anyna*.
- *catch.wt*, *landings.wt*, *discards.wt* and *stock.wt* > 0.
- *mat*, *m.spwn* and *harvest.spwn* values are between 0 and 1.
- *harvest* >= 0.
- *cohorts* in the stock.n slot contain decreasing numbers, except for the plusgroup age.
}
\examples{
data(ple4)
ruleset(ple4)
# Extract single rule by name
ruleset(ple4, 'anyna')
}
\author{
The FLR Team
}
\keyword{methods}
