% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\name{rwalk}
\alias{rwalk}
\title{Generate a random walk time series from a starting point}
\usage{
rwalk(x0, end = 1, sd = 0.05, delta = 0)
}
\arguments{
\item{x0}{The initial state of the random walk, 'FLQuant'.}

\item{end}{The number of years or the final year of the series. numeric.}

\item{sd}{The standard deviation of the random walk, numeric.}

\item{delta}{The drift of the random walk.}
}
\value{
An 'FLQuant' object.
}
\description{
The last year of an \code{FLQuant} object is used as atrating point to generate
a time series following a random walk with drift:
\deqn{z_t = z_{t-1} + \epsilon_t + \delta_t, t=1,2,...}{z[t] = z[t-1] + e[t] + d[t], t=1,2,...}
where \eqn{\epsilon}{e} is \eqn{\mathcal{N}(0, \sigma)}{N(0, sd)}
}
\details{
The length of the series is set by argument \emph{end}. This is taken as a number of years, if its value is smaller than the final 'year' of 'x0', or as a final year if larger or of class 'character'.
}
\examples{
data(ple4)
# Generate random walk recruitmrnt with positive drift
rwalk(rec(ple4), end=5, sd=0.08, delta=0.05)
# Use append() to add the new values at the end
append(rec(ple4), rwalk(rec(ple4), end=10, sd=0.04, delta=0))
# Use end as number of years
rwalk(rec(ple4), end=5)
# or as final year
rwalk(rec(ple4), end=2020)
}
\seealso{
\linkS4class{FLQuant} \link{rnorm}
}
\author{
Iago Mosqueira, WMR (2023)
}
\keyword{classes}
