% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLModel.R
\docType{methods}
\name{sd}
\alias{sd}
\alias{sd,FLModel,missing-method}
\title{Method sd}
\usage{
\S4method{sd}{FLModel,missing}(x)
}
\description{
Standard deviation of an \code{FLModel} object
}
\details{
\code{sd} computes the standard deviation of the parameter estimates in an
\code{FLModel} object by calculating either the square root of the diagonal
of the variance-covariance matrix[[, or, if there are multiple parameter
estimates, the standard deviation of each parameter]].
}
\section{Generic function}{
 sd(x, na.rm)
}
\examples{
data(nsher)
sd(nsher)
}
\author{
The FLR Team
}
\seealso{
\code{\link[stats]{sd}}, \code{\link{FLModel}}
}
\keyword{methods}

