% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R, R/FLStock.R, R/FLBiol.R
\docType{methods}
\name{setPlusGroup}
\alias{setPlusGroup}
\alias{setPlusGroup,FLBiol,numeric-method}
\alias{setPlusGroup,FLQuant,numeric-method}
\alias{setPlusGroup,FLStock,numeric-method}
\alias{setPlusGroup-methods}
\title{Method setPlusGroup}
\usage{
\S4method{setPlusGroup}{FLQuant,numeric}(x, plusgroup, na.rm = FALSE,
  by = "mean")

\S4method{setPlusGroup}{FLStock,numeric}(x, plusgroup, na.rm = FALSE,
  keepPlusGroup = TRUE)

\S4method{setPlusGroup}{FLBiol,numeric}(x, plusgroup, na.rm = FALSE)
}
\description{
Calculates the appropriate values for the plusgroup of an object and returns
a new object with the plusgroup set to the given age.
}
\details{
\emph{quant} of the given object must be 'age', and the selected age must
not be greater than the oldest age present in the object.
}
\section{Generic function}{
 sePlusGroup(x, plusgroup)
}
\examples{

data(ple4)
ple4.pg <- setPlusGroup(ple4, 6)


# For FLStock
data(ple4)
summary(setPlusGroup(ple4, 5))

# For an FLBiol
data(ple4.biol)
# Change spwn slot to make the calculation more interesting
summary(setPlusGroup(ple4.biol, 5))
}
\author{
The FLR Team
}
\seealso{
\linkS4class{FLStock}, \linkS4class{FLQuant}, \linkS4class{FLBiol}
}
\keyword{methods}

