% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLComp.R
\name{slim}
\alias{slim}
\alias{slim-methods}
\alias{slim,FLComp-method}
\title{Drop unnecesary 'iters'}
\usage{
slim(object, ...)

\S4method{slim}{FLComp}(object, ...)
}
\arguments{
\item{object}{A complex \strong{FLR} object to slim down.}
}
\value{
An object of the same class as the input.
}
\description{
Objects of FLR classes can vary in the length along the sixth dimension
in any slot of class \linkS4class{FLQuant}. This reduces object size and memory
usage. If an object has been extended fully, for example by using
\code{\link[=propagate]{propagate}}, we can slim down the object by reducing
any slot where all iters are identical and keeping only yhe first \emph{iter}.
}
\details{
The test for whether an slot can be slimmed is based on checking if the sum
of the variance along the 6th dimensions is equal to zero.
}
\examples{

data(ple4)
# Extend all of ple4 to 50 iters
ple4 <- propagate(ple4, 50)
# Add variability in catch.n
catch.n(ple4) <- rlnoise(50, log(catch.n(ple4)), log(catch.n(ple4))/10)
summary(ple4)
# slim object by dropping identical iters
sple4 <- slim(ple4)
summary(sple4)
}
\seealso{
\linkS4class{FLQuant} \code{\link[=propagate]{propagate}}
}
\author{
The FLR Team
}
\keyword{methods}
