% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{sop}
\alias{sop}
\title{Calculate sum of products for a slot}
\usage{
sop(stock, slot = "catch")
}
\arguments{
\item{stock}{An object of class FLStock}

\item{slot}{The slot prefix ("catch", "landings", or "discards")}
}
\value{
An FLQuant object
}
\description{
Computes the sum of products of numbers and weights for a specified slot
(catch, landings, or discards) divided by the total.
}
\seealso{
\linkS4class{FLStock}
}
\author{
FLR Team
}
\keyword{manip}
