% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R, R/FLComp.R
\name{split-methods}
\alias{split-methods}
\alias{split,FLComp,vector-method}
\title{splits \emph{x} along the \emph{iter} dimension into the groups defined by \emph{f}.}
\usage{
\S4method{split}{FLQuant,vector}(x, f)

\S4method{split}{FLComp,vector}(x, f)
}
\arguments{
\item{x}{The object to be split.}

\item{f}{The vector of group names.}
}
\value{
An object of the corresponding plural class (FLQuants from FLQuant).
}
\description{
Similar to  base::split, but working along the 6th, \emph{iter}, dimension of
any singular FLR object. The object is divided into as many objects as
unique values in \emph{f}, and returned as an FLlst-derived object, e.g. an
FLQuants object when applied to an FLQuant.
}
\examples{
# FROM FLQuant to FLQuants
flq <- rlnorm(20, FLQuant(seq(0.1, 0.8, length=10)), 0.2)
split(flq, c(rep(1, 5), rep(2,15)))
}
\author{
Iago Mosqueira (WMR).
}
\keyword{methods}
