% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{splom}
\alias{splom}
\alias{splom,FLPar,missing-method}
\title{Method splom}
\usage{
\S4method{splom}{FLPar,missing}(x, data, ...)
}
\description{
Draws a conditional scatter plot matrix.
}
\details{
See the help page in \code{\link[lattice]{lattice}} for a full description
of each plot and all possible arguments.
}
\section{Generic function}{
 splom(x,data)
}

\examples{

flp <- FLPar(c(t(mvrnorm(500, mu=c(0, 120, 0.01, 20),
  Sigma=matrix(.7, nrow=4, ncol=4) + diag(4) * 0.3))),
  dimnames=list(params=c('a','b','c','d'), iter=1:500), units="NA")

splom(flp)

}
\seealso{
\link[lattice]{splom}
}
\author{
The FLR Team
}
\keyword{methods}
