% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SRmodels.R
\docType{methods}
\name{spr0}
\alias{spr0}
\alias{spr0,FLQuant,FLQuant,FLQuant-method}
\alias{spr0,FLSR,missing,FLQuant-method}
\alias{spr0,FLStock,missing,missing-method}
\alias{spr0-methods}
\title{Method spr0}
\usage{
\S4method{spr0}{FLQuant,FLQuant,FLQuant}(ssb, rec, fbar)

\S4method{spr0}{FLStock,missing,missing}(ssb)

\S4method{spr0}{FLSR,missing,FLQuant}(ssb, fbar)
}
\description{
Calculates spawners per recruit at F=0.
}
\details{
This method regresses SSB/R against Fbar, and estimates the intercept as
\code{spr0}. The method currently does not work if any of the input objects
have multiple units, seasons or areas (i.e. if dim(object)[3:5] > 1).
}
\section{Generic function}{
 quant(ssb, rec, fbar)
}
\examples{
data(ple4)
spr0(ple4)

# equivalent to
spr0(ssb(ple4), rec(ple4), fbar(ple4))
}
\author{
The FLR Team
}
\seealso{
\linkS4class{FLStock}, \linkS4class{FLSR}
}
\keyword{methods}

