% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SRmodels.R
\name{SRModelName}
\alias{SRModelName}
\title{Method SRModelName}
\usage{
SRModelName(model)
}
\arguments{
\item{model}{A formula defining the model}
}
\value{
\item{name}{A character string or NULL}
}
\description{
Convenience function to identify an SR model by its formula
}
\details{
A supplied formula, representing a stock-recruitment relationship, is
matched against the list of all models defined in \code{FLCore} (See
\link{SRModels}).

If a match is found, a string character with the name of the model is
returned, otherwise \code{FALSE} is obtained.
}
\examples{
SRModelName(rec ~ a * ssb * exp(-b * ssb))
}
\author{
FLR Team
}
\seealso{
\code{\link{SRModels}}
}
\keyword{models}
\keyword{utilities}

