% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLBiol.R
\name{ssb}
\alias{ssb}
\alias{ssb-methods}
\alias{ssb,FLBiol-method}
\alias{ssb-FLBiol,method}
\title{Calculate or return the Spawning Stock Biomass}
\usage{
ssb(object, ...)

\S4method{ssb}{FLBiol}(object, ...)
}
\arguments{
\item{object}{Object on which \code{ssb} is calculated or extracted.}
}
\value{
An object, generally of class \code{\link{FLQuant}}.
}
\description{
The calculated Spawning Stock Biomass (SSB) of a fish population is returned
by this method. SSB is the combined weight of all individuals in a fish stock
that are capable of reproducing. In some classes this is calculated from
information stored in different slots, while in others \code{ssb()} is simply
an slot accessor. When the later is the case, the corresponding replacement
method also exists.
}
\details{
Objects of the *FLBiol* class do not contain any information on
catch or fishing mortality, so a call to `ssb()` will only correct abundances
for natural mortality to the moment of spawning. The method can also take
information on catches or fishing mortality and use them when calculating
abundances at spawning time. An *FLQuant* named either 'catch.n', 'f', 'hr' or
'harvest' can be used. The first three are self-explanatory, while for the last
units must be either 'f' or 'hr'. The quantities should refer to total yearly
values, as the value in the 'spwn' slot will be used to calculate what fraction
of fishing mortality to apply.
}
\examples{
data(ple4)
biol <- as(ple4, "FLBiol")
# SSB from FLBiol, abundances corrected only for M
ssb(biol)
# Provide catch-at-age, F or HR to correct N
ssb(biol, catch.n=catch.n(ple4))
ssb(biol, f=harvest(ple4))
ssb(biol, harvest=harvest(ple4))
ssb(biol, hr=catch.n(ple4) / stock.n(ple4))
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{methods}
