% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{ssb_next}
\alias{ssb_next}
\title{Calculate next yera's SSB from survivors and Fbar}
\usage{
ssb_next(x, fbar = 0, wts.nyears = 3, fbar.nyears = 3)
}
\arguments{
\item{x}{An FLStock object containing estimates of abundance and harvesting.}

\item{fbar}{The Fbar rate assumed on the extra year. Defaults to 0.}

\item{wts.nyears}{Number of years in calculation of mean weight-at-age and maturity for the extra year.}

\item{fbar.nyears}{Number of years in calculation of mean selectivity, natural mortality and fraction of F abnd M before spawning for the extra year.}
}
\value{
An FLQuant.
}
\description{
The spawning stock biomass (SSB) of the stock gets calculated from the
survivors of the previous year. This provides a value for the first year
after the end of the object. Weights-at-age, maturity in this extra year are
calculated as averages over the last \emph{wts.nyears}.
}
\details{
For stocks spawning later in the year, a value for the average fishing
mortality, \emph{fbar}, expected in that year can be provided. Mortality until
spawning is then calculated, with M and selectivity assumed in the extra year
to be an average of the last \emph{fbar.nyears}.
}
\examples{
data(ple4)
ssb_next(ple4)
# Compare with ssb()
ssb(ple4)[, ac(2014:2017)] / ssb_next(ple4)[, ac(2014:2017)]
}
