% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\docType{methods}
\name{ssbpurec}
\alias{ssbpurec}
\alias{ssbpurec,FLStock-method}
\alias{ssbpurec-methods}
\title{Method ssbpurec}
\usage{
\S4method{ssbpurec}{FLStock}(object, start = "missing", end = "missing",
  type = "non-param", recs = "missing", spwns = "missing",
  plusgroup = TRUE, ...)
}
\description{
Calculates the spawning stock biomass per unit recruit at zero fishing
mortality for an \code{FLStock} object.
}
\section{Generic function}{
 ssbpurec(object)
}
\examples{

data(ple4)
ssbpurec(ple4)
ssbpurec(ple4, start=1980, end=2000)

}
\author{
The FLR Team
}
\seealso{
\linkS4class{FLStock}
}
\keyword{methods}

