% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBiol.R
\docType{methods}
\name{ssn}
\alias{ssn}
\alias{ssn,FLBiol-method}
\alias{ssn-methods}
\title{Method ssn}
\usage{
\S4method{ssn}{FLBiol}(object, ...)
}
\description{
Returns the Spawning Stock Numbers of \code{\linkS4class{FLBiol}} objects.
}
\details{
\code{ssn} is calculated as follows:

\eqn{SSB = sum(N * mat * exp(-M*propM))}
}
\section{Generic function}{
 ssn(object)
}
\examples{

data(ple4.biol)

# Change spwn slot to make the calculation more interesting
spwn(ple4.biol)<-0.5

# Calculate ssn and check
ssn(ple4.biol)
ssnp <- quantSums(n(ple4.biol) * fec(ple4.biol) *
  exp(-m(ple4.biol)*spwn(ple4.biol)))
ssn(ple4.biol)-ssnp

}
\author{
The FLR Team
}
\seealso{
\link{FLComp} \link{FLStock}
}
\keyword{methods}

