% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLS.R, R/FLBiol.R
\name{standardUnits}
\alias{standardUnits}
\alias{standardUnits,FLS-method}
\alias{standardUnits,FLBiol-method}
\title{Standard units of measurement for a complex class object}
\usage{
standardUnits(object, ...)

\S4method{standardUnits}{FLS}(object, ...)

\S4method{standardUnits}{FLBiol}(object, ...)
}
\arguments{
\item{object}{for which the standard \emph{units} are to be returned}
}
\value{
A list with the corresponding \emph{units} value for each slot
}
\description{
Returns values for the \emph{units} of each \emph{FLQuant} slot according to the
standard adopted by the FLR Team for the supplied class.
}
\details{
For objects derived from class *FLS*, which currently includes
*FLStock* and *FLStockLen*, the adopted standard includes: 'kg' for individual
weights, '1000' for number of individuals, 't' for biomass, 'f' for harvest,
'm' for natural mortality, and an empty string for proportions (spwn, mat).

For objects derived of class *FLBiol* the adopted standard
units are: 'kg' for individual weights, '1000' for number of individuals,
'm' for natural mortality, and an empty string for proportions (spwn, mat).
}
\examples{
stk <- FLStock(catch=FLQuant(runif(20, 2, 120)))
# FLStock object has no units
summary(stk)
# Obtain standard units for the class as a list
standardUnits(stk)
# which can then be assigned to the object
units(stk) <- standardUnits(stk)
summary(stk)
# units<- methjod also accepts a function to be called to provide units
units(stk) <- standardUnits
bio <- FLBiol(n=FLQuant(runif(50, 2, 120), dim=c(5, 10)))
# Object has no units
summary(bio)
# Obtain standard units for the class as a list
standardUnits(bio)
# which can then be assigned to the object
units(bio) <- standardUnits(bio)
summary(stk)
}
\seealso{
\code{\link{units-FLCore}}
}
\author{
The FLR Team
}
\keyword{methods}
