\name{stripplot}
\docType{methods}
\alias{stripplot,ANY,ANY-method}
\alias{stripplot,formula,FLQuant-method}
\alias{stripplot,formula,FLComp-method}
\title{Method stripplot}
\description{
 My method is better then yours !
}
\section{Generic function}{
 stripplot(x,data) 
}
\section{Methods}{\describe{
\item{ signature(x=ANY,data=ANY) :}{Describe method}
\item{ signature(x=formula,data=FLQuant) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
