% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLQuantPoint.R, R/FLPar.R,
%   R/FLComp.R, R/FLQuants.R, R/predictModel.R, R/FLBiol.R, R/FLModel.R,
%   R/FLlst-methods.R
\docType{methods}
\name{summary,FLArray-method}
\alias{summary,FLArray-method}
\alias{summary,FLQuantPoint-method}
\alias{summary,FLPar-method}
\alias{summary,FLComp-method}
\alias{summary,FLQuants-method}
\alias{summary,FLQuants-methods}
\alias{summary,predictModel-method}
\alias{summary,FLBiol-method}
\alias{summary,FLModel-method}
\alias{summary,FLlst-method}
\alias{summary,FLlst-methods}
\title{Method summary}
\usage{
\S4method{summary}{FLArray}(object, .class = TRUE, ...)

\S4method{summary}{FLQuantPoint}(object, ...)

\S4method{summary}{FLPar}(object, title = TRUE, ...)

\S4method{summary}{FLComp}(object, ...)

\S4method{summary}{FLQuants}(object)

\S4method{summary}{predictModel}(object)

\S4method{summary}{FLBiol}(object)

\S4method{summary}{FLModel}(object, ...)

\S4method{summary}{FLlst}(object)
}
\description{
Outputs a general summary of the structure and content of an fwdControl
object. The method invisibly returns the data.frame shown on screen.
}
\section{Generic function}{
 summary(object)
}

\examples{

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
summary(flq)

data(ple4)
summary(ple4)

data(nsher)
summary(nsher)

}
\seealso{
\link[base]{summary}
}
\author{
The FLR Team
}
\keyword{methods}
