% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,FLArray-method}
\title{Method summary}
\usage{
\S4method{summary}{FLArray}(object, ...)
}
\description{
Outputs a general summary of the structure and content of the object. The
particular output obtined depends on the class of the argument object.
}
\section{Generic function}{
 summary(object)
}
\examples{

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
summary(flq)

data(ple4)
summary(ple4)

data(nsher)
summary(nsher)

}
\author{
The FLR Team
}
\seealso{
\link[base]{summary}
}
\keyword{methods}

