% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLBiol.R
\docType{methods}
\name{survprob}
\alias{survprob}
\alias{survprob,FLBiol-method}
\alias{survprob-methods}
\title{Method survprob}
\usage{
\S4method{survprob}{FLBiol}(object, by = "year", ...)
}
\arguments{
\item{object}{An object of type \code{\linkS4class{FLBiol}}.}

\item{\dots}{Extra arguments accepted by each implementation.}
}
\value{
An object of class \code{\linkS4class{FLQuant}}.
}
\description{
Calculating survival probabilties given mortality in the \code{FLBiol} object
}
\details{
For an \code{FLBiol} object with the mortality-at-quant present, this method
calculates the associated survival probability-at-quant. This can be used
later by the \code{r()} method. The calculation can be carried out either by
year or by cohort.

Individual survival probability (i.e. from one year to the next) is simply
exp(-M) and the survival probaiblity for a given quant is the cumulative
product along the quant dimension of the individual survival probabilities.
}
\examples{

data(ple4.biol)
survprob(ple4.biol,by='year')

}
\author{
FLR Team
}
\seealso{
\code{\linkS4class{FLBiol}}
}
\keyword{methods}

