% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R
\docType{methods}
\name{sweep,FLArray-method}
\alias{sweep,FLArray-method}
\alias{sweep,FLPar-method}
\title{Method sweep for FLCore classes}
\usage{
\S4method{sweep}{FLArray}(x, MARGIN, STATS, FUN = "-", check.margin = TRUE,
  ...)

\S4method{sweep}{FLPar}(x, MARGIN, STATS, FUN = "-", check.margin = TRUE,
  ...)
}
\description{
Use R's sweep method on FLCore classes
}
\details{
These methods call base R \code{\link[base]{sweep}} method on **FLCore** classes and then ensure
that the returned object is of same class.
}
\section{Generic function}{
 sweep(x, MARGIN, STATS, FUN = "-", check.margin = TRUE, ...)
}

\examples{

flq <- FLQuant(rlnorm(90), dim=c(3,10), units='kg')
# Get ratio of max value by year
sweep(flq, 2, apply(flq, 2, max), "/")
}
\seealso{
\link[base]{sweep}
}
\author{
The FLR Team
}
\keyword{methods}
