% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R, R/FLQuant.R
\docType{methods}
\name{sweep}
\alias{sweep}
\alias{sweep,FLArray-method}
\alias{sweep,FLPar-method}
\alias{sweep,FLQuant-method}
\title{Method sweep}
\usage{
\S4method{sweep}{FLArray}(x, MARGIN, STATS, FUN = "-", check.margin = TRUE,
  ...)

\S4method{sweep}{FLPar}(x, MARGIN, STATS, FUN = "-", check.margin = TRUE,
  ...)

\S4method{sweep}{FLQuant}(x, MARGIN, STATS, FUN = "-", check.margin = TRUE,
  ...)
}
\value{
An FLQuant or FLCohort with the same shape as \code{x}, but with the
summary statistics swept out.
}
\description{
Sweep out FLArray Summaries
}
\details{
Return an \code{FLQuant} or \code{FLCohort} obtained from an input object by
sweeping out a summary statistic along the selected dimensions.
}
\section{Generic function}{
 sweep(x, MARGIN, STATS, FUN="-",
check.margin=TRUE, \dots{})
}
\examples{
data(ple4)
mean.f <- apply(harvest(ple4),2,mean)
scaled.f <- sweep(harvest(ple4),2,mean.f,"/")
}
\author{
The FLR Team
}
\seealso{
\code{\link[base]{sweep}}
}
\keyword{methods}

