% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\name{tail,FLQuant-method}
\alias{tail,FLQuant-method}
\alias{head,FLQuant-method}
\title{Returns the first and last parts of an FLQuant.}
\usage{
\S4method{tail}{FLQuant}(x, n = 1, dim = 2, ...)

\S4method{head}{FLQuant}(x, n = 1, dim = 2, ...)
}
\arguments{
\item{x}{The object to extract from, FLQuant.}

\item{n}{The number of elements to extract, numeric.}

\item{dim}{Dimension to extract from, defaults to 2, 'year'.}
}
\value{
An FLQuant with the extracted elements.
}
\description{
Standard tail and head methods can be applied along any dimension of an
FLQuant object.
}
\examples{
x <- FLQuant(1:10)

# Extract the last 3 years
tail(x, 3)

# Extract all but the first 3 years
tail(x, -3)

# Extract the first 3 years
head(x, 3)

# Extract all but the last 3 years
head(x, -3)
}
\seealso{
\link[base:tail]{base::tail}
}
\author{
Iago Mosqueira (WMR)
}
\keyword{methods}
