% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLComp.R
\docType{methods}
\name{transform}
\alias{transform}
\alias{transform,FLComp-method}
\title{Method transform}
\usage{
\S4method{transform}{FLComp}(`_data`, ...)
}
\description{
Transform elements of a complex FLR object
}
\details{
Modification of individual elements of a complex FLR object can be carried
out using \code{transform}. A series of named arguments, corresponding to
the slots to modify, can be provided to the method. Existing slots can be
referred to simply by their names on the right-hand side of the argument
expressions (see example below).
}
\section{Generic function}{
 transform(\_data, \dots{})
}
\examples{
data(ple4)
ple4 <- transform(ple4, m.spwn=m.spwn+0.2)
m.spwn(ple4)
}
\author{
The FLR Team
}
\seealso{
\code{\link[base]{transform}}
}
\keyword{methods}

