% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R, R/FLComp.R
\docType{methods}
\name{units-FLCore}
\alias{units-FLCore}
\alias{units,FLArray-method}
\alias{units<-,FLArray,character-method}
\alias{units,FLPar-method}
\alias{units<-,FLPar,character-method}
\alias{units,FLComp-method}
\alias{units<-,FLComp,list-method}
\alias{units<-,FLComp,character-method}
\alias{units<-,FLComp,function-method}
\title{Method units}
\usage{
\S4method{units}{FLArray}(x)

\S4method{units}{FLArray,character}(x) <- value

\S4method{units}{FLPar}(x)

\S4method{units}{FLPar,character}(x) <- value

\S4method{units}{FLComp}(x)

\S4method{units}{FLComp,list}(x) <- value

\S4method{units}{FLComp,character}(x) <- value

\S4method{units}{FLComp,`function`}(x) <- value
}
\description{
\code{units} attribute for FLQuant and FLArray-derived objects
}
\details{
Objects of \code{FLArray}-based classes (e.g. \code{\link{FLQuant}}) contain a
\code{units} attribute of class \code{character}. This should be used to store
the corresponding units of measurement.  This attribute can be directly accessed
and modified using the \code{units} and \code{units<-} methods.

For complex objects, \code{units} will return a named list containing the
attributes of all \code{FLQuant} slots. \code{units} of a complex object can
be modified for all slots or a subset of them, by passing a named list with
the new values. See examples below.

The complete set of *units* for a complex object can be obtained
as a named *list*.

Assignment of *units* to the *FLQuant* slots of a complex object
can be carried out passing a named *list* or *character* vector containing
the units for the slots to be modified.
}
\section{Generic function}{
 units(x)

units<-(x,value)
}

\examples{

flq <- FLQuant(rnorm(100), dim=c(5,20), units='kg')
units(flq)
units(flq) <- 't'
summary(flq)

# units for a complex object
  data(ple4)
  units(ple4)
  units(ple4) <- list(harvest='hr')

data(ple4)
units(ple4) <- list(harvest="hr")
units(ple4) <- c(harvest="hr")
}
\seealso{
\linkS4class{FLQuant}, \linkS4class{FLPar}, \linkS4class{FLCohort}
}
\author{
The FLR Team
}
\keyword{methods}
