% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLlst-methods.R
\docType{methods}
\name{unlock}
\alias{unlock}
\alias{unlock-methods}
\alias{unlock,FLlst-method}
\title{Unlock an FLlst object so its length can be modified}
\usage{
unlock(object, ...)
}
\arguments{
\item{object}{An \code{\link{FLlst}} object.}

\item{...}{Further arguments passed to methods.}
}
\value{
A modified \code{\link{FLlst}} object with its \code{lock} slot updated.
}
\description{
\code{unlock()} reverses the effect of \code{\link{lock}()} and allows the
length of an \code{\link{FLlst}} object to be modified again.
}
\examples{
fll <- lock(FLlst(a=1:3, b=4:6))
slot(unlock(fll), "lock")
}
\seealso{
\link{lock}, \linkS4class{FLlst}
}
\author{
The FLR Team
}
\keyword{methods}
