% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLModel.R
\docType{methods}
\name{update}
\alias{update}
\alias{update,FLModel-method}
\title{Method update}
\usage{
\S4method{update}{FLModel}(object, ...)
}
\description{
\code{update} is a generic function for updating a model fit using the
same call that generated it. Input arguments can be provided that will alter
the FLModel object accordingly.
}
\section{Generic function}{
 update(object, ...)
}
\examples{

data(nsher)
nsher <- update(nsher, ssb=ssb(nsher) * 1.4)

}
\author{
The FLR Team
}
\seealso{
\linkS4class{FLModel}, \link[stats]{update}
}
\keyword{methods}

