% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLModel.R
\docType{methods}
\name{upperlower}
\alias{upperlower}
\alias{upper}
\alias{upper-methods}
\alias{upper<-}
\alias{upper<--methods}
\alias{lower,FLModel-method}
\alias{upper,FLModel-method}
\title{Extract and modify the \emph{lower} and \emph{upper} FLModel attibutes.}
\usage{
lower(object, ...)

\S4method{lower}{FLModel}(object)

\S4method{upper}{FLModel}(object)
}
\arguments{
\item{object}{Object to extract from or modify}

\item{...}{Other arguments}

\item{value}{New value}
}
\value{
RETURN Lorem ipsum dolor sit amet
}
\description{
Description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
}
\details{
Details: Aliquam sagittis feugiat felis eget consequat.
}
\seealso{
\code{\link{FLModel}}
}
\author{
The FLR Team
}
\keyword{methods}
