% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLPar.R
\docType{methods}
\name{var}
\alias{var}
\alias{var,FLPar,missing,missing,missing-method}
\alias{var,FLPar-method}
\title{Method var}
\usage{
\S4method{var}{FLPar}(x, y = NULL, na.rm = TRUE, use)
}
\description{
Variance of an FLPar
}
\details{
\code{var} computes the variance of an \code{\link{FLPar}} object along the
last dimension (\code{iter}) returning a value for each \code{param}

By default, arguments \code{na.rm} and \code{use} have values of
\code{FALSE} and \code{'all.obs'} respectively. See the
\code{\link[stats]{var}} help page for more information on possible argument
values.
}
\section{Generic function}{
 var(x, y, na.rm, use)
}
\examples{

flp <- FLPar(rnorm(200), params=c('a', 'b'))
var(flp)

}
\author{
The FLR Team
}
\seealso{
\code{\link[stats]{var}}, \code{\linkS4class{FLPar}}
}
\keyword{methods}

