% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\name{which_iter}
\alias{which_iter}
\title{which_iter}
\usage{
which_iter(x, rule)
}
\arguments{
\item{x}{An input object, class 'FLQuant'}

\item{rule}{A logical 'rule', either as a 'character' or a logical 'FLQuant'}
}
\value{
As vector of positions along the 'iter' (6th) dimension
}
\description{
Returns the positions along the 'iter' dimension that match a logical estatement
}
\details{
DETAILS
}
\examples{
# Creates an example object
x <- rnorm(50, FLQuant(runif(5, 10, 20)), 20)
# Looks for iters with a large value
which_iter(x, x > 60)
# rule as a character vector
which_iter(x, '> 60')
# rule has use & and |
which_iter(x, '> 50 & < 60')
which_iter(x, '< -40 | > 40')
}
