% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLComp.R, R/FLlst-methods.R
\docType{methods}
\name{window}
\alias{window}
\alias{window,FLComp-method}
\alias{window,FLQuant-method}
\alias{window,FLlst-method}
\title{Method window}
\usage{
\S4method{window}{FLArray}(x, start = dims(x)$minyear,
  end = dims(x)$maxyear, extend = TRUE, frequency = 1)

\S4method{window}{FLComp}(x, start = dims(x)$minyear, end = dims(x)$maxyear,
  extend = TRUE, frequency = 1)

\S4method{window}{FLlst}(x, ...)
}
\description{
Extract time (year) windows of an FLR object
}
\details{
This method extracts a section of, or extends an \code{FLQuant} or other FLR
objects along the \code{year} dimension. If a frequency is specified, the
new object contains data at only those year steps.

Although objects of class \code{\link{FLQuant}} do have another temporal
dimension, \code{season},  window currently only works along the \code{year}
dimension. To subset/extend along other dimensions, refer to
\link{Extract-FLCore}, \code{\link{trim}}, or \code{\link{expand}}.
}
\section{Generic function}{
 window(x)
}
\examples{
flq <- FLQuant(rnorm(50), dimnames=list(age=1:5, year=1991:2000))
window(flq, start=1995, end=1998)
window(flq, start=1990, end=2010, frequency=2)
}
\author{
The FLR Team
}
\seealso{
\link[stats]{window}, \link{Extract-FLCore}
}
\keyword{methods}

