% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{wireframe}
\alias{wireframe}
\alias{wireframe,FLQuant-method}
\title{3D plot for FLQuant objects}
\usage{
\S4method{wireframe}{formula,FLQuant}(x, data, ...)
}
\arguments{
\item{x}{a \code{formula} formula for lattice}

\item{data}{a \code{FLQuant} object with the values}

\item{...}{Additional argument list to be passed to \code{wireframe}}
}
\value{
a \code{wireframe} plot
}
\description{
Method to plot 3D representations of FLQuant objects
}
\examples{
data(ple4)
wireframe(data~age+year, data=harvest(ple4))
}

