% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\name{yearSample}
\alias{yearSample}
\title{Samples along the year dimension}
\usage{
yearSample(x, size = length(years), years, replace = TRUE, prob = NULL)
}
\arguments{
\item{x}{An FLQuant object.}

\item{size}{Number of samples (years), non-negative integer.}

\item{years}{Optional vector to set as 'year' dimnames in output.}

\item{replace}{should sampling be with replacement? Defaults to TRUE.}

\item{prob}{a vector of probability weights.}
}
\value{
RETURN Description, class
}
\description{
A resample from an FLQuant object along the 'year' dimension is returned. The
'year' dimnames of the output object can be specified, although that is not
needed if the resample is to be assigned in a slot.
}
\examples{
data(ple4)
# Take 20 samples of recent recruitment 
yearSample(rec(ple4)[, ac(2013:2017)], 20)
# Providing 'years' sets the output object dimnames
yearSample(rec(ple4)[, ac(2013:2017)], 20, year=2000:2019)
}
\seealso{
\linkS4class{FLQuant} \code{\link[=sample]{sample()}}
}
\author{
Iago Mosqueira (WMR)
}
\keyword{classes}
