% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLS.R
\docType{methods}
\name{z}
\alias{z}
\alias{z-methods}
\alias{z,FLS-method}
\title{Total mortality z}
\usage{
z(object, ...)

\S4method{z}{FLS}(object, ...)
}
\arguments{
\item{object}{Object to calculate on.}

\item{...}{Any extra arguments.}
}
\value{
An object of the corresponding class, usually \emph{FLQuant}.
}
\description{
Returns the calculation of total mortality, \emph{z}, usually as the sum of
fishing mortality, \emph{f}, and natural mortality, \emph{m}.
}
\examples{
data(ple4)

z(ple4)
}
\seealso{
\code{\link{FLQuant}}
}
\author{
The FLR Team
}
\keyword{methods}
