% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{aliases}
\alias{aliases}
\alias{flq}
\alias{flqp}
\alias{flqd}
\alias{flc}
\alias{flp}
\alias{fls}
\alias{flsl}
\alias{fli}
\alias{flib}
\alias{flsr}
\alias{flqs}
\alias{flcs}
\alias{flss}
\alias{flis}
\alias{flps}
\alias{flb}
\alias{flbs}
\alias{flms}
\alias{flmss}
\alias{pm}
\title{Short aliases for most FLCore classes construction methods}
\usage{
flq(object, ...)
flqp(object, ...)
flqd(object, var, ...)
flc(object, ...)
flp(object, ...)
fls(object, ...)
flsl(object, ...)
fli(object, ...)
flib(object, ...)
flsr(model, ...)
flqs(object, ...)
flcs(object, ...)
flss(object, ...)
flis(object, ...)
flps(object, ...)
flb(object, ...)
flbs(object, ...)
flms(object, ...)
flmss(object, ...)
pm(object, ...)
}
\value{
An object of the requested class
}
\description{
When working interactively, the naming convention of the FLCore classes can
become tiresome. This set of short, lowercase aliases make calling the class
construction methods a bit simpler.
}
\details{
We recomend you use the full name of classes and methods when developing code
that you intend to keep or distribute. A script written using these aliases
can be later 'corrected' by substituting the aliases used in your editor.

The aliases' equivalences are as follows:
\describe{
  \item{flq(...)}{For FLQuant(...)}
  \item{flqp(...)}{For FLQuantPoint(...)}
  \item{flqd(...)}{For FLQuantDistr(...)}
  \item{flc(...)}{For FLCohort(...)}
  \item{flp(...)}{For FLPar(...)}
  \item{fls(...)}{For FLStock(...)}
  \item{flsl(...)}{For FLStockLen(...)}
  \item{fli(...)}{For FLIndex(...)}
  \item{flib(...)}{For FLIndexBiomass(...)}
  \item{flsr(...)}{For FLSR(...)}
  \item{flqs(...)}{For FLQuants(...)}
  \item{flcs(...)}{For FLCohorts(...)}
  \item{flss(...)}{For FLStocks(...)}
  \item{flis(...)}{For FLIndices(...)}
  \item{flps(...)}{For FLPars(...)}
  \item{flb(...)}{For FLBiol(...)}
  \item{flbs(...)}{For FLBiols(...)}
  \item{flms(...)}{For FLModelSim(...)}
  \item{flmss(...)}{For FLModelSims(...)}
  \item{pm(...)}{For predictModel(...)}
}
}
\examples{

flq <- flc(1:10, units="kg")
}
\seealso{
\link{FLQuant}, \link{FLQuantPoint}, \link{FLQuantDistr},
\link{FLCohort}, \link{FLPar}, \link{FLStock}, \link{FLStockLen},
\link{FLIndex}, \link{FLIndexBiomass}, \link{FLSR}, \link{FLQuants},
\link{FLCohorts}, \link{FLStocks}, \link{FLIndices}, \link{FLPars}, 
\link{FLBiol}, \link{FLBiols}, \link{FLModelSim}, \link{FLModelSims},
\link{predictModel}
}
\author{
The FLR Team
}
\keyword{methods}
