% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{ar1rlnorm}
\alias{ar1rlnorm}
\title{Generates a time series of possible bias-corrected lognormal autocorrelated random values}
\usage{
ar1rlnorm(
  rho,
  years,
  iters = 1,
  meanlog = 0,
  sdlog = 1,
  bias.correct = TRUE,
  ...
)
}
\arguments{
\item{rho}{Autocorrelation coefficient.}

\item{years}{Vector of year names.}

\item{iters}{Number of iterations.}

\item{meanlog}{Mean of the series in log space.}

\item{sdlog}{Marginal standard deviation in log space.}

\item{bias.correct}{Should bias-correction be applied? Defaults to TRUE.}
}
\value{
An FLQuant object
}
\description{
Thorston, 2020.
}
\examples{
devs <- ar1rlnorm(rho=0.6, years=2000:2030, iter=500, meanlog=0, sdlog=1)
plot(devs)
}
\references{
Thorson, J. T. Predicting recruitment density dependence and intrinsic growth rate for all fishes worldwide using a data-integrated life-history model. Fish Fish. 2020; 21: 237– 251. https://doi-org.ezproxy.library.wur.nl/10.1111/faf.12427
}
\seealso{
\link{rlnorm}
}
\author{
Iago Mosqueira (WMR), Henning Winker (JRC).
}
\keyword{classes}
