% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{ar1rlnorm}
\alias{ar1rlnorm}
\title{Generates a time series of bias-corrected lognormal autocorrelated random values}
\usage{
ar1rlnorm(rho, years, iters = 1, mean = 0, margSD = 0.6)
}
\arguments{
\item{rho}{Autocorrelation coefficient.}

\item{years}{Vector of year names.}

\item{iters}{Number of iterations.}

\item{mean}{Mean of the series in log space.}

\item{margSD}{}
}
\value{
An FLQuant object
}
\description{
Thorston, 2020.
}
\examples{
resid <- ar1rlnorm(rho=0.6, years=2000:2030, iter=500, mean=0, margSD=0.6)
plot(resid)
}
\seealso{
\link{rlnorm}
}
\author{
Iago Mosqueira (WMR), Henning Winker (JRC).
}
\keyword{classes}
