% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\docType{methods}
\name{bias}
\alias{bias}
\alias{bias,FLParJK-method}
\title{Bias of estimates through jackknife}
\usage{
\S4method{bias}{FLQuantJK}(x)

\S4method{bias}{FLParJK}(x)
}
\arguments{
\item{x}{An object holding estimates obtained through jackknife}
}
\value{
A value for the mean bias
}
\description{
Description: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.
}
\details{
Details: Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa,
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien,
feugiat blandit arcu in, dapibus dictum felis.

\deqn{\widehat{Bias}_{(\theta)} = (n - 1)((\frac{1}{n}\sum\limits_{i=1}^n\hat{\theta}_{(i)})-\hat{\theta})}{}
}
\examples{

}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
