% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLModel.R
\docType{methods}
\name{BIC}
\alias{BIC}
\alias{BIC,FLModel-method}
\title{Method BIC}
\usage{
\S4method{BIC}{FLModel}(object)
}
\description{
Bayesian information criterion (BIC) method
}
\details{
A method to calculate the Bayesian information criterion (BIC), also known
as Schwarz's Bayesian criterion of an \link{FLModel} object from the value
of the obtained log-likelihood stored in its \code{logLik} slot.
}
\section{Generic function}{
 BIC(object)
}
\examples{
data(nsher)
BIC(nsher)
}
\author{
The FLR Team
}
\seealso{
\link[stats]{BIC}, \link[stats]{AIC}, \link{FLModel},
\link[stats]{logLik}
}
\keyword{methods}

