% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\docType{methods}
\name{coerce-methods}
\alias{coerce-methods}
\alias{coerce,FLArray,data.frame-method}
\alias{coerce,FLPar,data.frame-method}
\alias{coerce,data.frame,FLQuant-method}
\alias{coerce,FLBiol,FLStock-method}
\title{Convert Objects Between Classes}
\arguments{
\item{object}{Object to be converted.}

\item{Class}{Name of the class to convert the object to, \code{character}.}
}
\value{
An object of the requested class.
}
\description{
Objects of various \strong{FLCore} classes can be converted into other classes,
both basic R ones, like \code{data.frame}, and defined in the package. For the
specifics of the precise calculations carried out for each pair of classes,
see Details.
}
\section{FLArray to data.frame}{

 The six dimensions of an `FLArray` are converted into seven columns, named
 `quant` (or any other name given to the first dimension in the object),
 `year`, `unit`, `season`, `area`, `iter`  and `data`. The last one contains
 the actual numbers stored in the array. `units` are stored as an attribute
 to the `data.frame`. The `year` and `data` columns are of type `numeric`,
 while all others are `factor`.
}

\section{FLPar to data.frame}{

 The two or more dimensions of an *FLPar* objects are converted into three or
 more columns. For a 2D objects, they are named *params*, *iter* and *data*.
 The last one contains the actual numbers stored in the array, in a column
 type `numeric`, while all others are `factor`.
}

\section{data.frame to FLQuant}{

 A *data.frame* with the right column names is converted into an *FLQuant*
 object with missing values being added. Missing columns are assumed to
 contain the default dimnames in *FLQuant*.
}

\section{*FLBiol* to *FLStock*}{

 - *n* = *stock.n*
 - *wt* = *stock.wt*
 - *m* = *m*
 - *mat* = *mat()*
 - *m.spwm*, *harvest.spwn* = *spwn*
}

\examples{
# from FLQuant to data.frame
as(FLQuant(rnorm(100), dim=c(5, 20)), "data.frame")

# from FLPar to data.frame
as(FLPar(phi=rnorm(10), rho=rlnorm(10)), "data.frame")

# from data.frame to FLQuant
as(data.frame(age=rep(1:4, each=3), year=2011:2013, data=rnorm(12)), "FLQuant")

}
\seealso{
\link[base:as]{base::as}, \link[base:coerce]{base::coerce}
}
\author{
The FLR Team
}
\keyword{methods}
