% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{cpue}
\alias{cpue}
\alias{cpue-methods}
\alias{cpue,FLStock,missing-method}
\alias{cpue,FLStock-method}
\alias{survey,FLStock,FLIndex-method}
\alias{cpue-FLStock-method}
\alias{survey,FLStock,FLIndexBiomass-method}
\title{cpue, a method to generate an observation of a CPUE index of abundance}
\usage{
cpue(object, index, ...)

\S4method{cpue}{FLStock,missing}(
  object,
  sel.pattern = harvest(object),
  effort = units(harvest(object)),
  mass = TRUE
)

survey(object, index, ...)

\S4method{survey}{FLStock,FLIndex}(
  object,
  index,
  sel = sel.pattern(index),
  mass = FALSE,
  timing = mean(range(index, c("startf", "endf"))),
  index.q = index@index.q
)

\S4method{survey}{FLStock,FLIndexBiomass}(
  object,
  index,
  sel = sel.pattern(index),
  ages = ac(seq(range(index, c("min")), range(index, c("max")))),
  timing = mean(range(index, c("startf", "endf"))),
  index.q = index@index.q
)
}
\arguments{
\item{object}{The object on which to draw the observation}

\item{effort}{Units of index to use to mimic effort series in the fishery, "f" or "hr"}

\item{mass}{Is the index to be in weight at age?}

\item{sel}{The selectivity of the survey, defaults to be 1 for all ages.}
}
\value{
An FLQuant for the index of abundance, age-disaggregated

An FLQuant for the index of abundance
}
\description{
The observation of stock abundance by CPUE series from commercial fleets is an
important step in the generation of management advice that needs to replicated
on an Operating Model during any simulation exercise. This method gemnerates
an observation of biomass or numbers-at-age from an FLstock being used as OM.

Description: Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend
odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta
libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique.
}
\details{
Details: Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa, 
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in 
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae 
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien, 
feugiat blandit arcu in, dapibus dictum felis.
}
\examples{

data(ple4)

cpue(ple4)
# Am aggregated biomass CPUE
quantSums(cpue(ple4))

\dontrun{
plot(FLQuants(om=stock(ple4), cpue=quantSums(cpue(ple4)),
  hr=quantSums(cpue(ple4, effort="hr"))))
}

\dontrun{
plot(FLQuants(om=stock(ple4), survey=quantSums(survey(ple4) * stock.wt(ple4)),
 cpue=quantSums(cpue(ple4)), hr=quantSums(cpue(ple4, effort="hr"))))
}
data(ple4)
data(ple4.index)
# survey(ple4, ple4.index)

}
\seealso{
\link{FLComp}

\link{FLComp}
}
\author{
Laurie Kell & Iago Mosqueira, FLR Team.

The FLR Team
}
\keyword{classes}
\keyword{methods}
