% $Id: data.Rd,v 1.5.2.4 2006/10/19 15:40:36 iagoazti Exp $
\name{data}
\alias{ple4}
\alias{bt4}
\alias{ple4sex}
\alias{ple4.sr}
\alias{ple4.biol}
\alias{ple4.index}
\alias{ple4.indices}
\alias{north.beam}
\alias{south.beam}
\title{FLCore datasets}
\description{
    \item{ple4}{A dataset of North Sea (ICES Area IV) plaice catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
    \item{ple4sex}{A dataset of North Sea (ICES Area IV) plaice disaggregated by sex catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
    \item{ple4.sr}{A dataset of North Sea (ICES Area IV) plaice stock and recruitment.}
    \item{ple4.index}{A dataset of North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndex object.}
    \item{ple4.indices}{A dataset of two North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndices object.}
    \item{ple4.index}{A dataset of North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndex object.}
    \item{ple4.indices}{A dataset of two North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndices object.}
    \item{ple.biol}{A dataset of North Sea plaice population, numbers, natural mortality, mass and fecundity-at-age, contained in an FLBiol object.}
    \item{north.beam}{A dataset of class FLFleet corresponding to the northern North Sea beam trawl fleet fishing on plaice.}
    \item{south.beam}{A dataset of class FLFleet corresponding to the southern North Sea beam trawl fleet fishing on sole and plaice.}

	Datasets can be loaded by issuing \code{data(ple4)} and \code{data(bt4)} respectively.
}
\seealso{
\code{\link{FLStock}}, \code{\link{FLSR}}, \code{\link{FLIndex}, \code{\link{FLStock}}, \code{\link{FLIndex}}, \code{\link{FLBiol}}, \code{\link{FLFleet}}}
}
\references{
ICES.
}
\keyword{datasets}
