% $Id$
\name{data}
\alias{ple4}
\alias{ple4sex}
\alias{ple4.biol}
\alias{ple4.index}
\alias{ple4.indices}
\alias{nsher}
\alias{bt4}
\title{FLCore datasets}
\description{
    \item{ple4}{A dataset of North Sea (ICES Area IV) plaice catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
    \item{ple4sex}{A dataset of North Sea (ICES Area IV) plaice disaggregated by sex catch, yield, landings, discards, natural mortality, weight-at-age and maturity, together with the VPA estimated abundances and fishing mortalities, contained in an FLStock object.}
    \item{ple4.index}{A dataset of North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndex object.}
    \item{ple4.indices}{A dataset of two North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndices object.}
    \item{ple4.index}{A dataset of North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndex object.}
    \item{ple4.indices}{A dataset of two North Sea (ICES Area IV) plaice survey catch per unit effort, index and index variance, contained in an FLIndices object.}
    \item{ple.biol}{A dataset of North Sea plaice population, numbers, natural mortality, mass and fecundity-at-age, contained in an FLBiol object.}
	\item{nsher}{A dataset of class \code{FLStock} for autumn spawning North Sea herring.}
    \item{nsher.sr}{A dataset of North Sea autumn spawning herring stock and recruitment relationship. Fitted using the 'Ricker' model.}
	\item{nsher.biol}{A dataset of class \code{FLBiol} for North Sea herring.}
    \item{bt4}{A beam trawl fleet catching plaice and sole in North Sea area 4.}

	Datasets can be loaded by issuing the \code{data} command, like in \code{data(ple4)}.
}
\seealso{
\code{\link{FLStock-class}}, \code{\link{FLSR-class}}, \code{\link{FLIndex-class}, \code{\link{FLStock-class}}, \code{\link{FLIndex-class}}, \code{\link{FLBiol-class}}, \code{\link{FLFleet-class}}}
}
\references{
ICES.
}
\keyword{datasets}
