% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLArray.R
\name{dbind}
\alias{dbind}
\alias{dbind-methods}
\alias{dbind,FLArray,FLArray-method}
\alias{qbind}
\alias{ybind}
\alias{ubind}
\alias{sbind}
\alias{abind}
\alias{ibind}
\title{Methods for binding objects of array classes along a given dimension}
\usage{
dbind(x, y, ...)

\S4method{dbind}{FLArray,FLArray}(x, y, ..., dim = 1)

qbind(...)

ybind(...)

ubind(...)

sbind(...)

abind(...)

ibind(...)
}
\arguments{
\item{x}{First object to bind}

\item{y}{Second object to bind}

\item{...}{Other objects to bind}

\item{dim}{Dimension to bind on, *numeric* or *character*.}
}
\value{
An object of the same class as the inputs
}
\description{
These methods can bind two or more objects of array-based classes
(e.g. FLQuant), along the specified dimension.
}
\details{
The objects to bind must contain the same dimmames in all dimensions other
than that used to bind, while dimnames in the selected one must differ. See
the examples below for correct and incorrect uses.

Object are bound in the order they are provided, with no attempt to sort
according to the dimnames of the chosen dimension.

The implementation is based around a single method (\emph{dbind}), that operates
along the dimension position or name indicated by the \emph{dim} argument. A
series of shortcut functions call the method for specific dimensions, with
names related to the dimensions name they operate on (e.g. ybind for \emph{year}).
}
\examples{

# By iter
x <- FLQuant(rnorm(80000), dim=c(4,20,1,1,1,1000))
y <- FLQuant(rnorm(80000), dim=c(4,20,1,1,1,1000))
  dimnames(y) <- list(iter=1001:2000)
ibind(x,y)

# By quant (age)
x <- FLQuant(1, dimnames=list(age=1:3, year=1:10))
y <- FLQuant(2, dimnames=list(age=4:12, year=1:10))
qbind(x, y)

# By year
x <- FLQuant(1, dimnames=list(age=1:3, year=1:10))
y <- FLQuant(2, dimnames=list(age=1:3, year=11:20))
z <- FLQuant(3, dimnames=list(age=1:3, year=21:30))
ybind(x, y, z)

# By season
x <- FLQuant(1, dimnames=list(year=1:10, season=1:2))
y <- FLQuant(2, dimnames=list(year=1:10, season=3:4))
sbind(x, y)
}
\seealso{
\code{\link{FLQuant}} \code{\link{FLarray}}
}
\author{
Iago Mosqueira (EC JRC)
}
\keyword{methods}
