% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLQuant.R
\docType{methods}
\name{dimSummaries}
\alias{areaMeans}
\alias{areaSums}
\alias{areaSums,FLQuant-method}
\alias{areaVars}
\alias{dimMeans}
\alias{dimSummaries}
\alias{dimSums}
\alias{dimVars}
\alias{iterCVs}
\alias{iterMeans}
\alias{iterMedians}
\alias{iterVars}
\alias{quantMeans}
\alias{quantSums}
\alias{quantSums,FLQuant-method}
\alias{quantSums-methods}
\alias{quantVars}
\alias{seasonMeans}
\alias{seasonSums}
\alias{seasonSums,FLQuant-method}
\alias{seasonVars}
\alias{unitMeans}
\alias{unitSums}
\alias{unitSums,FLQuant-method}
\alias{unitVars}
\alias{yearMeans}
\alias{yearSums}
\alias{yearSums,FLQuant-method}
\alias{yearVars}
\title{Summaries by dimension}
\usage{
quantSums(x, ...)

yearSums(x, ...)

unitSums(x, ...)

seasonSums(x, ...)

areaSums(x, ...)

dimSums(x, ...)

quantMeans(x, ...)

yearMeans(x, ...)

unitMeans(x, ...)

seasonMeans(x, ...)

areaMeans(x, ...)

iterMeans(x, ...)

dimMeans(x, ...)

quantVars(x, ...)

yearVars(x, ...)

unitVars(x, ...)

seasonVars(x, ...)

areaVars(x, ...)

iterVars(x, ...)

dimVars(x, ...)

iterMedians(x, ...)

iterCVs(x, ...)

\S4method{quantSums}{FLQuant}(x, na.rm = TRUE)

\S4method{yearSums}{FLQuant}(x, na.rm = TRUE)

\S4method{unitSums}{FLQuant}(x, na.rm = TRUE)

\S4method{seasonSums}{FLQuant}(x, na.rm = TRUE)

\S4method{areaSums}{FLQuant}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An object.}

\item{na.rm}{Should NAs be removed before calculation? Defaults to TRUE.}
}
\description{
Methods to compute various summary calculations (sum, mean, variance) over
selected dimensions of objects from any array-based classes
(e.g. \code{FLQuant}, \code{FLPar}). These methods return an object of the
same dimensions as the input but with of length one in the dimension chosen
to operate along.
}
\details{
This set of methods computes three different summaries (sum, mean and
variance) of an \code{FLQuant} object along each of the six dimensions
(quant, year, unit, season, area, or iter). Medians can also be computed
along the sixth dimension, \code{iter}.

These methods simply encapsulate a call to \code{\link[base]{apply}} with
the corresponding dimensions and function.

Methods along the iter dimension are also defined for objects of class
\code{FLPar}.

Methods to operate over the first dimension refer to it as the \code{quant}
dimension, regardless of the actual name used in the object.
}
\section{Generic methods}{

quantSums(x), quantMeans(x), quantVars(x)
yearSums(x), yearMeans(x), yearVars(x)
unitSums(x), unitMeans(x), unitVars(x)
seasonSums(x), seasonMeans(x), seasonVars(x)
areaSums(x), areaMeans(x), areaVars(x)
iterMeans(x), iterVars(x), iterMedians(x)
dimSums(x), dimMeans(x), dimVars(x)
}
\examples{

 flq <- FLQuant(rnorm(4000), dim=c(5,10,2,2,2,10), quant='age')
 quantSums(flq)
 quantMeans(flq)
 yearSums(flq)
 iterMeans(flq)
 dim(quantSums(flq))
}
\author{
The FLR Team
}
\seealso{
\link{FLQuant}, \link[base]{sum}, \link[base]{mean},
\link[stats]{var}
}
\keyword{methods}

