% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLQuant.R,
%   R/FLQuantDistr.R, R/FLPar.R
\docType{methods}
\name{dimSummaries}
\alias{dimSummaries}
\alias{quantSums}
\alias{quantSums-methods}
\alias{quantSums,FLQuant-method}
\alias{yearSums}
\alias{unitSums}
\alias{seasonSums}
\alias{areaSums}
\alias{iterSums}
\alias{dimSums}
\alias{quantMeans}
\alias{yearMeans}
\alias{unitMeans}
\alias{seasonMeans}
\alias{areaMeans}
\alias{iterMeans}
\alias{dimMeans}
\alias{quantVars}
\alias{yearVars}
\alias{unitVars}
\alias{seasonVars}
\alias{areaVars}
\alias{iterVars}
\alias{dimVars}
\alias{iterMedians}
\alias{iterCVs}
\alias{iterProb}
\alias{quantSums,FLQuant-method}
\alias{yearSums,FLQuant-method}
\alias{unitSums,FLQuant-method}
\alias{seasonSums,FLQuant-method}
\alias{areaSums,FLQuant-method}
\alias{iterSums,FLQuant-method}
\alias{quantMeans,FLQuant-method}
\alias{yearMeans,FLQuant-method}
\alias{unitMeans,FLQuant-method}
\alias{seasonMeans,FLQuant-method}
\alias{areaMeans,FLQuant-method}
\alias{iterMeans,FLQuant-method}
\alias{iterMedians,FLQuant-method}
\alias{quantVars,FLQuant-method}
\alias{yearVars,FLQuant-method}
\alias{unitVars,FLQuant-method}
\alias{seasonVars,FLQuant-method}
\alias{areaVars,FLQuant-method}
\alias{iterVars,FLQuant-method}
\alias{iterCVs,FLQuant-method}
\alias{iterProb,FLQuant-method}
\alias{yearSums,FLQuantDistr-method}
\alias{unitSums,FLQuantDistr-method}
\alias{seasonSums,FLQuantDistr-method}
\alias{areaSums,FLQuantDistr-method}
\alias{yearMeans,FLQuantDistr-method}
\alias{unitMeans,FLQuantDistr-method}
\alias{seasonMeans,FLQuantDistr-method}
\alias{areaMeans,FLQuantDistr-method}
\alias{iterMeans,FLQuantDistr-method}
\alias{iterMedians,FLQuantDistr-method}
\alias{quantVars,FLQuantDistr-method}
\alias{yearVars,FLQuantDistr-method}
\alias{unitVars,FLQuantDistr-method}
\alias{seasonVars,FLQuantDistr-method}
\alias{areaVars,FLQuantDistr-method}
\alias{iterVars,FLQuantDistr-method}
\alias{iterMeans,FLPar-method}
\alias{iterMedians,FLPar-method}
\alias{iterVars,FLPar-method}
\alias{iterSums,FLPar-method}
\title{Summaries by dimension}
\usage{
quantSums(x, ...)

yearSums(x, ...)

unitSums(x, ...)

seasonSums(x, ...)

areaSums(x, ...)

iterSums(x, ...)

dimSums(x, ...)

quantMeans(x, ...)

yearMeans(x, ...)

unitMeans(x, ...)

seasonMeans(x, ...)

areaMeans(x, ...)

iterMeans(x, ...)

dimMeans(x, ...)

quantVars(x, ...)

yearVars(x, ...)

unitVars(x, ...)

seasonVars(x, ...)

areaVars(x, ...)

iterVars(x, ...)

dimVars(x, ...)

iterMedians(x, ...)

iterCVs(x, ...)

iterProb(x, ...)

\S4method{quantSums}{FLQuant}(x, na.rm = TRUE)

\S4method{yearSums}{FLQuant}(x, na.rm = TRUE)

\S4method{unitSums}{FLQuant}(x, na.rm = TRUE)

\S4method{seasonSums}{FLQuant}(x, na.rm = TRUE)

\S4method{areaSums}{FLQuant}(x, na.rm = TRUE)

\S4method{iterSums}{FLQuant}(x, na.rm = TRUE)

\S4method{quantMeans}{FLQuant}(x, na.rm = TRUE)

\S4method{yearMeans}{FLQuant}(x, na.rm = TRUE)

\S4method{unitMeans}{FLQuant}(x, na.rm = TRUE)

\S4method{seasonMeans}{FLQuant}(x, na.rm = TRUE)

\S4method{areaMeans}{FLQuant}(x, na.rm = TRUE)

\S4method{iterMeans}{FLQuant}(x, na.rm = TRUE)

\S4method{iterMedians}{FLQuant}(x, na.rm = TRUE)

\S4method{quantVars}{FLQuant}(x, na.rm = TRUE)

\S4method{yearVars}{FLQuant}(x, na.rm = TRUE)

\S4method{unitVars}{FLQuant}(x, na.rm = TRUE)

\S4method{seasonVars}{FLQuant}(x, na.rm = TRUE)

\S4method{areaVars}{FLQuant}(x, na.rm = TRUE)

\S4method{iterVars}{FLQuant}(x, na.rm = TRUE)

\S4method{iterCVs}{FLQuant}(x, na.rm = TRUE)

\S4method{iterProb}{FLQuant}(x, na.rm = TRUE)

\S4method{yearSums}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{unitSums}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{seasonSums}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{areaSums}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{yearMeans}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{unitMeans}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{seasonMeans}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{areaMeans}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{iterMeans}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{iterMedians}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{quantVars}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{yearVars}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{unitVars}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{seasonVars}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{areaVars}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{iterVars}{FLQuantDistr}(x, na.rm = TRUE)

\S4method{iterMeans}{FLPar}(x, na.rm = TRUE)

\S4method{iterMedians}{FLPar}(x, na.rm = TRUE)

\S4method{iterVars}{FLPar}(x, na.rm = TRUE)

\S4method{iterSums}{FLPar}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An object.}

\item{na.rm}{Should NAs be removed before calculation? Defaults to TRUE.}
}
\description{
Methods to compute various summary calculations (sum, mean, variance) over
selected dimensions of objects from any array-based classes
(e.g. \code{FLQuant}). These methods return an object of the
same dimensions as the input but with length one in the dimension chosen
to operate along.
}
\details{
This set of methods computes three different summaries (sum, mean and
variance) of an \code{FLQuant} object along each of the six dimensions
(quant, year, unit, season, area, or iter). Medians and CVs can also be
computed along the sixth dimension, \code{iter}.

These methods encapsulate a call to \code{\link[base]{apply}} with
the corresponding dimensions and function: \code{\link[base]{mean}}, 
\code{\link[stats]{median}}, \code{\link[stats]{var}}, and
\code{\link[base]{sum}}, while \code{iterCVs} are computed as
\code{sqrt(iterVars) / iterMeans}.

In contrast with R standard behaviour, the sum of a dimension where all
elements are \code{NA} will be \code{NA} and not 0. See example below.

Methods working along the iter dimension are also defined for objects of class
\code{FLPar}.

Methods to operate over the first dimension refer to it as the \code{quant}
dimension, regardless of the actual name used in the object.
}
\section{Generic methods}{

quantSums(x), quantMeans(x), quantVars(x)
yearSums(x), yearMeans(x), yearVars(x)
unitSums(x), unitMeans(x), unitVars(x)
seasonSums(x), seasonMeans(x), seasonVars(x)
areaSums(x), areaMeans(x), areaVars(x)
iterMeans(x), iterVars(x), iterMedians(x), iterSums(x)
dimSums(x), dimMeans(x), dimVars(x)
}

\examples{

flq <- FLQuant(rnorm(4000), dim=c(5,10,2,2,2,10), quant='age')

quantSums(flq)
quantMeans(flq)
yearSums(flq)
iterMeans(flq)
dim(quantSums(flq))

# NA dims stay as NA when summed along
x <- FLQuant(c(NA, NA, NA, rnorm(6)), dim=c(3, 3))
quantSums(x)
# although in fact a sum of no elements (as na.rm=TRUE) is zero
apply(x, 2:6, sum, na.rm=TRUE)
}
\seealso{
\link{FLQuant}, \link[base]{sum}, \link[base]{mean},
\link[stats]{var}
}
\author{
The FLR Team
}
\keyword{methods}
