% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{evalPredictModel}
\alias{evalPredictModel}
\title{Evaluates a predictModel slot inside the object cointaining it}
\usage{
evalPredictModel(object, slot)
}
\arguments{
\item{object}{The FLR S4 object holding the `predictModel` slot.}

\item{slot}{The name of the slot to be evaluated, as a character.}
}
\value{
The result of evaluating the model, usually an `FLQuant`
}
\description{
Models in objects of the \link{predictModel} class can make use of slots and
methods of the FLR class in which it is contained as a slot. This function can
be used by methods wishing to evaluate a single `predictModel` slot in the
context of the class it is part of.
}
\author{
The FLR Team
}
\seealso{
\link{predictModel}
}
\keyword{utilities}

