% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLComp.R, R/FLI.R, R/FLPar.R, R/FLQuantDistr.R, R/FLStock.R, R/FLlst-methods.R
\docType{methods}
\name{Extract}
\alias{$<-,FLlst,ANY-method}
\alias{$<-,FLlst-method}
\alias{Extract}
\alias{[,FLArray,ANY,ANY,ANY-method}
\alias{[,FLArray,array,missing,missing-method}
\alias{[,FLComp,ANY,ANY,ANY-method}
\alias{[,FLI,ANY,ANY,ANY-method}
\alias{[,FLPar,ANY,ANY,ANY-method}
\alias{[,FLQuantDistr,ANY,ANY,ANY-method}
\alias{[,FLQuantDistr,array,missing,missing-method}
\alias{[,FLStock,ANY,ANY,ANY-method}
\alias{[,FLlst,ANY,missing,ANY-method}
\alias{[<-,FLArray,ANY,ANY,ANY-method}
\alias{[<-,FLArray,ANY,ANY,FLArray-method}
\alias{[<-,FLComp,ANY,ANY,ANY-method}
\alias{[<-,FLQuantDistr,ANY,ANY,ANY-method}
\alias{[<-,FLQuantDistr,array,missing,missig-method}
\alias{[<-,FLlst,ANY,missing,ANY-method}
\alias{[[,FLlst,ANY,missing,ANY-method}
\alias{[[<-,FLlst,ANY,missing-method}
\alias{`[<-,FLArray,ANY,ANY,ANY-method`}
\alias{`[<-,FLArray,ANY,ANY,FLArray-method`}
\title{Extract}
\usage{
\S4method{[}{FLArray,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLArray,array,missing,missing}(x, i)

\S4method{[}{FLArray,ANY,ANY,ANY}(x, i, j, k, l, m, n, ...) <- value

\S4method{[}{FLArray,ANY,ANY,FLArray}(x, i, j, k, l, m, n, ...) <- value

\S4method{[}{FLComp,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLComp,ANY,ANY,ANY}(x, i, j, k, l, m, n, ...) <- value

\S4method{[}{FLI,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLPar,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLQuantDistr,ANY,ANY,ANY}(x, i, j, k, l, m, n)

\S4method{[}{FLQuantDistr,array,missing,missing}(x, i)

\S4method{[}{FLStock,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[[}{FLlst,ANY,missing}(x, i, j) <- value

\S4method{$}{FLlst}(x, name) <- value

\S4method{[}{FLlst,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{FLlst,ANY,missing,ANY}(x, i, j, drop)
}
\description{
Extract or replace parts of an FLR Object
}
\details{
Operators acting on FLQuant, FLCohort, FLPar, FLComp, and derived classes to
extract or replace sections of an object.

Please note the differences between referencing sections of an object by
position using values of class \code{numeric}, or by using dimnames of class
\code{character}. See examples below.

All classes that are derived from \code{FLComp} (for example, \code{FLStock}
and \code{FLBiol}) can be subset along the six dimensions of their
\code{FLQuant} slots.

Classes that are derived from \code{FLlst} (for example, \code{FLStocks} and
\code{FLBiols}) can be subset in a similar way to ordinary list objects.
}
\section{Generic function}{
 \describe{ \item{}{[x,i,j,drop]}
\item{}{[<-(x,i,j,value)} \item{}{[[<-(x,i,j,value)}
\item{}{\$<-(x,name,value)} }
}
\examples{
flq <- FLQuant(rnorm(200), dimnames=list(age=0:4, year=1991:2000,
  season=1:4))

# Extracting by position...
  flq[1,]
  flq[,1:5]
  flq[1:2,,,c(1,3)]

# ...by dimnames
  flq['0',]
  flq[,'1991']
  flq[,as.character(1991:1995),,'1']

# Replacing part of the object
  flq['0',,,1]<-0
data(ple4)
# Subset first five years
summary(ple4[,1:5])
# FLPar
flp <- FLPar(rlnorm(90), dimnames=list(params=letters[1:3], iter=30), units='NA')
# Select first parameter
flp[1,]
flp['a',]
# FLStock
data(ple4)
# Method for FLStock distinguishes first dimension 'age'
ple <- ple4[1:4,]
summary(ple4)
# Be AWARE that aggregated slots (catch, landings, ...) are NOT recomputed
all.equal(catch(ple), quantSums(catch.n(ple) * catch.wt(ple)))
# FLStocks (& any FLlst)
fll <- FLStocks(PLE4=ple4, OLD=ple4[,1:20])
summary(fll)
# Subset FLStocks
summary(fll[1])
# Extract single element
summary(fll[[1]])
}
\author{
The FLR Team
}
\seealso{
\link[base]{Extract}
}
\keyword{methods}

