% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R
\docType{methods}
\name{Extract}
\alias{Extract}
\alias{[,FLArray,ANY,ANY,ANY-method}
\alias{[,FLArray,array,missing,missing-method}
\alias{[<-,FLArray,ANY,ANY,ANY-method}
\alias{[<-,FLArray,ANY,ANY,FLArray-method}
\alias{`[<-,FLArray,ANY,ANY,ANY-method`}
\alias{`[<-,FLArray,ANY,ANY,FLArray-method`}
\title{Extract}
\usage{
\S4method{[}{FLArray,ANY,ANY,ANY}(x, i, j, k, l, m, n, ..., drop = FALSE)

\S4method{[}{FLArray,array,missing,missing}(x, i)

\S4method{[}{FLArray,ANY,ANY,ANY}(x, i, j, k, l, m, n, ...) <- value

\S4method{[}{FLArray,ANY,ANY,FLArray}(x, i, j, k, l, m, n, ...) <- value
}
\arguments{
\item{x}{object from which to extract or replace element(s)}

\item{i, j, k, l, m, n, ...}{indices specifying elements to extract or replace.}

\item{drop}{If 'TRUE' the result is coerced to the lowest possible dimension, and so
might change class (e.g. drop='TRUE' on an \code{FLQuant} might return an
\code{array} of less dimensions, a \code{matrix} or a \code{vector}.}

\item{value}{An object of a similar or simpler class than 'x'.}

\item{name}{See \link[base]{Extract} for further details.}
}
\description{
Extract or replace parts of an FLR Object
}
\details{
Operators acting on FLQuant, FLCohort, FLPar, FLComp, and derived classes to
extract or replace sections of an object.

Please note the differences between referencing sections of an object by
position using values of class \code{numeric}, or by using dimnames of class
\code{character}. See examples below.

All classes that are derived from \code{FLComp} (for example, \code{FLStock}
and \code{FLBiol}) can be subset along the six dimensions of their
\code{FLQuant} slots.

Classes that are derived from \code{FLlst} (for example, \code{FLStocks} and
\code{FLBiols}) can be subset in a similar way to ordinary list objects.
}
\section{Generic function}{
 \describe{ \item{}{[x,i,j,drop]}
\item{}{[<-(x,i,j,value)} \item{}{[[<-(x,i,j,value)}
\item{}{\$<-(x,name,value)} }
}
\examples{

flq <- FLQuant(rnorm(200), dimnames=list(age=0:4, year=1991:2000,
  season=1:4))

# Extracting by position...
  flq[1,]
  flq[,1:5]
  flq[1:2,,,c(1,3)]

# ...by dimnames
  flq['0',]
  flq[,'1991']
  flq[,as.character(1991:1995),,'1']

# Dimensions of length one can be drop
  flq[1, drop=TRUE]

# Replacing part of the object
  flq['0',,,1]<-0

}
\author{
The FLR Team
}
\seealso{
\link[base]{Extract}
}
\keyword{methods}

