\name{fbar}
\alias{fbar}
\alias{fbar-methods}
\alias{fbar,FLStock-method}
\title{Calculates mean harvest rate or fishing mortality}
\description{
  The mean harvest rate of fishing mortality
}
\usage{
fbar(object, ...)
}
\arguments{
  \item{object}{An FLStock object}
  \item{...}{Any extra arguments}
}
\details{
}
\value{
}
\author{FLR Team}
\examples{
}
\keyword{methods}
