% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{ffwd}
\alias{ffwd}
\title{Project forward an FLStock for a fbar target}
\usage{
ffwd(object, sr, fbar = control, control = fbar, deviances = "missing")
}
\arguments{
\item{object}{An *FLStock*}

\item{sr}{A stock-recruit relationship, *FLSR* or *predictModel*.}

\item{fbar}{Yearly target for average fishing mortality, *FLQuant*.}

\item{control}{Yearly target for average fishing mortality, *fwdControl*.}

\item{deviances}{Deviances for the strock-recruit relationsip, *FLQuant*.}
}
\value{
The projected *FLStock* object.
}
\description{
Projection of an FLStock object for a fishing mortality target does not
always require the features of fwd().Fast-forward an FLStock object for a fishing mortality yearly target only.
}
\examples{
data(ple4)
sr <- predictModel(model=bevholt, params=FLPar(a=140.4e4, b=1.448e5))
# Project for fixed Fbar=0.21
run <- ffwd(ple4, sr=sr, fbar=FLQuant(0.21, dimnames=list(year=1958:2017)))
plot(run)
}
\seealso{
\link{fwd}
}
\author{
Iago MOSQUEIRA (MWR), Henning WINKEL (JRC).
}
\keyword{classes}
