% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesArr.R
\docType{class}
\name{FLArray}
\alias{FLArray}
\alias{FLArray-class}
\title{Class FLArray}
\description{
A basic 6D array class. No objects of this class are created in
FLCore, as it is used only for method inheritance.
}
\section{Slots}{
 \describe{
\item{.Data}{Internal S4 data representation, of class \code{array}.}
}
}

\section{Validity}{
 \describe{
\item{Dimensions:}{Array must have 6 dimensions}
\item{Content:}{Array must be of class \code{numeric}}
}
}

\seealso{
\code{\linkS4class{FLQuant}}, \code{\linkS4class{FLCohort}}
}
\author{
The FLR Team
}
\keyword{classes}
