% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classesComp.R
\docType{class}
\name{FLBiol}
\alias{FLBiol}
\alias{FLBiol,FLQuant-method}
\alias{FLBiol,missing-method}
\alias{FLBiol-class}
\alias{FLBiol-methods}
\alias{desc,FLBiol-method}
\alias{desc<-,FLBiol,character-method}
\alias{fec,FLBiol-method}
\alias{fec<-,FLBiol,FLQuant-method}
\alias{m,FLBiol-method}
\alias{m<-,FLBiol,FLQuant-method}
\alias{n,FLBiol-method}
\alias{n<-,FLBiol,FLQuant-method}
\alias{name,FLBiol-method}
\alias{name<-,FLBiol,character-method}
\alias{range,FLBiol-method}
\alias{range<-,FLBiol,numeric-method}
\alias{spwn,FLBiol-method}
\alias{spwn<-,FLBiol,FLQuant-method}
\alias{wt,FLBiol-method}
\alias{wt<-,FLBiol,FLQuant-method}
\title{Class FLBiol}
\description{
A class for modelling age / length or biomass structured populations.
}
\details{
The \code{FLBiol} class is a representation of a biological fish population
(in contrast to \code{\link{FLStock}}, which represents a fish stock; the
two may not fully coincide). This includes information on abundances,
natural mortlity and fecundity.
}
\section{Slots}{
 \describe{ \item{n}{Numbers in the population.
\code{FLQuant}.} \item{m}{Mortality rate of the population. \code{FLQuant}.}
\item{wt}{Mean weight of an individual. \code{FLQuant}.}
\item{fec}{Fecundity/maturity/per capita birth rate. \code{FLQuant}.}
\item{spwn}{Proportion of mortality before spawning/birth. \code{FLQuant}.}
\item{name}{Name of the object. \code{character}.} \item{desc}{Brief
description of the object. \code{character}.} \item{range}{Named numeric
vector containing the quant and year ranges, and the plusgroup
(\code{numeric}).} }
}
\examples{
# An FLBiol example dataset
  data(ple4.biol)
  summary(ple4.biol)

# Create an FLBiol object from an existing FLStock object
  data(ple4)
  flb1 <- as.FLBiol(ple4)
  plot(flb1)

  # Just extract same dimensions without contents
    flb2 <- FLBiol(stock.n(ple4))
}
\author{
The FLR Team
}
\seealso{
\link{as.FLBiol}, \link{as.FLSR}, \link[methods]{coerce},
\link[graphics]{plot}, \link{ssb} \link{catch.n,FLBiol-method}
}
\keyword{classes}

