% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLBiol.R
\docType{class}
\name{FLBiol}
\alias{FLBiol}
\alias{FLBiol,FLQuant-method}
\alias{FLBiol,missing-method}
\alias{FLBiol-class}
\alias{FLBiol-methods}
\alias{FLBiolcpp-class}
\alias{desc,FLBiol-method}
\alias{desc<-,FLBiol,character-method}
\alias{fec,FLBiol-method}
\alias{fec<-,FLBiol,FLQuant-method}
\alias{m,FLBiol-method}
\alias{m<-,FLBiol,FLQuant-method}
\alias{n,FLBiol-method}
\alias{n<-,FLBiol,FLQuant-method}
\alias{name,FLBiol-method}
\alias{name<-,FLBiol,character-method}
\alias{range,FLBiol-method}
\alias{range<-,FLBiol,numeric-method}
\alias{spwn,FLBiol-method}
\alias{spwn<-,FLBiol,FLQuant-method}
\alias{wt,FLBiol-method}
\alias{wt<-,FLBiol,FLQuant-method}
\title{Class FLBiol}
\usage{
FLBiol(object, ...)

\S4method{FLBiol}{FLQuant}(object, plusgroup = dims(object)$max, ...)
}
\arguments{
\item{object}{FLQuant object used for sizing}

\item{...}{Other objects to be assigned by name to the class slots}

\item{plusgroup}{Plusgroup age, to be stored in range}
}
\description{
A class for modelling age / length or biomass structured populations.
}
\details{
The \code{FLBiol} class is a representation of a biological fish population.
This includes information on abundances, natural mortality and fecundity.
}
\section{Slots}{
 \describe{
  \item{n}{Numbers in the population. \code{FLQuant}.}
  \item{m}{Mortality rate of the population. \code{FLQuant}.}
  \item{wt}{Mean weight of an individual. \code{FLQuant}.}
  \item{mat}{\code{predictModel}.}
  \item{fec}{\code{predictModel}.}
  \item{rec}{\code{predictModel}.}
  \item{spwn}{Proportion of time step at which spawning ocurrs. \code{FLQuant}.}
  \item{name}{Name of the object. \code{character}.}
  \item{desc}{Brief description of the object. \code{character}.}
  \item{range}{Named numeric vector describing the range of the object. \code{numeric}.} }
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Validity}{
 \describe{
    \item{Dimensions}{All FLQuant slots must have iters equal to 1 or 'n'.}
    \item{Iters}{The dimname for iter[1] should be '1'.}
    \item{Dimnames}{The name of the quant dimension must be the same for all FLQuant slots.}
}
}
\examples{

# An FLBiol example dataset
data(ple4.biol)

summary(ple4.biol)

}
\author{
The FLR Team
}
\seealso{
\link{as.FLBiol}, \link{as.FLSR}, \link[methods]{coerce}, \link[graphics]{plot}, \link{ssb} \link{catch.n,FLBiol-method}
}
\keyword{classes}

