\name{FLBiols}
\docType{class}
\alias{FLBiols}
\alias{FLBiols-class}
\title{Class FLBiols}
\description{
 A list of \code{FLBiol} objects.
}
\section{Slots}{\describe{
\item{.Data (\code{list}):}{Describe slot}
\item{names (\code{character}):}{Describe slot}
\item{desc (\code{character}):}{Describe slot}
\item{lock (\code{logical}):}{Describe slot}
}}
\section{Extends}{\describe{
FLlst
list
vector
}}
\section{Methods}{\describe{
All methods are inherited.
}}
\author{The FLR Team}
\seealso{ \link{FLlst}, \link[base]{list}, \link[base]{vector} }
\keyword{classes}
