% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLBiol.R
\docType{class}
\name{FLBiols}
\alias{FLBiols}
\alias{FLBiols-methods}
\alias{FLBiols}
\alias{FLBiols-class}
\alias{FLBiols,FLBiol-method}
\alias{FLBiols,missing-method}
\alias{FLBiols,list-method}
\title{Class FLBiols}
\usage{
FLBiols(object, ...)

\S4method{FLBiols}{FLBiol}(object, ...)

\S4method{FLBiols}{missing}(object, ...)

\S4method{FLBiols}{list}(object, ...)
}
\arguments{
\item{object}{unnamed object to be added to the list}

\item{...}{other named or unnamed objects}
}
\description{
A list of \code{FLBiol} objects.
}
\section{Slots}{
 \describe{
  \item{.Data}{Internal S4 data representation, of class \code{list}.}
  \item{desc}{As textual description of the object contents}
  \item{lock}{Can the object be extended/trimmed? \code{TRUE} or \code{FALSE}.}
  \item{names}{A character vector for the element names} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}

\seealso{
\code{\link{FLlst}}, \code{\link[base]{list}},
  \code{\link[base]{vector}}
}
\author{
The FLR Team
}
\keyword{classes}
