% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLBiol.R
\docType{class}
\name{FLBiols}
\alias{FLBiols}
\alias{FLBiols,ANY-method}
\alias{FLBiols,list-method}
\alias{FLBiols,missing-method}
\alias{FLBiols-class}
\alias{FLBiols-methods}
\title{Class FLBiols}
\usage{
FLBiols(object, ...)
}
\description{
A list of \code{FLBiol} objects.
}
\section{Slots}{
 \describe{
  \item{.Data}{Internal S4 data representation, of class \code{list}.}
  \item{desc}{As textual description of the object contents}
  \item{lock}{Can the object be extended/trimmed? \code{TRUE} or \code{FALSE}.}
  \item{names}{A character vector for the element names} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}
\author{
The FLR Team
}
\seealso{
\link{FLlst}, \link[base]{list}, \link[base]{vector}
}
\keyword{classes}

