% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{FLComp}
\alias{FLComp}
\alias{FLComp-class}
\title{Class FLComp}
\description{
A virtual class that forms the basis for most FLR classes composed of slots
of class \code{\linkS4class{FLQuant}}. No objects of this class can be
constructed.
}
\section{Validity}{
 \describe{
    \item{Dimensions}{All FLQuant slots must have iters equal to 1 or 'n'.}
    \item{Iters}{The dimname for iter[1] should be '1'.}
    \item{Dimnames}{The name of the quant dimension must be the same for all FLQuant slots.}
}
}

\section{Slots}{
 \describe{
   \item{name}{A character vector for the object name.}
   \item{desc}{A textual description of the object contents.}
   \item{range}{A named numeric vector with various values of quant and year ranges, plusgroup, fishing mortality ranges, etc.}
}
}
\author{
The FLR Team
}
\seealso{
\link[base]{[}, \link[base]{[<-}, \link[base]{as.data.frame},
\link{iter}, \link{propagate}, \link{qapply}, \link[base]{summary},
\link[base]{transform}, \link{trim}, \link{units,FLComp-method},
\link{units<-,FLComp,list-method}, \link[stats]{window}
}
\keyword{classes}

