\name{FLCore-accesors}
\alias{FLCore-accesors}
\alias{name}
\alias{name-methods}
\alias{name<-}
\alias{"name<-"-methods}
\alias{desc}
\alias{desc-methods}
\alias{desc<-}
\alias{"desc<-"-methods}
\alias{distribution}
\alias{distribution-methods}
\alias{distribution<-}
\alias{"distribution<-"-methods}
\alias{index}
\alias{index-methods}
\alias{index<-}
\alias{"index<-"-methods}
\alias{index.var}
\alias{index.var-methods}
\alias{index.var<-}
\alias{"index.var<-"-methods}
\alias{n}
\alias{n-methods}
\alias{n<-}
\alias{"n<-"-methods}
\alias{m}
\alias{m-methods}
\alias{m<-}
\alias{"m<-"-methods}
\alias{wt}
\alias{wt-methods}
\alias{wt<-}
\alias{"wt<-"-methods}
\alias{fec}
\alias{fec-methods}
\alias{fec<-}
\alias{"fec<-"-methods}
\alias{spwn}
\alias{spwn-methods}
\alias{spwn<-}
\alias{"spwn<-"-methods}
\alias{model}
\alias{model-methods}
\alias{model<-}
\alias{"model<-"-methods}
\alias{ssb-methods}
\alias{ssb<-}
\alias{"ssb<-"-methods}
\alias{rec}
\alias{rec-methods}
\alias{rec<-}
\alias{"rec<-"-methods}
\alias{rechat}
\alias{rechat-methods}
\alias{rechat<-}
\alias{"rechat<-"-methods}
\alias{residuals}
\alias{residuals-methods}
\alias{residuals<-}
\alias{"residuals<-"-methods}
\alias{params}
\alias{params-methods}
\alias{params<-}
\alias{"params<-"-methods}
\alias{se}
\alias{se-methods}
\alias{se<-}
\alias{"se<-"-methods}
\alias{covar}
\alias{covar-methods}
\alias{covar<-}
\alias{"covar<-"-methods}
\alias{varacorr}
\alias{varacorr-methods}
\alias{varacorr<-}
\alias{"varacorr<-"-methods}
\alias{aic}
\alias{aic-methods}
\alias{aic<-}
\alias{"aic<-"-methods}
\alias{mcmc}
\alias{mcmc-methods}
\alias{mcmc<-}
\alias{"mcmc<-"-methods}
\alias{catch.n}
\alias{catch.n-methods}
\alias{catch.wt}
\alias{catch.wt-methods}
\alias{effort}
\alias{effort<-}
\alias{effort-methods}
\alias{"effort<-"-methods}
\alias{sel.pattern}
\alias{sel.pattern-methods}
\alias{type}
\alias{type-methods}
\alias{type<-}
\alias{"type<-"-methods}
\alias{index.q}
\alias{index.q-methods}
\alias{catch}
\alias{catch-methods}
\alias{catch.n}
\alias{catch.n-methods}
\alias{catch.wt}
\alias{catch.wt-methods}
\alias{catch<-}
\alias{"catch<-"-methods}
\alias{catch.n<-}
\alias{"catch.n<-"-methods}
\alias{catch.wt<-}
\alias{"catch.wt<-"-methods}
\alias{discards}
\alias{discards-methods}
\alias{discards<-}
\alias{"discards<-"-methods}
\alias{discards.n}
\alias{discards.n-methods}
\alias{discards.n<-}
\alias{"discards.n<-"-methods}
\alias{discards.wt}
\alias{discards.wt-methods}
\alias{discards.wt<-}
\alias{"discards.wt<-"-methods}
\alias{landings}
\alias{landings-methods}
\alias{landings.n}
\alias{landings.n-methods}
\alias{landings.wt}
\alias{landings.wt-methods}
\alias{landings<-}
\alias{"landings<-"-methods}
\alias{landings.n<-}
\alias{"landings.n<-"-methods}
\alias{landings.wt<-}
\alias{"landings.wt<-"-methods}
\alias{mat}
\alias{mat-methods}
\alias{mat<-}
\alias{"mat<-"-methods}
\alias{m.spwn}
\alias{m.spwn-methods}
\alias{m.spwn<-}
\alias{"m.spwn<-"-methods}
\alias{stock}
\alias{stock-methods}
\alias{stock<-}
\alias{"stock<-"-methods}
\alias{stock.n}
\alias{stock.n-methods}
\alias{stock.n<-}
\alias{"stock.n<-"-methods}
\alias{stock.wt}
\alias{stock.wt-methods}
\alias{stock.wt<-}
\alias{"stock.wt<-"-methods}
\alias{harvest}
\alias{harvest-methods}
\alias{harvest<-}
\alias{"harvest<-"-methods}
\alias{harvest.spwn}
\alias{harvest.spwn-methods}
\alias{harvest.spwn<-}
\alias{"harvest.spwn<-"-methods}
\alias{catch.sel}
\alias{catch.sel-methods}
\alias{catch.sel<-}
\alias{"catch.sel<-"-methods}
\alias{catchability}
\alias{catchability-methods}
\alias{catchability<-}
\alias{"catchability<-"-methods}
\alias{discards.sel}
\alias{discards.sel-methods}
\alias{discards.sel<-}
\alias{"discards.sel<-"-methods}
\alias{gear}
\alias{gear-methods}
\alias{gear<-}
\alias{"gear<-"-methods}
\alias{landings.sel}
\alias{landings.sel-methods}
\alias{landings.sel<-}
\alias{"landings.sel<-"-methods}
\alias{price}
\alias{price-methods}
\alias{price<-}
\alias{"price<-"-methods}



\title{Accesor and replacement methods for slots of complex objects}
\description{
}
\usage{
	name(object, ...)
	name(object) <- value
}
\arguments{
	\item{object}{The object whose slot is to be returned.}
	\item{value}{The new value to be placed in that slot.}
	\item{...}{Extra arguments.}
}
\details{
}
\value{
	Accesor functions return the content of the selected slot, whic could be of various classes.
}
\seealso{
	\code{\link{FLStock-class}}, \code{\link{FLBiol-class}}, \code{\link{FLIndex-class}},
	\code{\link{FLFleet-class}},\code{\link{FLSR-class}}
}
\examples{
}
\author{FLR Team}
\keyword{methods}
