\name{FLCore-accesors}
\docType{methods}
% name
\alias{name}
\alias{name-methods}
\alias{name,FLStock-method}
\alias{name,FLCatch-method}
\alias{name,FLMetier-method}
\alias{name,FLBiol-method}
\alias{name,FLFleet-method}
\alias{name,FLIndex-method}
\alias{name,FLModel-method}
% name<- 
\alias{name<-}
\alias{name<--methods}
\alias{name<-,FLStock,character-method}
\alias{name<-,FLBiol,character-method}
\alias{name<-,FLFleet,character-method}
\alias{name<-,FLCatch,character-method}
\alias{name<-,FLMetier,character-method}
\alias{name<-,FLIndex,character-method}
\alias{name<-,FLModel,character-method}
% desc
\alias{desc}
\alias{desc-methods}
\alias{desc,FLStock-method}
\alias{desc,FLCatch-method}
\alias{desc,FLMetier-method}
\alias{desc,FLBiol-method}
\alias{desc,FLFleet-method}
\alias{desc,FLIndex-method}
\alias{desc,FLModel-method}
% desc<- 
\alias{desc<-}
\alias{desc<--methods}
\alias{desc<-,FLStock,character-method}
\alias{desc<-,FLBiol,character-method}
\alias{desc<-,FLFleet,character-method}
\alias{desc<-,FLCatch,character-method}
\alias{desc<-,FLMetier,character-method}
\alias{desc<-,FLIndex,character-method}
\alias{desc<-,FLModel,character-method}
% spp
\alias{spp}
\alias{spp-methods}
\alias{spp,FLCatch-method}
\alias{spp,FLCatches-method}
\alias{spp,FLMetier-method}
\alias{spp,FLMetiers-method}
\alias{spp,FLFleet-method}
\alias{spp,FLFleets-method}
% spp<-
\alias{spp<--methods}
\alias{spp<-,FLCatch,character-method}
% catch
\alias{catch}
\alias{catch-methods}
\alias{catch,FLStock-method}
% catch<-
\alias{catch<-}
\alias{catch<--methods}
\alias{catch<-,FLStock,FLQuant-method}
\alias{catch<-,FLStock,numeric-method}
\alias{catch<-,FLStock,FLQuants-method}
%catch.n
\alias{catch.n}
\alias{catch.n-methods}
\alias{catch.n,FLStock-method}
\alias{catch.n,FLIndex-method}
% catch.n<-
\alias{catch.n<-}
\alias{catch.n<--methods}
\alias{catch.n<-,FLStock,FLQuant-method}
\alias{catch.n<-,FLStock,numeric-method}
\alias{catch.n<-,FLIndex,FLQuant-method}
\alias{catch.n<-,FLIndex,numeric-method}
%catch.wt
\alias{catch.wt}
\alias{catch.wt-methods}
\alias{catch.wt,FLStock-method}
\alias{catch.wt,FLIndex-method}
% catch.wt<-
\alias{catch.wt<-}
\alias{catch.wt<--methods}
\alias{catch.wt<-,FLStock,FLQuant-method}
\alias{catch.wt<-,FLStock,numeric-method}
\alias{catch.wt<-,FLIndex,FLQuant-method}
\alias{catch.wt<-,FLIndex,numeric-method}
% catch.q
\alias{catch.q}
\alias{catch.q-methods}
% catch.q<-
\alias{catch.q<-}
\alias{catch.q<--methods}
% catch.sel
\alias{catch.sel}
\alias{catch.sel-methods}
% catch.sel<-
\alias{catch.sel<-}
\alias{catch.sel<--methods}
% catches
\alias{catches}
\alias{catches-methods}
% catches<-
\alias{catches<-}
\alias{catches<--methods}
% capacity
\alias{capacity}
\alias{capacity-methods}
\alias{capacity,FLFleet-method}
% capacity<-
\alias{capacity<-}
\alias{capacity<--methods}
\alias{capacity<-,FLFleet,FLQuant-method}
\alias{capacity<-,FLFleet,numeric-method}
% crewshare
\alias{crewshare}
\alias{crewshare-methods}
\alias{crewshare,FLFleet-method}
% crewshare<-
\alias{crewshare<-}
\alias{crewshare<--methods}
\alias{crewshare<-,FLFleet,FLQuant-method}
\alias{crewshare<-,FLFleet,numeric-method}
% covar
\alias{covar}
\alias{covar-methods}
\alias{covar,FLSR-method}
% covar<-
\alias{covar<-}
\alias{covar<--methods}
\alias{covar<-,FLSR-method}
\alias{covar<-,FLSR,FLQuants-method}
% details
\alias{details}
\alias{details-methods}
\alias{details,FLModel-method}
% details<-
\alias{details<-}
\alias{details<--methods}
\alias{details<-,FLModel,list-method}
% discards
\alias{discards}
\alias{discards-methods}
\alias{discards,FLStock-method}
% discards<-
\alias{discards<-}
\alias{discards<--methods}
\alias{discards<-,FLStock,FLQuant-method}
\alias{discards<-,FLStock,numeric-method}
\alias{discards<-,FLStock,FLQuants-method}
%discards.n
\alias{discards.n}
\alias{discards.n-methods}
\alias{discards.n,FLStock-method}
\alias{discards.n,FLIndex-method}
% discards.n<-
\alias{discards.n<-}
\alias{discards.n<--methods}
\alias{discards.n<-,FLStock,FLQuant-method}
\alias{discards.n<-,FLStock,numeric-method}
\alias{discards.n<-,FLIndex,FLQuant-method}
% discards.wt
\alias{discards.wt}
\alias{discards.wt-methods}
\alias{discards.wt,FLStock-method}
\alias{discards.wt,FLIndex-method}
% discards.wt<-
\alias{discards.wt<-}
\alias{discards.wt<--methods}
\alias{discards.wt<-,FLStock,FLQuant-method}
\alias{discards.wt<-,FLStock,numeric-method}
\alias{discards.wt<-,FLIndex,FLQuant-method}
\alias{discards.wt<-,FLIndex,numeric-method}
% discards.q
\alias{discards.q}
\alias{discards.q-methods}
% discards.q<-
\alias{discards.q<-}
\alias{discards.q<--methods}
% discards.sel
\alias{discards.sel}
\alias{discards.sel-methods}
% discards.sel<-
\alias{discards.sel<-}
\alias{discards.sel<--methods}
% effort
\alias{effort}
\alias{effort-methods}
\alias{effort,FLIndex,missing-method}
\alias{effort,FLFleet,missing-method}
\alias{effort,FLFleet,character-method}
% effort<-
\alias{effort<-}
\alias{effort<--methods}
\alias{effort<-,FLIndex,FLQuant-method}
\alias{effort<-,FLIndex,numeric-method}
\alias{effort<-,FLFleet,FLQuant-method}
\alias{effort<-,FLFleet,numeric-method}
% effshare
\alias{effshare}
\alias{effshare-methods}
\alias{effshare,FLMetier-method}
\alias{effshare,FLMetiers-method}
\alias{effshare,FLFleet-method}
% effshare<-
\alias{effshare<-}
\alias{effshare<--methods}
\alias{effshare<-,FLMetier,FLQuant-method}
\alias{effshare<-,FLMetier,numeric-method}
% fcost
\alias{fcost}
\alias{fcost-methods}
\alias{fcost,FLFleet-method}
% fcost<-
\alias{fcost<-}
\alias{fcost<--methods}
\alias{fcost<-,FLFleet,FLQuant-method}
\alias{fcost<-,FLFleet,numeric-method}
% fec
\alias{fec}
\alias{fec-methods}
\alias{fec,FLBiol-method}
% fec<-
\alias{fec<-}
\alias{fec<--methods}
\alias{fec<-,FLBiol,FLQuant-method}
\alias{fec<-,FLBiol,numeric-method}
% fitted
\alias{fitted,ANY-method}
\alias{fitted,FLModel-method}
% fitted<-
\alias{fitted<-}
\alias{fitted<--methods}
\alias{fitted<-,FLModel,FLArray-method}
\alias{fitted<-,FLModel,numeric-method}
% gear
\alias{gear}
\alias{gear-methods}
\alias{gear,ANY-method}
\alias{gear,FLMetier-method}
% gear<-
\alias{gear<-}
\alias{gear<--methods}
\alias{gear<-,FLMetier,character-method}
% gr
\alias{gr}
\alias{gr-methods}
\alias{gr,FLModel-method}
% gr<-
\alias{gr<-}
\alias{gr<--methods}
\alias{gr<-,FLModel,function-method}
% harvest
\alias{harvest}
\alias{harvest-methods}
\alias{harvest,FLStock,missing-method}
% harvest<-
\alias{harvest<-}
\alias{harvest<--methods}
\alias{harvest<-,FLStock,character-method}
\alias{harvest<-,FLStock,FLQuant-method}
\alias{harvest<-,FLStock,numeric-method}
% harvest.spwn
\alias{harvest.spwn}
\alias{harvest.spwn-methods}
\alias{harvest.spwn,FLStock-method}
% harvest.spwn<-
\alias{harvest.spwn<-}
\alias{harvest.spwn<--methods}
\alias{harvest.spwn<-,FLStock,FLQuant-method}
\alias{harvest.spwn<-,FLStock,numeric-method}
% index
\alias{index}
\alias{index-methods}
\alias{index,FLIndex-method}
% index<-
\alias{index<-}
\alias{index<--methods}
\alias{index<-,FLIndex,FLQuant-method}
\alias{index<-,FLIndex,numeric-method}
% index.q
\alias{index.q}
\alias{index.q-methods}
\alias{index.q,FLIndex-method}
% index.q<-
\alias{index.q<-}
\alias{index.q<--methods}
\alias{index.q<-,FLIndex,FLQuant-method}
\alias{index.q<-,FLIndex,numeric-method}
% index.var
\alias{index.var}
\alias{index.var-methods}
\alias{index.var,FLIndex-method}
% index.var<-
\alias{index.var<-}
\alias{index.var<--methods}
\alias{index.var<-,FLIndex,FLQuant-method}
\alias{index.var<-,FLIndex,numeric-method}
% initial
\alias{initial}
\alias{initial-methods}
\alias{initial,FLModel-method}
% initial<-
\alias{initial<-}
\alias{initial<--methods}
\alias{initial<-,FLModel,function-method}
% landings
\alias{landings}
\alias{landings-methods}
\alias{landings,FLStock-method}
% landings<-
\alias{landings<-}
\alias{landings<--methods}
\alias{landings<-,FLStock,FLQuant-method}
\alias{landings<-,FLStock,numeric-method}
\alias{landings<-,FLStock,FLQuants-method}
%landings.n
\alias{landings.n}
\alias{landings.n-methods}
\alias{landings.n,FLStock-method}
\alias{landings.n,FLIndex-method}
% landings.n<-
\alias{landings.n<-}
\alias{landings.n<--methods}
\alias{landings.n<-,FLStock,FLQuant-method}
\alias{landings.n<-,FLStock,numeric-method}
\alias{landings.n<-,FLIndex,FLQuant-method}
%landings.wt
\alias{landings.wt}
\alias{landings.wt-methods}
\alias{landings.wt,FLStock-method}
\alias{landings.wt,FLIndex-method}
% landings.wt<-
\alias{landings.wt<-}
\alias{landings.wt<--methods}
\alias{landings.wt<-,FLStock,FLQuant-method}
\alias{landings.wt<-,FLStock,numeric-method}
\alias{landings.wt<-,FLIndex,FLQuant-method}
% landings.q
\alias{landings.q}
\alias{landings.q-methods}
% landings.q<-
\alias{landings.q<-}
\alias{landings.q<--methods}
% landings.sel
\alias{landings.sel}
\alias{landings.sel-methods}
% landings.sel<-
\alias{landings.sel<-}
\alias{landings.sel<--methods}
% logl
\alias{logl}
\alias{logl-methods}
\alias{logl,FLModel-method}
% logl<-
\alias{logl<-}
\alias{logl<--methods}
\alias{logl<-,FLModel,function-method}
% logLik
\alias{logLik,ANY-method}
\alias{logLik,FLModel-method}
% logLik
\alias{logLik<-}
\alias{logLik<--methods}
\alias{logLik<-,FLModel,logLik-method}
\alias{logLik<-,FLModel,numeric-method}
% m
\alias{m}
\alias{m-methods}
\alias{m,FLStock-method}
\alias{m,FLBiol-method}
% m<-
\alias{m<-}
\alias{m<--methods}
\alias{m<-,FLStock,FLQuant-method}
\alias{m<-,FLStock,numeric-method}
\alias{m<-,FLBiol,FLQuant-method}
\alias{m<-,FLBiol,numeric-method}
% m.spwn
\alias{m.spwn}
\alias{m.spwn-methods}
\alias{m.spwn,FLStock-method}
% m.spwn<-
\alias{m.spwn<-}
\alias{m.spwn<--methods}
\alias{m.spwn<-,FLStock,FLQuant-method}
\alias{m.spwn<-,FLStock,numeric-method}
% mat
\alias{mat}
\alias{mat-methods}
\alias{mat,FLStock-method}
% mat<-
\alias{mat<-}
\alias{mat<--methods}
\alias{mat<-,FLStock,FLQuant-method}
\alias{mat<-,FLStock,numeric-method}
% metiers
\alias{metiers}
\alias{metiers-methods}
\alias{metiers,FLFleet-method}
% metiers<-
\alias{metiers<-}
\alias{metiers<--methods}
\alias{metiers<-,FLFleet,FLMetiers-method}
% model
\alias{model}
\alias{model-methods}
\alias{model,FLModel-method}
% model<-
\alias{model<-}
\alias{model<--methods}
\alias{model<-,FLModel,formula-method}
\alias{model<-,FLModel,character-method}
\alias{model<-,FLModel,function-method}
\alias{model<-,FLModel,list-method}
% n
\alias{n}
\alias{n-methods}
\alias{n,FLBiol-method}
% n<-
\alias{n<-}
\alias{n<--methods}
\alias{n<-,FLBiol,FLQuant-method}
\alias{n<-,FLBiol,numeric-method}
% params
\alias{params}
\alias{params-methods}
\alias{params,FLModel-method}
% params<-
\alias{params<-}
\alias{params<--methods}
\alias{params<-,FLModel,FLPar-method}
% price
\alias{price}
\alias{price-methods}
% price<-
\alias{price<-}
\alias{price<--methods}
% rec
\alias{rec}
\alias{rec-methods}
\alias{rec,FLSR-method}
\alias{rec,FLBiol-method}
\alias{rec,FLStock-method}
% rec<-
\alias{rec<-}
\alias{rec<--methods}
\alias{rec<-,FLSR,FLQuant-method}
\alias{rec<-,FLSR,numeric-method}
% residuals
\alias{residuals,ANY-method}
\alias{residuals,FLModel-method}
% residuals<-
\alias{residuals<-}
\alias{residuals<--methods}
\alias{residuals<-,FLModel,FLArray-method}
\alias{residuals<-,FLModel,numeric-method}
% sel.pattern
\alias{sel.pattern}
\alias{sel.pattern-methods}
\alias{sel.pattern,FLIndex-method}
% sel.pattern<-
\alias{sel.pattern<-}
\alias{sel.pattern<--methods}
\alias{sel.pattern<-,FLIndex,FLQuant-method}
\alias{sel.pattern<-,FLIndex,numeric-method}
%spwn
\alias{spwn}
\alias{spwn-methods}
\alias{spwn,FLBiol-method}
% spwn<-
\alias{spwn<-}
\alias{spwn<--methods}
\alias{spwn<-,FLBiol,FLQuant-method}
\alias{spwn<-,FLBiol,numeric-method}
% ssb
\alias{ssb,FLSR-method}
% ssb<-
\alias{ssb<-}
\alias{ssb<--methods}
\alias{ssb<-,FLSR,FLQuant-method}
\alias{ssb<-,FLSR,numeric-method}
% stock
\alias{stock}
\alias{stock-methods}
\alias{stock,FLStock-method}
% stock<-
\alias{stock<-}
\alias{stock<--methods}
\alias{stock<-,FLStock,FLQuant-method}
\alias{stock<-,FLStock,numeric-method}
% stock.n
\alias{stock.n}
\alias{stock.n-methods}
\alias{stock.n,FLStock-method}
% stock.n<-
\alias{stock.n<-}
\alias{stock.n<--methods}
\alias{stock.n<-,FLStock,FLQuant-method}
\alias{stock.n<-,FLStock,numeric-method}
% stock.wt
\alias{stock.wt}
\alias{stock.wt-methods}
\alias{stock.wt,FLStock-method}
% stock.wt<-
\alias{stock.wt<-}
\alias{stock.wt<--methods}
\alias{stock.wt<-,FLStock,FLQuant-method}
\alias{stock.wt<-,FLStock,numeric-method}
% type
\alias{type}
\alias{type-methods}
\alias{type,FLIndex-method}
% type<-
\alias{type<-}
\alias{type<--methods}
\alias{type<-,FLIndex,character-method}
% vcost
\alias{vcost}
\alias{vcost-methods}
\alias{vcost,FLMetier-method}
\alias{vcost,FLMetiers-method}
\alias{vcost,FLFleet-method}
% vcost <-
\alias{vcost<-}
\alias{vcost<--methods}
\alias{vcost<-,FLMetier,FLQuant-method}
\alias{vcost<-,FLMetier,numeric-method}
%vcov
\alias{vcov,ANY-method}
\alias{vcov,FLModel-method}
% vcov<-
\alias{vcov<-}
\alias{vcov<--methods}
\alias{vcov<-,FLModel,array-method}
% wt
\alias{wt}
\alias{wt-methods}
\alias{wt,FLBiol-method}
% wt<-
\alias{wt<-}
\alias{wt<--methods}
\alias{wt<-,FLBiol,FLQuant-method}
\alias{wt<-,FLBiol,numeric-method}

\title{Accesor and replacement methods for complex S4 classes}
\description{
	Slots in all S4 classes defined in FLCore can be accessed and modified using
	these accessor and replacement methods. Methods are named after the slot they
	apply to.

	These are generic methods that can be extended for other classes.

	Accessors for objects of class \link{FLCatch} work slightly differently, as
	FLQuant slots in the embeded classes. See \link{FLCatch-accesors}.

  The replacement function \code{model<-} can simultaneously modify a number of slots
  in objects of class \code{\link{FLModel}} if a list is passed to it with named elements
  corresponding to slots in the class. This is used to modify model-related slots in
  a single operation: \code{model}, \code{logl} and \code{initial}. A function, or a
  character with the name of a function returning such a list can also be passed.
}
\section{Generic function}{
capacity(object) 
capacity<-(object,value) 
catch(object) 
catch<-(object,value) 
catch.n(object) 
catch.n<-(object,value) 
catch.wt(object) 
catch.wt<-(object,value) 
catch.q(object) 
catch.q<-(object,value) 
catch.sel(object) 
crewshare<-(object,value) 
crewshare(object) 
details(object) 
details<-(object,value) 
discards(object) 
discards<-(object,value) 
discards.n(object) 
discards.n<-(object,value) 
discards.wt(object) 
discards.wt<-(object,value) 
discards.q(object) 
discards.q<-(object,value) 
discards.sel(object) 
effort(object) 
effort<-(object,value) 
effshare(object) 
effshare<-(object,value) 
fcost(object) 
fcost<-(object,value) 
fec(object) 
fec<-(object,value) 
fitted(object) 
fitted<-(object,value) 
harvest(object) 
harvest<-(object,value) 
harvest.spwn(object) 
harvest.spwn<-(object, value) 
index(object) 
index<-(object,value) 
index.q(object) 
index.q<-(object,value) 
index.var(object) 
index.var<-(object,value) 
landings(object) 
landings<-(object,value) 
landings.n(object) 
landings.n<-(object,value) 
landings.wt(object) 
landings.wt<-(object,value) 
landings.q(object) 
landings.q<-(object,value) 
landings.sel(object) 
logl(object) 
logl<-(object,value) 
logLik(object) 
logLik<-(object,value) 
model(object) 
model<-(object,value) 
m(object) 
m<-(object,value) 
m.spwn(object) 
m.spwn<-(object,value) 
n(object) 
n<-(object,value) 
params(object) 
params<-(object,value) 
price(object) 
price<-(object,value) 
residuals(object) 
residuals<-(object,value) 
sel.pattern(object) 
sel.pattern<-(object,value) 
spwn(object) 
spwn<-(object,value) 
stock(object) 
stock<-(object,value) 
stock.n(object) 
stock.n<-(object,value) 
stock.wt(object) 
stock.wt<-(object,value) 
wt(object) 
wt<-(object,value) 
vcost(object) 
vcost<-(object,value) 
vcov(object) 
vcov<-(object,value) 
}
\section{Methods}{\describe{
\item{object}{An FLR complex object for which we want to extract or modify a slot.}
\item{value}{An object to be placed in a given slot. This object needs to be of the
	corresponding class.}
}}
\author{The FLR Team}
\seealso{\link{FLComp}, \link{FLCatch-accesors} }
\examples{
data(ple4)
summary(catch(ple4))
catch(ple4) <- landings(ple4)

# model<- modifies one or more slots
flm <- FLSR(rec~ssb*a)
model(flm)
model(flm) <- rec~ssb^a
model(flm) <- ricker()
model(flm) <- list(model=rec~ssb*a+ssb*b, initial=function() return(c(a=1, b=1)))
}
\keyword{methods}
