% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCore-package.R
\docType{package}
\name{FLCore-package}
\alias{FLCore}
\alias{FLCore-package}
\title{Core package of FLR, fisheries modelling in R.}
\description{
FLCore contains the core classes and methods for FLR, a framework for
fisheries modelling and management strategy simulation in R. Developed by a
team of fisheries scientists in various countries. More information can be
found at http://flr-project.org/, including a development mailing list.
}
\author{
FLR Team \email{flrteam@flr-project.org},
Iago Mosqueira \email{iago.mosqueira@jrc.ec.europa.eu},
Laurence T. Kell \email{laurie.kell@iccat.int},
Philippe Grosjean \email{Philippe.Grosjean@umons.ac.be}

Maintainer: Iago Mosqueira \email{iago.mosqueira@jrc.ec.europa.eu},
}
\references{
Website at \url{http://flr-project.org/}.

Kell L.T., Mosqueira I., Grosjean P., Fromentin J-M., Garcia D., Hillary R.,
Jardim E., Pastoors M., Poos J.J., Scott F. & Scott R.D. 2007.  FLR: an
open-source framework for the evaluation and development of management
strategies. ICES J. of Mar. Sci. 20: 289-290. doi: 10.1093/icesjms/fsm012
}
\keyword{package}

