\name{FLIndex}
\docType{methods}
\alias{FLIndex}
\alias{FLIndex-methods}
\alias{FLIndex,missing-method}
\alias{FLIndex,FLQuant-method}
\title{Create FLIndex objects}
\description{
  Constructor method for the FLIndex class
}
\section{Arguments}{
  \describe{
  \item{object}{An initial FLQuant used for sizing.}
  \item{plusgroup}{The age/length to be considered as plusgroup.}
  }
}
\section{Generic function}{
 FLIndex(object, ...)
}
\section{Methods}{\describe{
\item{ signature(object=missing) :}{Describe method}
\item{ signature(object=FLQuant) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLIndex-class} }
\examples{
}
\keyword{methods}
