% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classesComp.R
\docType{class}
\name{FLIndexBiomass}
\alias{FLIndexBiomass}
\alias{FLIndexBiomass,FLQuant-method}
\alias{FLIndexBiomass,missing-method}
\alias{FLIndexBiomass-methods}
\alias{FLIndexBiomassBiomass}
\alias{FLQuant}
\alias{catch.n,FLIndexBiomass-method}
\alias{catch.n<-,FLIndexBiomass,FLQuant-method}
\alias{catch.wt,FLIndexBiomass-method}
\alias{catch.wt<-,FLIndexBiomass,FLQuant-method}
\alias{desc,FLIndexBiomass-method}
\alias{desc<-,FLIndexBiomass,character-method}
\alias{distribution,FLIndexBiomass-method}
\alias{distribution<-,FLIndexBiomass,character-method}
\alias{effort,FLIndexBiomass-method}
\alias{effort<-,FLIndexBiomass,FLQuant-method}
\alias{index,FLIndexBiomass-method}
\alias{index.q,FLIndexBiomass-method}
\alias{index.q<-,FLIndexBiomass,FLQuant-method}
\alias{index.var,FLIndexBiomass-method}
\alias{index.var<-,FLIndexBiomass,FLQuant-method}
\alias{index<-,FLIndexBiomass,FLQuant-method}
\alias{name,FLIndexBiomass-method}
\alias{name<-,FLIndexBiomass,character-method}
\alias{range,FLIndexBiomass-method}
\alias{range<-,FLIndexBiomass,numeric-method}
\alias{sel.pattern,FLIndexBiomass-method}
\alias{sel.pattern<-,FLIndexBiomass,FLQuant-method}
\alias{type,FLIndexBiomass-method}
\alias{type<-,FLIndexBiomass,character-method}
\title{Class FLIndexBiomass}
\description{
A class for modelling biomass indices.
}
\details{
The \code{FLIndexBiomass} object holds data and parameters related to
biomass indices.
}
\section{Slots}{
 \describe{ \item{distribution}{Statistical distribution of
the index values (\code{character}).} \item{index}{Index values
(\code{FLQuant}).} \item{index.var}{Variance of the index (\code{FLQuant}).}
\item{catch.n}{Catch numbers used to create the index (\code{FLQuant}).}
\item{catch.wt}{Catch weight of the index (\code{FLQuant}).}
\item{effort}{Effort used to create the index (\code{FLQuant}).}
\item{sel.pattern}{Selection pattern for the index (\code{FLQuant}).}
\item{index.q}{Catchability of the index (\code{FLQuant}).} \item{name}{Name
of the stock (\code{character}).} \item{desc}{General description of the
object (\code{character}).} \item{range}{Named numeric vector containing the
quant and year ranges, the plusgroup, and the period of the year, expressed
as proportions of a year, that corresponds to the index (\code{numeric}).} }
}
\examples{
# Create an FLIndexBiomass object from scratch...
  idx <- FLIndexBiomass(index=FLQuant(1:10, quant='age'))

# ...from an existing FLStock object
  data(ple4)
  ida <- FLIndexBiomass(index=(catch(ple4)/fbar(ple4))*
    exp(rnorm(1,catch(ple4)*0,0.3)), catch.n=catch.n(ple4),
    catch.wt=catch.wt(ple4))
}
\author{
The FLR Team
}
\seealso{
\link{computeCatch}, \link{dims}, \link{iter},
\link[graphics]{plot}, \link{propagate}, \link[base]{summary},
\link[base]{transform}, \link{trim}, \link[stats]{window}, \link{FLComp}
}
\keyword{classes}

