\name{FLIndices}
\docType{methods}
\alias{FLIndices-methods}
\alias{FLIndices,ANY-method}
 \alias{FLIndices,missing-method}
 \alias{FLIndices,list-method}
\title{Method FLIndices}
\description{
 My method is better then yours !
}
\section{Generic function}{
 FLIndices(object) 
}
\section{Methods}{\describe{
\item{ signature(object=ANY) :}{Describe method}
\item{ signature(object=missing) :}{Describe method}
\item{ signature(object=list) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
