\name{FLlst}
\docType{methods}
\alias{FLlst-methods}
\alias{FLlst,ANY-method}
 \alias{FLlst,missing-method}
 \alias{FLlst,list-method}
\title{Method FLlst}
\description{
 My method is better then yours !
}
\section{Generic function}{
 FLlst(object) 
}
\section{Methods}{\describe{
\item{ signature(object=ANY) :}{Describe method}
\item{ signature(object=missing) :}{Describe method}
\item{ signature(object=list) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
