\name{FLModel}
\docType{methods}
\alias{FLModel-methods}
\alias{FLModel,formula-method}
\alias{FLModel,missing-method}
\alias{FLModel,character-method}
\alias{FLModel,function-method}
\title{Method FLModel}
\description{
 My method is better then yours !
}
\section{Generic function}{
 FLModel(object) 
}
\section{Methods}{\describe{
\item{ signature(object=formula) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
