% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLModelSim.R
\docType{class}
\name{FLModelSim}
\alias{FLModelSim}
\alias{FLModelSim-methods}
\alias{FLModelSim}
\alias{FLModelSim-class}
\alias{FLModelSim-methods}
\alias{FLModelSim,formula-method}
\alias{FLModelSim,missing-method}
\alias{FLModelSim,character-method}
\alias{FLModelSim,function-method}
\alias{FLModelSim,missing-method}
\title{Class FLModelSim}
\usage{
FLModelSim(object, ...)

\S4method{FLModelSim}{missing}(object, ...)
}
\description{
A virtual class for statistical simulation models
}
\details{
The \code{FLModelSim} class provides a virtual class that developers of various
statistical models can use to implement classes that allow those models to
be tested, fitted and presented.

Slots in this class attempt to map all the usual outputs for a modelling
exercise, together with the standard inputs. Input data are stored in slots
created by a specified class that is based on \code{FLModelSim}. See for example
\code{\linkS4class{FLSR}} for a class used for stock-recruitment models.

Various fitting algorithms, similar to those present in the basic R packages,
are currently available for \code{FLModelSim}, including \code{\link{fmle}},
\code{\link{nls-FLCore}} and \code{\link[stats]{glm}}.
}
\section{Slots}{
 \describe{
\item{params}{Estimated parameter values. \code{FLPar}.}
\item{distr}{\code{character}}
\item{vcov}{\code{array}}
\item{model}{\code{formula}}}
}

\seealso{
\link[stats]{AIC}, \link[stats]{BIC}, \link{fmle},
\link[stats]{nls}
}
\author{
The FLR Team
}
\keyword{classes}
