\name{FLPar}
\docType{methods}
\alias{FLPar}
\alias{FLPar-methods}
\alias{FLPar,array-method}
\alias{FLPar,missing-method}
\alias{FLPar,vector-method}
\alias{FLPar,FLPar-method}
\title{Method FLPar}
\description{
 Creates an object of class \code{FLPar}.
}
\section{Generic function}{
 FLPar(object) 
}
\section{Methods}{\describe{
\item{signature(object=array) :}{Describe method}
\item{signature(object=missing) :}{Describe method}
\item{signature(object=vector) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
