% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesArr.R, R/jackknife.R
\docType{class}
\name{FLParJK}
\alias{FLParJK}
\alias{FLParJK-class}
\alias{FLParJK,ANY-method}
\alias{orig,FLParJK-method}
\title{Class FLParJK}
\usage{
\S4method{FLParJK}{ANY}(object, orig)

\S4method{orig}{FLParJK}(object)
}
\description{
A class for storing parameters of a jackknifed model fit.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{Jackknifed object, \code{FLPar}.}

\item{\code{units}}{units of measurement, \code{character}.}

\item{\code{orig}}{original object being jackknifed, \code{FLPar}.}
}}

\section{Validity}{

You can inspect the class validity function by using
\code{getValidity(getClassDef('FLParJK'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

Objects of this class are commonly created by calling the \code{\link[=jackknife]{jackknife()}} method
A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity.
}

\seealso{
\code{\link{FLPar}}
}
\author{
The FLR Team
}
\keyword{classes}
