% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/classesLst.R
\docType{class}
\name{FLPars}
\alias{FLPars}
\alias{FLPars,ANY-method}
\alias{FLPars,FLPars-method}
\alias{FLPars,list-method}
\alias{FLPars,missing-method}
\alias{FLPars-class}
\alias{FLPars-methods}
\title{Class FLPars}
\usage{
FLPars(object, ...)

\S4method{FLPars}{ANY}(object, ...)

\S4method{FLPars}{missing}(object, ...)

\S4method{FLPars}{list}(object)

\S4method{FLPars}{FLPars}(object)
}
\description{
A list of \code{FLPar} objects.
}
\section{Slots}{
 \describe{
  \item{.Data}{Internal S4 data representation, of class \code{list}.}
  \item{desc}{As textual description of the object contents}
  \item{lock}{Can the object be extended/trimmed? \code{TRUE} or \code{FALSE}.}
  \item{names}{A character vector for the element names} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}
\author{
The FLR Team
}
\seealso{
\link{FLlst}, \link[base]{list}, \link[base]{vector}
}
\keyword{classes}

