% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesArr.R, R/jackknife.R
\docType{class}
\name{FLQuantJK}
\alias{FLQuantJK}
\alias{FLQuantJK-class}
\alias{FLQuantJK,ANY-method}
\alias{orig,FLQuantJK-method}
\alias{orig,FLQuants-method}
\title{A class for jackknifing fisheries data}
\usage{
\S4method{FLQuantJK}{ANY}(object, orig)

\S4method{orig}{FLQuantJK}(object)

\S4method{orig}{FLQuants}(object)
}
\arguments{
\item{object}{Input numeric object}

\item{...}{Additonal arguments}
}
\description{
This extended FLQuant class holds both a jackknifed FLQuant, one in which each
iter is missing one element, and the original object, as a separate
\code{FLQuant} in the \code{orig} slot.
}
\section{Slots}{

\describe{
\item{.Data}{Unnamed slot containing the jackknifed object(\linkS4class{FLQuant}).}
\item{orig}{Original object, (\linkS4class{FLQuant}).}
}
}

\section{Validity}{


\describe{
\item{slot dims}{.Data and orig slots must have the same dimensions 1-5.}
\item{slot dimnames}{.Data and var slots must have the same dimnames 1-5.}
}

You can inspect the class validity function by using
\code{getValidity(getClassDef('FLQuantJK'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

Objects of this class must be constructed from an \linkS4class{FLQuant} that
is to be jackknifed, through the \code{\link{jackknife}} method.
}

\section{Methods}{

All methods defined for the \linkS4class{FLQuant} class are available, but
they will operate only on the jackknifed (\code{.Data}) slot. Please use
\code{orig()} to apply them to the original object stored in the class.
}

\examples{

data(ple4)
fjk <- jackknife(stock(ple4))
# New object has as many iters as length of jackknifed dimension (defaults to 'year')
dim(fjk)

}
\seealso{
\link{FLQuant}
}
\author{
The FLR Team
}
\keyword{classes}
