% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/classesArr.R,
%   R/FLQuantPoint.R
\docType{class}
\name{FLQuantPoint}
\alias{FLQuantPoint}
\alias{FLQuantPoint-methods}
\alias{FLQuantPoint-class}
\alias{FLQuantPoint,FLQuant-method}
\alias{mean,FLQuantPoint-method}
\alias{mean<-,FLQuantPoint,FLQuant-method}
\alias{median,FLQuantPoint-method}
\alias{median<-,FLQuantPoint,FLQuant-method}
\alias{var,FLQuantPoint-method}
\alias{var<-,FLQuantPoint,FLQuant-method}
\alias{lowq,FLQuantPoint-method}
\alias{lowq<-,FLQuantPoint,FLQuant-method}
\alias{uppq,FLQuantPoint-method}
\alias{uppq<-,FLQuantPoint,FLQuant-method}
\alias{quantile,FLQuantPoint-method}
\alias{FLQuantPoint,missing-method}
\alias{n,FLQuantPoint-method}
\title{Class FLQuantPoint}
\usage{
FLQuantPoint(object, ...)

\S4method{FLQuantPoint}{missing}(..., units = "NA", n = 1)

\S4method{FLQuantPoint}{FLQuant}(object, ..., probs = c(0.25, 0.75))

\S4method{n}{FLQuantPoint}(object, ...)
}
\arguments{
\item{object}{Input numeric object}

\item{...}{Additonal arguments}
}
\description{
The \code{FLQuantPoint} class summarizes the contents of an \code{FLQuant}
object with multiple iterations along its sixth dimension using a number of
descriptive statistics.
}
\details{
An object of this class has a set structure along its sixth dimension
(\emph{iter}), which will always be of length 5, and with dimnames
\emph{mean}, \emph{median}, \emph{var}, \emph{uppq} and \emph{lowq}. They
refer, respectively, to the sample mean, sample median, variance, and lower
(0.25) and upper (0.75) quantiles.

Objects of this class wil be typically created from an \code{FLQuant}. The
various statistics are calculated along the \emph{iter} dimension of the
original \code{FLQuant} using \code{\link[base]{apply}}.
}
\section{Slots}{
 \describe{
\item{.Data}{The main array holding the computed statistics. \code{array}.}
\item{units}{Units of measurement. \code{character}.}
}
}

\section{Accesors}{
 \describe{
\item{mean,mean<-:}{'mean' element on 6th dimension, arithmetic mean.}
\item{median,median<-:}{'median' element on 6th dimension, median.}
\item{var,var<-:}{'var' element on 6th dimension, variance.}
\item{lowq,lowq<-:}{'lowq' element on 6th dimension, lower quantile (0.25 by default).}
\item{uppq,uppq<-:}{'uppq' element on 6th dimension, upper quantile (0.75 by default).}
\item{quantile:}{returns the 'lowq' or 'uppq' iter, depending on the value of 'probs' (0.25 or 0.75).}
}
}

\section{Constructor}{

Inputs can be of class:
\describe{
\item{\code{FLQuant}:}{An FLQuant object with iters (i.e. dim\link{6} > 1)}
}
}

\section{Validity}{
 \describe{
\item{iter:}{iter dimension is of length 5.}
\item{Dimnames:}{iter dimnames are 'mean', 'median', 'var', 'uppq' and'lowq'}
}
}

\examples{

flq <- FLQuant(rlnorm(2000), dim=c(10,20,1,1,1,200), units="kg")
flqp <- FLQuantPoint(flq)
flqp <- FLQuantPoint(flq, probs=c(0.05, 0.95))
summary(flqp)
mean(flqp)
var(flqp)
rnorm(200, flqp)
}
\seealso{
\link{FLQuant}
}
\author{
The FLR Team
}
\keyword{classes}
