% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{FLQuants}
\alias{FLQuants}
\alias{FLQuants,ANY-method}
\alias{FLQuants,FLQuants-method}
\alias{FLQuants,list-method}
\alias{FLQuants,missing-method}
\alias{FLQuants-class}
\alias{FLQuants-methods}
\title{Class FLQuants}
\description{
\code{FLQuants} is a \code{list} of \code{FLQuant} objects.
It is very similar to the standard \code{list} class.
It implements a lock mechanism that, when turned on, does
not allow the user to increase or decrease the object length.
The elements of \code{FLQuants} must all be of class  \code{FLQuant}.
}
\section{Slots}{
 \describe{
    \item{.Data}{The data. \code{list}.}
    \item{names}{Names of the list elements. \code{character}.}
    \item{desc}{Description of the object. \code{character}.}
    \item{lock}{Lock mechanism, if turned on the length of the list can not be modified by adding or removing elements. \code{logical}.}
}
}

\section{Accesors}{

Elements in the list can be accessed using '[' and '[['
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}
\author{
The FLR Team
}
\seealso{
\link[base]{*}, \link[methods]{Arith}, \link[base]{as.data.frame},
\link{bubbles}, \link{catch<-}, \link{iter}, \link[stats]{model.frame},
\link[methods]{show}, \link[base]{summary}, \link[lattice]{xyplot},
\link{FLlst}, \link[base]{list}
}
\keyword{classes}

