% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/classesLst.R
\docType{class}
\name{FLQuants}
\alias{FLQuants}
\alias{FLQuants-methods}
\alias{FLQuants-class}
\alias{FLQuants,ANY-method}
\alias{FLQuants,FLComp-method}
\alias{FLQuants,missing-method}
\alias{FLQuants,list-method}
\alias{FLQuants,FLQuants-method}
\title{Class FLQuants}
\usage{
FLQuants(object, ...)

\S4method{FLQuants}{ANY}(object, ...)

\S4method{FLQuants}{FLComp}(object, ...)

\S4method{FLQuants}{missing}(object, ...)

\S4method{FLQuants}{list}(object, ...)

\S4method{FLQuants}{FLQuants}(object, ...)
}
\arguments{
\item{object}{unnamed object to be added to the list}

\item{...}{other named or unnamed objects}
}
\description{
\code{FLQuants} is a \code{list} of \code{FLQuant} objects. It is very
similar to the standard \code{list} class. It implements a lock mechanism
that, when turned on, does not allow the user to increase or decrease the
object length. The elements of \code{FLQuants} must all be of class
\code{FLQuant}.
}
\section{Slots}{
 \describe{ \item{.Data}{The data. \code{list}.}
\item{names}{Names of the list elements. \code{character}.}
\item{desc}{Description of the object. \code{character}.} \item{lock}{Lock
mechanism, if turned on the length of the list can not be modified by adding
or removing elements. \code{logical}.} }
}

\section{Constructor}{

A constructor method exists for this class that can take named arguments for
any of the list elements.
}

\examples{

# Compute various FLStock indicators
  data(ple4)
  fqs <- FLQuants(ssb=ssb(ple4), catch=catch(ple4), rec=rec(ple4),
    f=fbar(ple4))
  summary(fqs)
  xyplot(data~year|qname, fqs, type='b', scales=list(relation='free'))

}
\seealso{
\link[base]{*}, \link[methods]{Arith}, \link[base]{as.data.frame},
\link{bubbles}, \link{catch<-}, \link{iter}, \link[stats]{model.frame},
\link[methods]{show}, \link[base]{summary}, \link[lattice]{xyplot},
\link{FLlst}, \link[base]{list}
}
\author{
The FLR Team
}
\keyword{classes}
