\name{FLQuants}
\docType{methods}
\alias{FLQuants-methods}
\alias{FLQuants,ANY-method}
 \alias{FLQuants,missing-method}
 \alias{FLQuants,list-method}
 \alias{FLQuants,FLQuants-method}
\title{Method FLQuants}
\description{
 My method is better then yours !
}
\section{Generic function}{
 FLQuants(object) 
}
\section{Methods}{\describe{
\item{ signature(object=ANY) :}{Describe method}
\item{ signature(object=missing) :}{Describe method}
\item{ signature(object=list) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
