\name{FLSR}
\docType{class}
\alias{FLSR}
\alias{FLSR-class}
\title{Class FLSR}
\description{
 My class is better then yours !
}
\section{Slots}{\describe{
\item{rec (\code{FLQuant}):}{Describe slot}
\item{ssb (\code{FLQuant}):}{Describe slot}
\item{rechat (\code{FLQuant}):}{Describe slot}
\item{residuals (\code{FLQuant}):}{Describe slot}
\item{covar (\code{FLQuants}):}{Describe slot}
\item{model (\code{formula}):}{Describe slot}
\item{lkhd (\code{function}):}{Describe slot}
\item{initial (\code{function}):}{Describe slot}
\item{params (\code{FLPar}):}{Describe slot}
\item{logLik (\code{logLik}):}{Describe slot}
\item{vcov (\code{array}):}{Describe slot}
\item{details (\code{list}):}{Describe slot}
\item{name (\code{character}):}{Describe slot}
\item{desc (\code{character}):}{Describe slot}
\item{range (\code{numeric}):}{Describe slot}
}}
\section{Extends}{\describe{
FLModel
FLComp
}}
\section{Methods}{\describe{
All methods are inherited.
}}
\author{The FLR Team}
\seealso{ \link{FLModel}, \link{FLComp} }
\examples{}
\keyword{classes}
