% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLSR.R
\docType{class}
\name{FLSR}
\alias{FLSR}
\alias{FLSR,ANY-method}
\alias{FLSR,missing-method}
\alias{FLSR-class}
\alias{FLSR-methods}
\title{Class FLSR}
\description{
A class for stock-recruitment models.
}
\details{
A series of commonly-used stock-recruitment models are already available,
including the corresponding likelihood functions and calculation of initial
values. See \code{\link{SRModels}} for more details and the exact
formulation implemented for each of them.
}
\section{Slots}{
 \describe{ \item{name}{Name of the object
(\code{character}).} \item{desc}{Description of the object
(\code{character}).} \item{range}{Named numeric vector containing the quant
and year ranges (\code{numeric}).} \item{rec}{Recruitment series
(\code{FLQuant}).} \item{ssb}{Index of reproductive potential, e.g. SSB or
egg oor egg production (\code{FLQuant}).} \item{fitted}{Estimated values for
rec (\code{FLQuant}).} \item{residuals}{Residuals obtained from the model
fit (\code{FLArray}).} \item{covar}{Covariates for SR model
(\code{FLQuants}).} \item{model}{Model formula (\code{formula}).}
\item{gr}{Function returning the gradient of the likelihood
(\code{function}).} \item{logl}{Log-likelihood function (\code{function}).}
\item{initial}{Function returning initial parameter values for the optimizer
(\code{function}).} \item{params}{Estimated parameter values (\code{FLPar}).}
\item{logLik}{Value of the log-likelihood (\code{logLik}).}
\item{vcov}{Variance-covariance matrix (\code{array}).}
\item{details}{Extra information on the model fit procedure (\code{list}).}
\item{logerror}{Is the error on a log scale (\code{logical}).}
\item{distribution}{(\code{factor}).} \item{hessian}{Resulting Hessian
matrix from the fit (\code{array}).} }
}
\examples{

# Create an empty FLSR object.
  sr1 <- FLSR()
  slotNames(sr1)

  # Create an FLSR object using the existing SR models.
    sr2 <- FLSR(model = 'ricker')
    sr2@model
    sr2@initial
    sr2@logl

    sr3 <- FLSR(model = 'bevholt')
    sr3@model
    sr3@initial
    sr3@logl

  # Create an FLSR using a function.
    mysr1 <- function(){
        model <- rec ~ a*ssb^b
        return(list(model = model))}

    sr4 <- FLSR(model = mysr1)

  # Create an FLSR using a function and check that it works.
    mysr2 <- function(){
        formula <- rec ~ a+ssb*b

        logl <- function(a, b, sigma, rec, ssb) sum(dnorm(rec,
            a + ssb*b, sigma, TRUE))

       initial <- structure(function(rec, ssb) {
            a     <- min(rec)
            b     <- 1
            sigma <- sqrt(var(rec-a-ssb*b))

            return(list(a = a, b = b, sigma = sigma))}, lower = c(0, 1e-04, 1e-04), upper = rep(Inf, 3))

       return(list(model = formula, initial = initial, logl = logl))
    }

    ssb <- FLQuant(runif(10, 10000, 100000))
    rec <- 10000 + 2*ssb + rnorm(10,0,1)
    sr5 <- FLSR(model = mysr2, ssb = ssb, rec = rec)

    sr5.mle <- fmle(sr5)
    sr5.nls <- nls(sr5)

# NS Herring stock-recruitment dataset
  data(nsher)

  # already fitted with a Ricker SR model
    summary(nsher)
    plot(nsher)

  # change model
    model(nsher) <- bevholt()

  # fit through MLE
    nsher <- fmle(nsher)
    plot(nsher)

}
\author{
The FLR Team
}
\seealso{
\link{FLModel}, \link{FLComp}
}
\keyword{classes}

