\name{FLStock}
\docType{methods}
\alias{FLStock}
\alias{FLStock-methods}
\alias{FLStock,missing-method}
\alias{FLStock,FLQuant-method}
\title{Create FLStock objects}
\description{
  Constructor method for the FLStock class
}
\section{Arguments}{
  \describe{
  \item{object}{An initisal FLQuant used for sizing.}
  \item{plusgroup}{The age/length to be considred as plusgroup.}
  }
}
\section{Generic function}{
 FLStock(object, ...)
}
\section{Methods}{\describe{
\item{ signature(object=missing) :}{Describe method}
\item{ signature(object=FLQuant) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{
}
\keyword{methods}
