% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classesComp.R
\docType{class}
\name{FLStockLen}
\alias{FLStockLen}
\alias{FLStockLen,FLQuant-method}
\alias{FLStockLen,missing-method}
\alias{FLStockLen-class}
\alias{FLStockLen-methods}
\alias{catch,FLStockLen-method}
\alias{catch.n,FLStockLen-method}
\alias{catch.n<-,FLStockLen,FLQuant-method}
\alias{catch.wt,FLStockLen-method}
\alias{catch.wt<-,FLStockLen,FLQuant-method}
\alias{catch<-,FLStockLen,FLQuant-method}
\alias{desc,FLStockLen-method}
\alias{desc<-,FLStockLen,character-method}
\alias{discards,FLStockLen-method}
\alias{discards.n,FLStockLen-method}
\alias{discards.n<-,FLStockLen,FLQuant-method}
\alias{discards.wt,FLStockLen-method}
\alias{discards.wt<-,FLStockLen,FLQuant-method}
\alias{discards<-,FLStockLen,FLQuant-method}
\alias{halfwidth,FLStockLen-method}
\alias{halfwidth<-,FLStockLen,numeric-method}
\alias{harvest,FLStockLen-method}
\alias{harvest.spwn,FLStockLen-method}
\alias{harvest.spwn<-,FLStockLen,FLQuant-method}
\alias{harvest<-,FLStockLen,FLQuant-method}
\alias{landings,FLStockLen-method}
\alias{landings.n,FLStockLen-method}
\alias{landings.n<-,FLStockLen,FLQuant-method}
\alias{landings.wt,FLStockLen-method}
\alias{landings.wt<-,FLStockLen,FLQuant-method}
\alias{landings<-,FLStockLen,FLQuant-method}
\alias{m,FLStockLen-method}
\alias{m.spwn,FLStockLen-method}
\alias{m.spwn<-,FLStockLen,FLQuant-method}
\alias{m<-,FLStockLen,FLQuant-method}
\alias{mat,FLStockLen-method}
\alias{mat<-,FLStockLen,FLQuant-method}
\alias{name,FLStockLen-method}
\alias{name<-,FLStockLen,character-method}
\alias{range,FLStockLen-method}
\alias{range<-,FLStockLen,numeric-method}
\alias{stock,FLStockLen-method}
\alias{stock.n,FLStockLen-method}
\alias{stock.n<-,FLStockLen,FLQuant-method}
\alias{stock.wt,FLStockLen-method}
\alias{stock.wt<-,FLStockLen,FLQuant-method}
\alias{stock<-,FLStockLen,FLQuant-method}
\title{Class FLStockLen}
\description{
A class for modelling a length structured fish stock.
}
\details{
The \code{FLStockLen} object contains a length based representation of a fish stock
This includes information on removals (i.e. catches, landings and discards),
maturity, natural mortality and the results of an analytical assessment (i.e.
estimates of abundance and removal rates).
}
\section{Slots}{

    \describe{
    \item{halfwidth}{The middle of the length bins (\code{numeric}).}
    \item{catch}{Total catch weight (\code{FLQuant}).}
    \item{catch.n}{Catch numbers (\code{FLQuant}).}
    \item{catch.wt}{Mean catch weights (\code{FLQuant}).}
    \item{discards}{Total discards weight (\code{FLQuant}).}
    \item{discards.n}{Discard numbers (\code{FLQuant}).}
    \item{discards.wt}{Mean discard weights (\code{FLQuant}).}
    \item{landings}{Total landings weight (\code{FLQuant}).}
    \item{landings.n}{Landing numbers (\code{FLQuant}).}
    \item{landings.wt}{Landing weights (\code{FLQuant}).}
    \item{stock}{Total stock weight (\code{FLQuant}).}
    \item{stock.n}{Stock numbers (\code{FLQuant}).}
    \item{stock.wt}{Mean stock weights (\code{FLQuant}).}
    \item{m}{Natural mortality (\code{FLQuant}).}
    \item{mat}{Proportion mature (\code{FLQuant}).}
    \item{harvest}{Harvest rate or fishing mortality. The units of the FLQuant should be set to 'harvest' or 'f' accordingly (\code{FLQuant}).}
    \item{harvest.spwn}{Proportion of harvest/fishing mortality before spawning (\code{FLQuant}).}
    \item{m.spwn}{Proportion of natural mortality before spawning (\code{FLQuant}).}
    \item{name}{Name of the stock (\code{character}).}
    \item{desc}{Description of stock (\code{character}).}
    \item{range}{Named numeric vector containing the quant and year ranges, the plusgroup and the quant range that the average fishing mortality is calculated over (\code{numeric}).}
}
}

\section{Accesors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity. If an unnamed \code{FLQuant} object is provided, this is used
for sizing but not stored in any slot.
}

\section{Validity}{
 \describe{
    \item{Dimensions}{All FLQuant slots must have iters equal to 1 or 'n'.}
    \item{Iters}{The dimname for iter[1] should be '1'.}
    \item{Dimnames}{The name of the quant dimension must be the same for all FLQuant slots.}
    \item{Totals}{The length of the quant dimension for the totals slots (catch, landings and discards) must be equal to 1.}
}
}
\author{
The FLR Team
}
\seealso{
\link[base]{[}, \link[base]{[<-}, \link{as.FLSR},
\link{computeCatch}, \link{computeDiscards}, \link{computeLandings}, \link[graphics]{plot}, \link{ssb},
\link{ssbpurec}, \link{trim}, \link{FLComp}
}
\keyword{classes}

