\name{FLStocks}
\docType{methods}
\alias{FLStocks-methods}
\alias{FLStocks,ANY-method}
 \alias{FLStocks,missing-method}
 \alias{FLStocks,list-method}
\title{Method FLStocks}
\description{
 My method is better then yours !
}
\section{Generic function}{
 FLStocks(object) 
}
\section{Methods}{\describe{
\item{ signature(object=ANY) :}{Describe method}
\item{ signature(object=missing) :}{Describe method}
\item{ signature(object=list) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
