% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{FUNCTION}
\alias{FUNCTION}
\title{Extract and modify the recruitment time series}
\usage{
\S4method{rec}{FLStock}(object, rec.age = as.character(object@range["min"]))
}
\arguments{
\item{object}{An object of class 'FLStock'}

\item{rec.age}{What age to extract, defaults to first one. As 'character' to select by name or as 'numeric' by position.}
}
\value{
RETURN Lorem ipsum dolor sit amet
}
\description{
Recruitment in number of fish is the first row of the 'stock.n' slot of
an age-structured 'FLStock'. These convenience functions allow a clearer
syntax when retrieving of altering the content of 'stock.n\link{rec.age,}', where
'rec.age' is usually the first age in the object.
}
\examples{
data(ple4)
rec(ple4)
# Multiple recruitment by a factor of 2
rec(ple4) <- rec(ple4) * 2
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
