% $Id: is.FLStock.Rd,v 1.6.2.1 2005/12/19 11:13:48 iagoazti Exp $
\name{is.FLStock}
\alias{is.FLStock}
\title{FLStock}
\description{Testing of FLStock objects}
\usage{
  is.FLStock(x)
}
\arguments{
  \item{x}          {object to test.}
}
\details{
}
\value{
	\code{is.FLStock} returns \code{TRUE} or \code{FALSE}
}
\seealso{
    \code{\link{FLStock-class}}
}
\examples{\dontrun{
    data(ple4)
    is.FLStock(ple4)
}
}
\author{FLR Team}
\keyword{classes}
\keyword{manip}
