% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLArray.R
\docType{methods}
\name{iter}
\alias{iter}
\alias{iter-methods}
\alias{iter,FLArray-method}
\alias{iter,FLQuant,ANY-method}
\alias{iter,FLCohort,ANY-method}
\title{Methods iter}
\usage{
iter(obj, ...)

\S4method{iter}{FLArray}(obj, iter)
}
\description{
Select or modify iterations of an FLR object
}
\details{
To extract or modify a subset of the iterations contained in an FLR object,
the \code{iter} and \code{iter<-} methods can be used.

In complex objects with various \code{FLQuant} slots, the \code{iter} method
checks whether individual slots contain more than one iteration, i.e.
\code{dims(object)[6] > 1}. If a particular slot contains a single
iteration, that is returned, otherwise the chosen iteration is selected.
This is in contrast with the subset operator \code{[}, which does not carry
out this check.

For objects of class \code{\link{FLModel}}, iters are extracted for slots of
classes \code{FLQuant}, \code{FLCohort} and \code{FLPar}.
}
\section{Generic function}{
 iter(object) iter<-(object,value)
}

\examples{
# For an FLQuant
  flq <- FLQuant(rnorm(800), dim=c(4,10,2), iter=10)
  iter(flq, 2)
# For the more complex FLStock object
  data(ple4)
  fls <- propagate(ple4, 10)
  # Extraction using iter...
    fls2 <- iter(fls, 2)
    summary(fls2)
}
\seealso{
\linkS4class{FLComp}, \linkS4class{FLQuant}
}
\author{
The FLR Team
}
\keyword{methods}
