% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\docType{methods}
\name{jackknife}
\alias{jackknife}
\alias{jackknife-methods}
\alias{jackknife,FLQuant-method}
\alias{jackknife,FLQuants-method}
\alias{jackknife,FLModel-method}
\title{Method jackknife}
\usage{
\S4method{jackknife}{FLQuant}(object, dim = "year", na.rm = TRUE)

\S4method{jackknife}{FLQuants}(object, ...)

\S4method{jackknife}{FLModel}(object, slot)
}
\description{
Jackknife resampling
}
\details{
The \code{jackknife} method sets up objects ready for jackknifing, i.e. to
systematically recompute a given statistic leaving out one observation at a
time. From this new set of "observations" for the statistic, estimates for
the bias and variance of the statstic can be calculated.

Input objects cannot have length > 1 along the \code{iter} dimension, and
the main slot in the resulting object will have as many \code{iter}s as the
number of elements in the original object that are not \code{NA}.
}
\section{Generic function}{
 jackknife(object, ...)
}

\examples{

flq <- FLQuant(1:8)
flj <- jackknife(flq)
iters(flj)

}
\seealso{
\code{\link{FLQuantJK}} \code{\link{FLParJK}}
}
\author{
The FLR Team
}
\keyword{methods}
