% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\docType{methods}
\name{jackknife}
\alias{jackknife}
\alias{jackknife-methods}
\alias{jackknife,FLQuant-method}
\title{Method jackknife}
\usage{
\S4method{jackknife}{FLQuant}(object, dim = "year", na.rm = TRUE)
}
\description{
Jackknife resampling
}
\details{
The \code{jackknife} method sets up objects ready for jackknifing, i.e. to
systematically recompute a given statistic leaving out one observation at a
time. From this new set of "observations" for the statistic, estimates for
the bias and variance of the statstic can be calculated.

Input objects cannot have length > 1 along the \code{iter} dimension, and
the resulting object will have one more \code{iter} than the number of
elements in the original object.
}
\section{Generic function}{
 jackknife(object, ...)
}

\examples{

flq <- FLQuant(1:8)
flj <- jackknife(flq)
iters(flj)

}
\seealso{
\code{\link{FLQuant}}
}
\author{
The FLR Team
}
\keyword{methods}
