\name{median}
\docType{methods}
\alias{upper}
\alias{upper,FLModel-method}
\alias{upper<-}
\alias{upper<-,FLModel,numeric-method}
\alias{lower}
\alias{lower,FLModel-method}
\alias{lower<-}
\alias{lower<-,FLModel,numeric-method}
\title{Methods upper and lower}
\description{
 My method is better then yours !
}
\section{Generic function}{
upper(object) 
upper<-(object, value) 
lower(object) 
lower<-(object, value) 
}
\section{Methods}{\describe{
\item{signature(object=FLModel) :}{Describe method}
\item{signature(object=FLModel, value=numeric) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
