\name{lowess}
\docType{methods}
\alias{lowess,FLSR,missing-method}
\title{Method lowess}
\description{
  LOWESS smoother based on locally-weighted polynomial regression for FLSR objects
}
\section{Generic function}{
 lowess(x, y, f=2/3, iter=3, delta=0.01 * diff(range(xy\$x[o])))

}
\section{Methods}{\describe{
\item{signature(x=FLSR,y=missing) :}{lowessb(x, y, f=2/3, iter=3, delta=0.01 * diff(range(ssb(x))))
}
}}
\author{The FLR Team}
\seealso{ \link{lowess} }
\examples{}
\keyword{methods}
