% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{mase}
\alias{mase}
\alias{mase,FLQuant,FLQuants-method}
\alias{mase,FLIndices,list-method}
\title{Compute mean absolute scaled error (MASE)}
\usage{
mase(ref, preds, ...)

\S4method{mase}{FLQuant,FLQuants}(ref, preds, order = c("inverse", "ahead"))

\S4method{mase}{FLIndices,list}(ref, preds, order = "inverse", wt = "missing")
}
\arguments{
\item{ref}{Reference or naive prediction.}

\item{preds}{Predicitions to compare to reference.}

\item{...}{Extra arguments.}

\item{order}{Are predictions in 'inverse' (default) or 'ahead' order.}

\item{wt}{Mean weights-at-age to use with indices.}
}
\value{
A numeric vector of the same length as 'preds'.
}
\description{
Franses, PH. "A note on the Mean Absolute Scaled Error". International Journal of Forecasting. 32 (1): 20–22. doi:10.1016/j.ijforecast.2015.03.008.
}
