% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{meanage}
\alias{meanage}
\alias{meanageCatch}
\title{Calculate the mean age in the stock and catch}
\usage{
meanage(object)

meanageCatch(object)
}
\arguments{
\item{object}{An age-structured FLStock object}
}
\value{
An FLQuant object
}
\description{
Average age in the stock numbers or catch-at-age.
}
\examples{
data(ple4)
meanage(ple4)
meanageCatch(ple4)
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{ts}
