% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLComp.R, R/FLS.R
\docType{methods}
\name{metrics}
\alias{metrics}
\alias{metrics-methods}
\alias{metrics,FLComp,list-method}
\alias{metrics,FLS,missing-method}
\title{Extract simply-defined metrics from compex objects}
\usage{
metrics(object, metrics, ...)

\S4method{metrics}{FLComp,list}(object, metrics)

\S4method{metrics}{FLS,missing}(object, metrics, ...)
}
\arguments{
\item{object}{A complex \strong{FLR} object from which to extract time series metrics.}
}
\value{
An object, generally of class \code{\link{FLQuants}}.
}
\description{
Time series summaries of complex objects are commonly needed, for example for
plotting the inputs and outputs of a class like \code{\link{FLStock}}. These
methods allow for simple specification of those metrics by means of function
calls and formulas.
}
\examples{
data(ple4)
# missing
metrics(ple4)
# metrics = function
metrics(ple4, metrics=function(x) FLQuants(SSB=ssb(x), REC=rec(x),
  F=fbar(x), SSBREC=ssb(x) / rec(x)))
# metrics = formula
metrics(ple4, metrics=SSB~ssb)
metrics(ple4, metrics=~ssb)
# metrics = list
metrics(ple4, metrics=list(SSB=ssb, REC=rec, F=fbar))
metrics(ple4, metrics=list(SSB=~ssb, REC=rec, F=fbar))
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{methods}
