% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{mohnMatrix}
\alias{mohnMatrix}
\title{Generate a matrix to compute Mohn's rho for a single metric}
\usage{
mohnMatrix(stocks, metric = "fbar", ...)
}
\arguments{
\item{stocks}{An FLStocks object from a restrospective analysis}

\item{metric}{Metric to be computed, as a character vector or function}
}
\value{
A metrics of n + 2 x n, where n is the numbers of objects in stocks.
}
\description{
A common measure of the strength of stock assessment retrospective
patterns is Mohn's rho. This function does not carry out the calculation
but returns a matrix with the metrics value for the n restrospective
runs, in columns, and n + 2 years, in rows.
}
