% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R
\docType{methods}
\name{names}
\alias{names}
\alias{names,FLArray-method}
\alias{names,FLPar-method}
\alias{names<-,FLPar,character-method}
\title{Method names}
\usage{
\S4method{names}{FLArray}(x)

\S4method{names}{FLPar}(x)

\S4method{names}{FLPar,character}(x) <- value
}
\description{
The \code{names} method returns the names of the dimnames of an object. For
some classes, the names attribute can be modified directly using names<-.
}
\section{Generic function}{
 names(x) names<-(x, value)
}

\examples{
# FLQuant
data(ple4)
names(catch.n(ple4))

# Contrast this with
dimnames(catch.n(ple4))

}
\seealso{
\link[base]{names}
}
\author{
The FLR Team
}
\keyword{methods}
