\name{names}
\docType{methods}
\alias{names,ANY-method}
\alias{names,FLArray-method}
\alias{names,FLlst-method}
\title{Method names}
\description{
 the \code{names} returns the names of the dimnames of an object
}
\section{Generic function}{
 names(x) 
}
\section{Methods}{\describe{
\item{ signature(x=FLQuant) :}{Returns the names of the dimnames of x}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
