\name{plot}
\docType{methods}
\alias{plot,ANY,ANY-method}
 \alias{plot,FLQuant,missing-method}
 \alias{plot,FLQuantPoint,missing-method}
 \alias{plot,FLPar,missing-method}
 \alias{plot,FLStock,missing-method}
 \alias{plot,FLStocks,missing-method}
 \alias{plot,FLStocks,FLPar-method}
 \alias{plot,FLBiol,missing-method}
 \alias{plot,FLCohort,ANY-method}
 \alias{plot,FLIndex,missing-method}
 \alias{plot,FLIndices,missing-method}
 \alias{plot,FLSR,missing-method}
\title{Method plot}
\description{
 My method is better then yours !
}
\section{Generic function}{
 plot(x,y) 
}
\section{Methods}{\describe{
\item{ signature(x=ANY,y=ANY) :}{Describe method}
\item{ signature(x=FLQuant,y=missing) :}{Describe method}
\item{ signature(x=FLQuantPoint,y=missing) :}{Describe method}
\item{ signature(x=FLPar,y=missing) :}{Describe method}
\item{ signature(x=FLStock,y=missing) :}{Describe method}
\item{ signature(x=FLStocks,y=missing) :}{Describe method}
\item{ signature(x=FLBiol,y=missing) :}{Describe method}
\item{ signature(x=FLCohort,y=ANY) :}{Describe method}
\item{ signature(x=FLIndex,y=missing) :}{Describe method}
\item{ signature(x=FLSR,y=missing) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
