\name{predict}
\docType{methods}
\alias{predict,FLModel-method}
\title{Method predict}
\description{
 \code{predict} is a generic function for predictions from the results of various model fitting functions.
 The function invokes particular \textit{methods} which depend on the class of the first argument.
}
\section{Generic function}{
 predict(object, ...)
}
\section{Methods}{\describe{
\item{ signature(object=FLModel) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
