% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\docType{class}
\name{predictModel}
\alias{predictModel}
\alias{predictModel-methods}
\alias{predictModel-class}
\alias{predictModel,FLQuants,formula-method}
\alias{predictModel,FLQuants,missing-method}
\alias{predictModel,FLQuants,character-method}
\alias{predictModel,FLQuants,function-method}
\alias{predictModel,FLQuants,list-method}
\alias{predictModel,missing,ANY-method}
\title{A class for model prediction}
\usage{
\S4method{predictModel}{FLQuants,formula}(object, model,
  params = FLPar())

\S4method{predictModel}{FLQuants,missing}(object, params = FLPar())

\S4method{predictModel}{FLQuants,character}(object, model,
  params = FLPar())

\S4method{predictModel}{FLQuants,`function`}(object, model,
  params = FLPar())

\S4method{predictModel}{FLQuants,list}(object, model, params = FLPar())

\S4method{predictModel}{missing,ANY}(object, model, ...)
}
\description{
Object of the predictModel class are used in various FLR classes to allow
flexible modelling of the dynamics of different biological and technological
processes.
}
\details{
The dependency of life history processes, such as maturity and fecundity, to biological
and environmental factors, can be represented in objects of this class via a simple model
(represented by a `formula`) and the corresponding paramaters (`FLPar`) and inputs
(`FLQuants`).
}
\section{Slots}{

    \describe{
    \item{.Data}{Inputs to the model not found in enclosing class (\code{FLQuants}).}
    \item{model}{Model representation (\code{formula}).}
    \item{params}{Model paramaters (\code{FLPar}).}
}
}

\section{Validity}{


  \describe{
    \item{VALIDITY}{Neque porro quisquam est qui dolorem ipsum.}
}

You can inspect the class validity function by using
   \code{getValidity(getClassDef('predictModel'))}
}

\section{Accessors}{

All slots in the class have accessor and replacement methods defined that
allow retrieving and substituting individual slots.

The values passed for replacement need to be of the class of that slot.
A numeric vector can also be used when replacing FLQuant slots, and the
vector will be used to substitute the values in the slot, but not its other
attributes.
}

\section{Constructor}{

A construction method exists for this class that can take named arguments for
any of its slots. All slots are then created to match the requirements of the
class validity.
}

\section{Methods}{

Methods exist for various calculations based on values stored in the class:

\describe{
    \item{METHOD}{Neque porro quisquam est qui dolorem ipsum.}
}
}

\examples{
fec <- FLQuants(fec=FLQuant(rlnorm(10, 20, 5),
  dimnames=list(year=2000:2009), units='1'))
predictModel(fec, model=~fec)
predictModel(fec)
predictModel(fec, model="bevholt")
predictModel(fec, model=bevholt)
predictModel(fec, model=bevholt())
predictModel(model=rec~a*ssb, params=FLPar(a=1.234))
predictModel(model=bevholt, params=FLPar(a=1.234))
predictModel(model="bevholtss3", params=FLPar(a=1.234))
}
\seealso{
\code{\link{FLQuants}} \code{\link{FLPar}} \code{\link{FLBiol}}
}
\author{
The FLR Team
}
\keyword{classes}
