% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLQuant.R
\docType{methods}
\name{propagate}
\alias{propagate}
\alias{propagate-methods}
\alias{propagate,FLQuant-method}
\title{Method propagate}
\usage{
propagate(object, ...)

\S4method{propagate}{FLQuant}(object, iter, fill.iter = TRUE)
}
\arguments{
\item{object}{Object to be propagated.}

\item{fill.iter}{Should first array be copied to others? Defaults to FALSE.}

\item{iters}{No. of iterations in output.}
}
\description{
Methods to extend objects of various FLR classes along the \code{iter}
(6th FLQuant) dimension. Objects must generally have a single
\code{iter} to be extended. The new iterations can be filled with copies
of the existing, or remain as \code{NA}.
}
\section{Generic function}{
 propagate(object, ...)
}

\examples{

# An FLQuant with one iter (dim(flq)[6] == 1)
flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age')

# can now be extended along the `iter` dimension, with
#' copies of the first
propagate(flq, 100)

# or without
iter(propagate(flq, 100, fill.iter=FALSE), 2)
}
\seealso{
\linkS4class{FLQuant}
}
\author{
The FLR Team
}
\keyword{methods}
