% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{pv}
\alias{pv}
\alias{pv,FLQuant-method}
\alias{pv-methods}
\title{Method pv}
\usage{
\S4method{pv}{FLQuant}(object, dist = FALSE)
}
\description{
Population variability
}
\details{
The \code{pv} method computes the population variability (\emph{pv}) of an
\code{FLQuant} object.
}
\section{Generic function}{
 pv(object)
}
\examples{

flq <- FLQuant(rnorm(40), dim=c(1,40))
pv(flq)

data(ple4)
pv(stock(ple4))

}
\author{
The FLR Team
}
\references{
Heath, J.P. 2006. Quantifying temporal variability in population
abundances. \emph{Oikos} \bold{115 (3)}: 573--581.
}
\seealso{
\link{FLComp}
}
\keyword{methods}

