% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLArray.R
\docType{methods}
\name{quant}
\alias{quant}
\alias{quant,FLArray-method}
\alias{quant-methods}
\alias{quant<-,FLArray,character-method}
\title{Method quant}
\usage{
quant(object, ...)

\S4method{quant}{FLArray}(object)

\S4method{quant}{FLArray,character}(object) <- value
}
\description{
Function to get or set the name of the first dimension (quant) in an object
of any FLArray-based class, like \code{\link{FLQuant}} or \code{\link{FLCohort}}.
}
\section{Generic function}{
 quant(object) quant<-(object,value)
}
\examples{

# quant is 'quant' by default
  quant(FLQuant())

flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age')
quant(flq)
quant(flq) <- 'length'
summary(flq)


# quant is 'quant' by default
  quant(FLQuant())

flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age')
quant(flq)
quant(flq) <- 'length'
summary(flq)

}
\author{
The FLR Team
}
\seealso{
\linkS4class{FLQuant}, \linkS4class{FLCohort}
}
\keyword{methods}

