\name{quantile}
\docType{methods}
\alias{quantile,ANY-method}
 \alias{quantile,FLQuant-method}
 \alias{quantile,FLQuantPoint-method}
\title{Method quantile}
\description{
 My method is better then yours !
}
\section{Generic function}{
 quantile(x) 
}
\section{Methods}{\describe{
\item{ signature(x=ANY) :}{Describe method}
\item{ signature(x=FLQuant) :}{Describe method}
\item{ signature(x=FLQuantPoint) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
