\name{range}
\docType{methods}
\alias{range-methods}
\alias{range,FLComp,missing-method}
\alias{range,FLComp-method}
\alias{range,FLlst-method}
\alias{range<-}
\alias{range<--methods}
\alias{range<-,FLComp-method}
\title{Method range}
\description{
  Extraction and modification of range slot in complex FLR classes
}
\section{Generic function}{
 range(x, i) 
 quant<-(x, i, value) 
}
\section{Methods}{\describe{
\item{ signature(x=FLComp) :}{Get the name of the first dimension of an \code{FLQuant} object.}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
