% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.VPAsuite.R
\name{readVPAIntercatch}
\alias{readVPAIntercatch}
\title{Reads a single file with one year of data in VPA format as output by
ICES Intercatch}
\usage{
readVPAIntercatch(file)
}
\arguments{
\item{file}{Intercatch VPA file to load}
}
\value{
An object of class FLQuant.
}
\description{
Reads a single file with one year of data in VPA format as output by
ICES Intercatch
}
