% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLQuant.R
\docType{methods}
\name{residuals-FLQuant}
\alias{residuals-FLQuant}
\title{residuals}
\usage{
\S4method{residuals}{FLQuant}(object, fit, type = "log")
}
\description{
residuals
}
\examples{
data(ple4)
fit <- rlnorm(1, log(catch(ple4)), 0.1)
residuals(catch(ple4), fit)
residuals(catch(ple4), fit, type="student")
rraw(catch(ple4), fit)
rlogstandard(catch(ple4), fit)
rstandard(catch(ple4), fit)
rstudent(catch(ple4), fit)
}
