\name{rgamma}
\docType{methods}
\alias{rgamma,ANY,ANY,ANY,ANY-method}
 \alias{rgamma,numeric,FLQuantPoint,missing,missing-method}
\title{Method rgamma}
\description{
 My method is better then yours !
}
\section{Generic function}{
 rgamma(n,shape,rate,scale) 
}
\section{Methods}{\describe{
\item{ signature(n=ANY,shape=ANY,rate=ANY,scale=ANY) :}{Describe method}
\item{ signature(n=numeric,shape=FLQuantPoint,rate=missing,scale=missing) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
