\name{rlnorm}
\docType{methods}
\alias{rlnorm,ANY,ANY,ANY-method}
 \alias{rlnorm,numeric,FLQuant,FLQuant-method}
 \alias{rlnorm,numeric,FLQuant,numeric-method}
 \alias{rlnorm,numeric,numeric,FLQuant-method}
 \alias{rlnorm,numeric,FLQuantPoint,missing-method}
\title{Method rlnorm}
\description{
 My method is better then yours !
}
\section{Generic function}{
 rlnorm(n,meanlog,sdlog) 
}
\section{Methods}{\describe{
\item{ signature(n=ANY,meanlog=ANY,sdlog=ANY) :}{Describe method}
\item{ signature(n=numeric,meanlog=FLQuant,sdlog=FLQuant) :}{Describe method}
\item{ signature(n=numeric,meanlog=FLQuant,sdlog=numeric) :}{Describe method}
\item{ signature(n=numeric,meanlog=numeric,sdlog=FLQuant) :}{Describe method}
\item{ signature(n=numeric,meanlog=FLQuantPoint,sdlog=missing) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
