% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{runstest}
\alias{runstest}
\alias{runstest,FLQuants,missing-method}
\alias{runstest,FLQuants,FLQuants-method}
\alias{runstest,FLQuant,FLQuant-method}
\alias{runstest,FLQuant,missing-method}
\alias{runstest,numeric,numeric-method}
\alias{runstest,numeric,missing-method}
\title{Computes Runs Test p-values}
\usage{
runstest(fit, obs, ...)

\S4method{runstest}{FLQuants,missing}(fit, combine = TRUE)

\S4method{runstest}{FLQuants,FLQuants}(fit, obs, combine = TRUE)

\S4method{runstest}{FLQuant,FLQuant}(fit, obs, combine = TRUE)

\S4method{runstest}{FLQuant,missing}(fit, combine = TRUE)

\S4method{runstest}{numeric,numeric}(fit, obs, combine = TRUE)

\S4method{runstest}{numeric,missing}(fit, obs, combine = TRUE)
}
\arguments{
\item{fit}{The result of a model fit.}

\item{obs}{The observations used in the fit.}

\item{...}{Extra arguments.}

\item{combine}{Should ages be combined by addition, defaults to TRUE.}
}
\value{
A list with elements 'p.values' and 'pass'.
}
\description{
Computes Runs Test p-values
}
\examples{
data(nsher)
# Compute 'runstest' for FLSR fit
runstest(fit=fitted(nsher), obs=rec(nsher))
# Example runstest by age
data(ple4)
runstest(catch.n(ple4), landings.n(ple4), combine=FALSE)
runstest(fit=FLQuants(D=residuals(catch(ple4), discards(ple4)),
  L=residuals(catch(ple4), landings(ple4))))
runstest(fit=residuals(fitted(nsher), rec(nsher)))
runstest(FLQuants(residuals(fitted(nsher), rec(nsher))))
runstest(ssb(nsher))
runstest(rnorm(1, FLQuant(1, dimnames=list(year=1973:2021))))
runstest(rep(0.1, 10), cumsum(rnorm(10, 0.1, 0.01)))
runstest(rnorm(10, 0, 0.1))
}
