% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FLArray.R, R/FLPar.R, R/FLQuant.R, R/FLQuantDistr.R, R/FLQuantPoint.R, R/FLQuants.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,FLArray-method}
\alias{show,FLPar-method}
\alias{show,FLQuant-method}
\alias{show,FLQuantDistr-method}
\alias{show,FLQuantPoint-method}
\alias{show,FLQuants-method}
\title{Method show}
\usage{
\S4method{show}{FLArray}(object)

\S4method{show}{FLPar}(object)

\S4method{show}{FLQuant}(object)

\S4method{show}{FLQuantDistr}(object)

\S4method{show}{FLQuantPoint}(object)

\S4method{show}{FLQuants}(object)
}
\description{
Standard display of an object contents in an interactive session. Objects of
class \code{\linkS4class{FLQuant}} with length > 1 along the sixth dimension
(\emph{iter}) are output in a summarised form, as \code{median(mad)}, where
mad is the median absolute deviation. See \code{\link[stats]{mad}}.
}
\details{
The same format is used for objects of class \code{\linkS4class{FLPar}} with
length > 1 on the last dimension (\emph{iter}).
}
\section{Generic function}{
 show(object)
}
\examples{
# no 'iter'
  flq <- FLQuant(rnorm(80), dim=c(4,20), quant='age', units='kg')
  flq

# with 'iter'
  flq <- FLQuant(rnorm(800), dim=c(4,20,1,1,1,10), quant='age', units='kg')
  flq
}
\author{
The FLR Team
}
\seealso{
\link{FLComp}
}
\keyword{methods}

