% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{sigma3}
\alias{sigma3}
\title{Compute the 3-sigma limits and the corresponding p-value}
\usage{
sigma3(x, mixing = "less")
}
\arguments{
\item{x}{An object of class FLQuant.}

\item{mixing}{Alternative hypothesis to be tested. One of "two.sided", "less" (default) or "greater".}
}
\value{
A list with elements 'lcl', 'ucl' and 'p.value'.
}
\description{
Compute the 3-sigma limits and the corresponding p-value
}
\examples{
data(ple4)
sigma3(catch(ple4))
}
