% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLStock.R
\name{simplify}
\alias{simplify}
\alias{simplify-methods}
\alias{simplify,FLStock-method}
\title{Aggregate or select along unwanted dimensions}
\usage{
simplify(object, ...)

\S4method{simplify}{FLStock}(
  object,
  dims = c("unit", "season", "area")[dim(object)[3:5] > 1],
  spwn.season = 1,
  harvest = TRUE
)
}
\arguments{
\item{object}{A complex \strong{FLR} object to aggregate.}
}
\value{
An object of the same class as the input.
}
\description{
Objects of many FLR classes might be aggregated along the "unit", "season",
and/or "area" dimensions according to the type of data they contain.
}
\author{
The FLR Team
}
\keyword{methods}
