% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R
\name{sop}
\alias{sop}
\title{Method sop}
\usage{
sop(stock, slot = "catch")
}
\arguments{
\item{stock}{An FLStock object}

\item{slot}{Name of the slot group, i.e. "catch", "landings" or "discards"
for an FLStock object.}
}
\value{
Returns the ratio as an FLQuant
}
\description{
Calculates the sum of products correction
}
\details{
Calculates the \code{sop} correction for quantities such as catch,
discards and landings.  For example, in an object of class
\code{\linkS4class{FLStock}}, there are slots such as \emph{catch.n},
\emph{catch.wt} and \emph{catch}. \emph{catch} should equal the product
\emph{catch.n}*\emph{catch.wt} summed over ages. This function returns the
ratio (i.e. the correction) of \emph{catch.n}*\emph{catch.wt} : \emph{catch},
which can then be used to correct either \emph{catch.n} or \emph{catch.wt}.

Can be used for any class or slot where there are the three FLQuant slots
\emph{x}, \emph{x.n} and \emph{x.wt}, representing totals added over
all quants (ages), numbers by quant, and weight by quant.
}
\examples{

data(ple4)
sop(ple4,"catch")
sop(ple4,"landings")

}
\author{
FLR Team
}
\keyword{methods}

