\name{splom}
\docType{methods}
\alias{splom,ANY,ANY-method}
 \alias{splom,FLPar,missing-method}
\title{Method splom}
\description{
 My method is better then yours !
}
\section{Generic function}{
 splom(x,data) 
}
\section{Methods}{\describe{
\item{ signature(x=ANY,data=ANY) :}{Describe method}
\item{ signature(x=FLPar,data=missing) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
