\name{splom}
\docType{methods}
\alias{splom,ANY,ANY-method}
 \alias{splom,FLPar,missing-method}
\title{Method splom}
\description{
  Draws a conditional scatter plot matrix.

 See the help page in \code{lattice} for a full description of each plot and all
 possible arguments.
}
\section{Generic function}{
 splom(x,data) 
}
\section{Methods}{\describe{
\item{ signature(x=FLPar,data=missing) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{splom} }
\examples{}
\keyword{methods}
