\name{spr0}
\docType{methods}
\alias{spr0}
\alias{spr0-methods}
\alias{spr0,FLQuant,FLQuant,FLQuant-method}
\alias{spr0,FLStock,missing,missing-method}
\alias{spr0,FLSR,missing,FLQuant-method}
\title{Method spr0}
\description{
 My method is better then yours !
}
\section{Generic function}{
 quant(ssb, rec, f) 
}
\section{Methods}{\describe{
\item{ signature(ssb=FLQuant, rec=FLQuant, f=FLQuant) :}{Describe method}
\item{ signature(ssb=FLStock, rec=missing, f=missing) :}{Describe method}
\item{ signature(ssb=FLSR, rec=missing, f=FLQuant) :}{Describe method}
}}
\author{The FLR Team}
\seealso{ \link{FLComp} }
\examples{}
\keyword{methods}
