% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRmodels.R
\name{SRModels}
\alias{SRModels}
\alias{ricker}
\alias{ab2sv}
\alias{bevholt}
\alias{bevholt.ar1}
\alias{bevholt.c.a}
\alias{bevholt.c.b}
\alias{bevholt.d}
\alias{bevholt.ndc}
\alias{bevholt.sv}
\alias{Bevholt.SV}
\alias{geomean}
\alias{logl.ar1}
\alias{ricker.ar1}
\alias{ricker.c.a}
\alias{ricker.c.b}
\alias{ricker.d}
\alias{ricker.sv}
\alias{Ricker.SV}
\alias{segreg}
\alias{shepherd}
\alias{shepherd.ar1}
\alias{shepherd.d}
\alias{shepherd.d.ar1}
\alias{shepherd.ndc}
\alias{shepherd.ndc.ar1}
\alias{sv2ab}
\alias{SRModels}
\alias{bevholt}
\alias{SRModels}
\alias{segreg}
\alias{SRModels}
\alias{geomean}
\alias{SRModels}
\alias{shepherd}
\alias{SRModels}
\alias{cushing}
\alias{SRModels}
\alias{rickerSV}
\alias{SRModels}
\alias{bevholtSV}
\alias{SRModels}
\alias{shepherdSV}
\alias{SRModels}
\alias{bevholtAR1}
\alias{SRModels}
\alias{rickerAR1}
\alias{SRModels}
\alias{segregAR1}
\alias{SRModels}
\alias{rickerCa}
\title{Stock-Recruitment models}
\usage{
ricker()

bevholt()

segreg()

geomean()

shepherd()

cushing()

rickerSV()

bevholtSV()

shepherdSV()

bevholtAR1()

rickerAR1()

segregAR1()

rickerCa()
}
\arguments{
\item{rho}{Autoregression}

\item{sigma2}{Autoregression}

\item{obs}{Observed values}

\item{hat}{estimated values}

\item{steepness}{Steepness.}

\item{vbiomass}{Virgin biomass.}

\item{spr0}{Spawners per recruit at F=0, see \code{\link{spr0}}.}

\item{model}{character vector with model name, either 'bevholt' or 'ricker'.}
}
\description{
A range of stock-recruitment (SR) models commonly used in fisheries science
are provided in FLCore.
}
\examples{

# inspect the output of one of the model functions
  bevholt()
  names(bevholt())
  bevholt()$logl

# once an FLSR model is in the workspace ...
  data(nsher)

# the three model-definition slots can be modified
# at once by calling 'model<-' with
# (1) a list
  model(nsher) <- bevholt()

# (2) the name of the function returning this list
  model(nsher) <- 'bevholt'

# or (3) the function itself that returns this list
  model(nsher) <- bevholt

}
\references{
Beverton, R.J.H. and Holt, S.J. (1957) On the dynamics of
exploited fish populations. MAFF Fish. Invest., Ser: II 19, 533.

Needle, C.L. Recruitment models: diagnosis and prognosis.  Reviews in Fish
Biology and Fisheries 11: 95-111, 2002.

Ricker, W.E. (1954) Stock and recruitment. J. Fish. Res. Bd Can. 11,
559-623.

Shepherd, J.G. (1982) A versatile new stock-recruitment relationship for
fisheries and the construction of sustainable yield curves.  J. Cons. Int.
Explor. Mer 40, 67-75.
}
\seealso{
\linkS4class{FLSR}, \linkS4class{FLModel}
}
\author{
The FLR Team
}
\keyword{models}
