% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLStock.R, R/FLBiol.R
\docType{methods}
\name{ssb}
\alias{ssb}
\alias{ssb-methods}
\alias{ssb,FLStock-method}
\alias{ssb-FLStock,method}
\alias{ssb,FLBiol-method}
\alias{ssb-FLBiol,method}
\title{Calculate or return the Spawning Stock Biomass}
\usage{
ssb(object, ...)

\S4method{ssb}{FLStock}(object, ...)

\S4method{ssb}{FLBiol}(object, ...)
}
\arguments{
\item{object}{Object on which \code{ssb} is calculated or extracted.}
}
\value{
An object, generally of class \code{\link{FLQuant}}.
}
\description{
The calculated Spawning Stock Biomass (SSB) of a fish population is returned
by this method. SSB is the combined weight of all individuals in a fish stock
that are capable of reproducing. In some classes this is calculated from
information stored in different slots, while in others \code{ssb()} is simply
an slot accessor. When the later is the case, the corresponding replacement
method also exists.
}
\details{
For an object of class \code{\link{FLStock}}, the calculation of SSB depends
on the value of the 'units' attribute in the \code{harvest} slot. If this is
in terms of fishing mortality (\code{units(harvest(object)) == 'f'}), and
assuming an object structured by age, then SSB is calculated as
\deqn{SSB_{y} = \sum\nolimits_{a} N_{a,y} \cdot e^{-(F_{a,y} \cdot Hs_{a,y} + M_{a,y} \cdot Ms_{a,y})} \cdot W_{a,y} \cdot T_{a,y} }{SSB_y = sum_a(N_ay * exp(-(F_ay * Hs_ay + M_ay * Ms_ay)) * W_ay * T_ay)}
where \eqn{N_{a,y}}{N_ay} is the abundance in numbers (\code{stock.n}) by
age (a) and year (y), \eqn{F_{a,y}}{F_ay} is the fishing mortality (\code{harvest}), 
\eqn{Hs_{a,y}}{Hs_ay} is the proportion of fishing mortality before spawning
(\code{harvest.spwn}),
\eqn{M_{a,y}}{M_ay} is the natural mortality (\code{m}), 
\eqn{Ms_{a,y}}{Ms_ay} is the proportion of natural mortality before spawning
(\code{m.spwn}),
\eqn{W_{a,y}}{W_ay} is the mean weight at age in the stock (\code{m}), and
\eqn{T_{a,y}}{T_ay} is the proportion mature at age in the stock (\code{mat}).
For \code{\link{FLStock}} objects with other dimensions (\code{area},
\code{unit}), the calculation is carried out along those dimensions too. To
obtain a global value please use the corresponding summing method.

If the harvest slot contains estimates in terms of harvest rates
(\code{units(harvest(object)) == "hr"}), SSB will be computed as
\deqn{SSB_{y} = \sum\nolimits_{a} N_{a,y} \cdot (1 - H_{a,y} \cdot Hs_{a,y}) \cdot e^{-(M_{a,y} \cdot Ms_{a,y})} \cdot W_{a,y} \cdot T_{a,y} }{SSB_y = sum_a(N_ay * (1 - H_ay * Hs_ay) * exp(-(M_ay * Ms_ay)) * W_ay * T_ay)}
where \eqn{H_{a,y}}{H_ay} is the harvest rate (proportion of catch in weight
over total biomass).

For an object of class \code{\link{FLBiol}}, the calculation of SSB is as
follows:
\deqn{SSB_{a,y} = \sum\nolimits_{a} N_{a,y} \cdot W_{a,y} \cdot e^{-S_{a,y} \cdot M_{a,y}}}{SSB_ay = sum_a N_ay * W_ay * exp(-S_ay * M_ay)}
where \eqn{S_{a,y}}{S_ay} is the fraction of natural mortality before spawning (\code{spwn}). The method for this class does not correct the estimate, based on
abundances at the start of the time period, for fishing mortality ocurring
between that time and spawning, as the class holds no information on fishing
mortality.
}
\examples{

data(ple4)
ssb(ple4)
}
\seealso{
\link{FLComp}

\code{\link{areaSums}}
}
\author{
The FLR Team
}
\keyword{methods}
