% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R, R/FLStock.R, R/FLBiol.R
\name{ssb}
\alias{ssb}
\alias{ssb-methods}
\alias{ssb,FLStock-method}
\alias{ssb,FLBiol-method}
\alias{ssb-FLBiol,method}
\title{Calculate or return the Spawning Stock Biomass}
\usage{
ssb(object, ...)

\S4method{ssb}{FLStock}(object, ...)

\S4method{ssb}{FLBiol}(object, ...)
}
\arguments{
\item{object}{Object on which \code{ssb} is calculated or extracted.}
}
\value{
An object, generally of class \code{\link{FLQuant}}.
}
\description{
The calculated Spawning Stock Biomass (SSB) of a fish population is returned
by this method. SSB is the combined weight of all individuals in a fish stock
that are capable of reproducing. In some classes this is calculated from
information stored in different slots, while in others \code{ssb()} is simply
an slot accessor. When the later is the case, the corresponding replacement
method also exists.

For an object of class \code{\link{FLStock}}, the calculation of SSB depends
on the value of the 'units' attribute in the \code{harvest} slot. If this is
in terms of fishing mortality (\code{units(harvest(object)) == 'f'}), and
assuming an object structured by age, then SSB is calculated as
\deqn{SSB_{y} = \sum\nolimits_{a} N_{a,y} \cdot e^{-(F_{a,y} \cdot Hs_{a,y} + M_{a,y} \cdot Ms_{a,y})} \cdot W_{a,y} \cdot T_{a,y} }{SSB_y = sum_a(N_ay * exp(-(F_ay * Hs_ay + M_ay * Ms_ay)) * W_ay * T_ay)}
where \eqn{N_{a,y}}{N_ay} is the abundance in numbers (\code{stock.n}) by
age (a) and year (y), \eqn{F_{a,y}}{F_ay} is the fishing mortality (\code{harvest}), 
\eqn{Hs_{a,y}}{Hs_ay} is the proportion of fishing mortality before spawning
(\code{harvest.spwn}),
\eqn{M_{a,y}}{M_ay} is the natural mortality (\code{m}), 
\eqn{Ms_{a,y}}{Ms_ay} is the proportion of natural mortality before spawning
(\code{m.spwn}),
\eqn{W_{a,y}}{W_ay} is the mean weight at age in the stock (\code{m}), and
\eqn{T_{a,y}}{T_ay} is the proportion mature at age in the stock (\code{mat}).
For \code{\link{FLStock}} objects with other dimensions (\code{area},
\code{unit}), the calculation is carried out along those dimensions too. To
obtain a global value please use the corresponding summing method.
If the harvest slot contains estimates in terms of harvest rates
(\code{units(harvest(object)) == "hr"}), SSB will be computed as
\deqn{SSB_{y} = \sum\nolimits_{a} N_{a,y} \cdot (1 - H_{a,y} \cdot Hs_{a,y}) \cdot e^{-(M_{a,y} \cdot Ms_{a,y})} \cdot W_{a,y} \cdot T_{a,y} }{SSB_y = sum_a(N_ay * (1 - H_ay * Hs_ay) * exp(-(M_ay * Ms_ay)) * W_ay * T_ay)}
where \eqn{H_{a,y}}{H_ay} is the harvest rate (proportion of catch in weight
over total biomass).
}
\details{
Objects of the *FLBiol* class do not contain any information on
catch or fishing mortality, so a call to `ssb()` will only correct abundances
for natural mortality to the moment of spawning. The method can also take
information on catches or fishing mortality and use them when calculating
abundances at spawning time. An *FLQuant* named either 'catch.n', 'f', 'hr' or
'harvest' can be used. The first three are self-explanatory, while for the last
units must be either 'f' or 'hr'. The quantities should refer to total yearly
values, as the value in the 'spwn' slot will be used to calculate what fraction
of fishing mortality to apply.
}
\examples{
data(ple4)
# SSB from FLStock
ssb(ple4)
biol <- as(ple4, "FLBiol")
# SSB from FLBiol, abundances corrected only for M
ssb(biol)
# Provide catch-at-age, F or HR to correct N
ssb(biol, catch.n=catch.n(ple4))
ssb(biol, f=harvest(ple4))
ssb(biol, harvest=harvest(ple4))
ssb(biol, hr=catch.n(ple4) / stock.n(ple4))
}
\seealso{
\link{FLComp}

\code{\link{areaSums}}
}
\author{
The FLR Team
}
\keyword{methods}
