% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oem.R
\name{survey}
\alias{survey}
\alias{survey,FLStock,FLIndex-method}
\alias{survey,FLStock,FLIndexBiomass-method}
\alias{survey,FLStock,missing-method}
\alias{survey,FLStock,FLIndices-method}
\title{A method to generate observations of abundance at age.}
\usage{
survey(object, index, ...)

\S4method{survey}{FLStock,FLIndex}(
  object,
  index,
  sel = sel.pattern(index),
  ages = dimnames(index)$age,
  timing = mean(range(index, c("startf", "endf"))),
  index.q = index@index.q,
  stability = 1
)

\S4method{survey}{FLStock,FLIndexBiomass}(
  object,
  index,
  sel = sel.pattern(index),
  ages = ac(seq(range(index, c("min")), range(index, c("max")))),
  timing = mean(range(index, c("startf", "endf"))),
  catch.wt = stock.wt(object)[, dimnames(index)$year],
  index.q = index@index.q,
  stability = 1
)

\S4method{survey}{FLStock,missing}(
  object,
  sel = catch.sel(object),
  ages = dimnames(sel)$age,
  timing = 0.5,
  index.q = 1,
  biomass = FALSE,
  stability = 1
)

\S4method{survey}{FLStock,FLIndices}(object, index, ...)
}
\arguments{
\item{object}{The object on which to draw the observation}
}
\value{
An FLQuant for the index of abundance
}
\description{
A method to generate observations of abundance at age.
}
\examples{
data(ple4)
data(ple4.index)
# CONSTRUCT a survey from stock and index
survey(ple4, ple4.index)
# Create FLIndexBiomass
ple4.biom <- as(ple4.index, "FLIndexBiomass")
survey(ple4, ple4.biom)
data(ple4)
survey(ple4)
survey(ple4, biomass=TRUE)
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
