% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom.R
\docType{methods}
\name{uom}
\alias{uom}
\alias{uomUnits}
\title{uom Units of Measurement}
\usage{
uom(op, u1, u2)

uomUnits(unit = missing)
}
\arguments{
\item{op}{The arithmetic operator to be used, one of '+', '-', '*' or '/'}

\item{u1}{The units of measurement string of the first object}

\item{u2}{The units of measurement string of the second object}

\item{unit}{A character vector for one or more units to be compared with those known to \code{uom}.}
}
\value{
\code{uom} returns a string with the corresponding units of measurement, or a character vector, showing the operation carried out, when units are not known to \code{uom} or not compatible, e.g. "100 * d".

\code{uomUnits} returns TRUE or FALSE if \code{unit} is given, otherwise a character vector with all units known to \code{uom}.
}
\description{
The 'units' attribute of FLQuant objects provides a mechanism for keeping track
of the units of measurement of that particular piece of data.
}
\details{
Arithmetic operators for 'FLQuant' objects are aware of a limited set of units
of measurement and will output the right unit when two object are arithmetically
combined. For example, the product of object with units of 'kg' and '1000' will
output an object with 'units' of 't' (for metric tonnes).

Operations involving combinations of units not defined will issue a warning, and
the resulting 'units' attribute will simply keep a string indicating the input
units of measurement and the operation carried out, as in '10 * 1000'.

Note that no scaling or modification of the values in the object takes place.

Conversion across units is carried out by the \code{uom()} function

The list of units known to \code{uom} is stored internally but can be queried by calling \code{uomUnits()} with no arguments. If a character vector is provided, a logical is returned telling whether the string is included or not in that table.
}
\examples{

# Conversion between weights
FLQuant(1, units='kg') * FLQuant(1000, units='1')

# Conversion between mortalities
FLQuant(0.2, units='m') + FLQuant(0.34, units='f')

# Check if units are known
uomUnits('kg')
uomUnits('kell')
}
\seealso{
\code{\linkS4class{FLQuant}} \code{\link{units,FLArray-method}}
}
\author{
The FLR Team
}
\keyword{function}
