% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom.R
\docType{data}
\name{uomTable}
\alias{uomTable}
\title{Table for conversions and operations between units of measurement}
\format{
An object of class array
}
\description{
\itemize{
\item uom defaults to NA unless defined below.
\item unit +/- itself, returns the same unit (e.g. kg + kg = kg)
\item numeric unit * 1 returns same unit (e.g. 1e4 * 1 = 1e4)
\item numeric unit * numeric unit returns product (e.g. 10 * 100 = 1000)
\item unit / unit returns "" (e.g. 100 / 100 = "")
\item numeric unit / smaller numeric unit returns division (e.g. 100 / 10 = 10)
\item 100 times kg returns t
\item numeric unit * 'kg' returns the product in tonnes (e.g. kg * 1e4 = t * 10)
\item units with divisions are parsed (e.g. days/boat * boat = days)
\item 
\item 
}
}
\keyword{datasets}
