% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLStock.R, R/FLQuants.R
\name{weighted.mean,FLQuants,FLQuants-method}
\alias{weighted.mean,FLQuants,FLQuants-method}
\title{Weighted means along a FLQuants.}
\usage{
\S4method{weighted.mean}{FLQuants,FLQuants}(x, w)

\S4method{weighted.mean}{FLQuants,FLQuants}(x, w)
}
\arguments{
\item{x}{Values to be averaged, as an object of class \code{FLQuants}.}

\item{w}{weights to be used, as an object of class \code{FLQuants}.}
}
\value{
A single \code{FLQuant} object.
}
\description{
Facilitates the calculation of weighted means across a FLQuants object.
}
\details{
An object of class FLQuants containing elements over which an average is to
computed, is combined with another one, of the same length, containing
values to be used as weights. The overall weighted mean is calculated by
computing the product of each element to its corresponding weight, and
dividing by the sum of all weights.
NAs in the value elements are substituted for zeroes, so do not influence
the mean. Weights are addedn a small quantity (1e-36) to avoid division over zero
errors
}
\examples{
data(ple4)
x <- FLQuants(landings.wt(ple4), discards.wt(ple4))
w <- FLQuants(landings.n(ple4), discards.n(ple4))
# Computes weighted mean of landings and discards weights-at-age
weighted.mean(x, w)
data(ple4)
x <- FLQuants(landings.wt(ple4), discards.wt(ple4))
w <- FLQuants(landings.n(ple4), discards.n(ple4))
# Weighted mean of landings and discards weights-at-age
weighted.mean(x, w)
}
\seealso{
\link[FLCore::FLQuants stats:weighted.mean]{FLCore::FLQuants stats::weighted.mean}
}
\author{
The FLR Team
}
\keyword{methods}
